% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{add_copy_comment_europe}
\alias{add_copy_comment_europe}
\title{add_copy_comment_europe}
\usage{
add_copy_comment_europe(x, copy_name)
}
\arguments{
\item{x}{An object}

\item{copy_name}{A string with the name that has been copied}
}
\value{
\code{x} with comments appended to any existing comments.
}
\description{
Add comment to indicate object was copied and filtered to european regions
}
