% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{add_legacy_name}
\alias{add_legacy_name}
\title{add_legacy_name}
\usage{
add_legacy_name(x, legacy_name)
}
\arguments{
\item{x}{An object}

\item{legacy_name}{Legacy name (character)}
}
\value{
\code{x} with legacy name set.
}
\description{
Add the legacy (old data system) name to a data system object.
}
