% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{add_node_leaf_names}
\alias{add_node_leaf_names}
\title{add_node_leaf_names}
\usage{
add_node_leaf_names(
  data,
  nesting_table,
  leaf_name,
  ...,
  LT_name = "Land_Type",
  append_GLU = TRUE
)
}
\arguments{
\item{data}{Data, tibble}

\item{nesting_table}{Table of node names (as column names) and leaf data (contents), tibble}

\item{leaf_name}{Name of leaf name column in \code{nesting_table}, character}

\item{...}{Names of columns to add leaf nodes for, character}

\item{LT_name}{Name of land type column in \code{data}, character}

\item{append_GLU}{Append GLU column to new leaf name columns? Logical}
}
\value{
Data with new leaf name columns added.
}
\description{
Match in the node and leaf names from a land nesting table
}
