% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{add_price_units}
\alias{add_price_units}
\title{add_price_units}
\usage{
add_price_units(x, units)
}
\arguments{
\item{x}{An object}

\item{units}{Price Units (character)}
}
\value{
\code{x} with units appended to any existing comments.
}
\description{
Add character price units to a data system object. Units are written out
with the data when the file is saved.
}
