% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{add_xml_data_generate_levels}
\alias{add_xml_data_generate_levels}
\title{Add a table to convert to XML but generate additional levels of XML nesting}
\usage{
add_xml_data_generate_levels(
  dot,
  data,
  header,
  old_tag,
  new_tag,
  num_levels,
  rename_final,
  column_name = old_tag,
  column_order_lookup = header
)
}
\arguments{
\item{dot}{The current state of the pipeline started from \code{create_xml}}

\item{data}{The tibble of data to add to the conversion}

\item{header}{The base header tag to can be looked up in the header file to
convert \code{data} besides the additional levels of nesting.}

\item{old_tag}{The XML tag which is being expanded to add more levels}

\item{new_tag}{The XML tag which will serve as the additional levels}

\item{num_levels}{The number of additional levels to generate}

\item{rename_final}{Whether to rename the final \code{old_tag} to
\code{new_tag} in the original header}

\item{column_name}{The base name of the column in \code{data} which will get
expanded by appending \code{paste0(column_name, (seq_len(num_levels) - 1))}}

\item{column_order_lookup}{A tag that can be used to look up \code{LEVEL2_DATA_NAMES}
to reorder the columns of data before XML conversion to ensure they correspond
with the ModelInterface header.  Note by default the \code{header} is used then we
append the additional columns generated by \code{column_name} + \code{num_levels}. If
given \code{NULL} no column reordering will be done.}
}
\value{
A "data structure" to hold the various parts needed to run the model
interface CSV to XML conversion.
}
\description{
We have the ability to create as many levels of "subsector" nesting as we
want but in doing so we want to avoid having to make an explict copy of
the headers / LEVEL2_DATA_NAMES. So we provide this method which will instruct
the ModelIntercae to generate the addtional levels automatically.  It assumes
the data in the base header is provided as is and the additional columns needed
for nesting will be moved to the end (see detail for the \code{column_name} and
\code{column_order_lookup} params).
}
\author{
Pralit Patel
}
