% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{append_peat_carbon_info}
\alias{append_peat_carbon_info}
\title{append_peat_carbon_info}
\usage{
append_peat_carbon_info(
  data,
  peat_table,
  join_vars = intersect(names(data), names(peat_table))
)
}
\arguments{
\item{data}{Data table to modify}

\item{peat_table}{table with relevant peatland region, land leaf, soil type, and share}

\item{join_vars}{"by" variables for the join between the peat table and the data table}
}
\value{
full data table with peat info appended
}
\description{
Helper function to append carbon density information to peatlands to an existing table of carbon density information
}
