% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{append_peat_technologies}
\alias{append_peat_technologies}
\title{append_peat_technologies}
\usage{
append_peat_technologies(
  data,
  peat_table,
  join_vars = intersect(names(data), names(peat_table))
)
}
\arguments{
\item{data}{Data to expand}

\item{peat_table}{table with peatland region, supplysector, subsector}

\item{join_vars}{"by" variables for the join between the peat table and the data table}
}
\value{
full data table with peat info appended
}
\description{
Helper function to copy an existing table to corresponding technologies/subsectors on peat soil,
based on a separate table with the relevant subsectors on peat
}
