% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{apply_peat_emissions}
\alias{apply_peat_emissions}
\title{apply_peat_emissions}
\usage{
apply_peat_emissions(
  data,
  peat_table,
  join_vars = intersect(names(data), names(peat_table))
)
}
\arguments{
\item{data}{Data table to modify}

\item{peat_table}{table with relevant peatland region, land leaf, soil type, and share}

\item{join_vars}{"by" variables for the join between the peat table and the data table}
}
\value{
full data table with peat info appended
}
\description{
Helper function to apply land share to a table of burning emissions, appending the "Peat" string as appropriate,
multiplying emissions quantities by peat (or mineral) shares
}
