% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-helpers.R
\name{approx_fun_constant}
\alias{approx_fun_constant}
\title{approx_fun_constant}
\usage{
approx_fun_constant(year, value, rule = 1)
}
\arguments{
\item{year}{Numeric year, in a melted tibble or data frame}

\item{value}{Numeric value to interpolate}

\item{rule}{Rule to use; see \code{\link{approx}} and details}
}
\value{
Interpolated values.
}
\description{
\code{\link{approx}} (interpolation) for use in a dplyr pipeline.
}
\details{
This is a slight change to approx_fun that can be used if there is only one value, not two to interpolate between.

This function will apply the one value you do have to all other years in a grouping.
}
\examples{
df <- data.frame(year = 1:5, value = c(1, 2, NA, 4, 5))
approx_fun_constant(df$year, df$value, rule = 2)
}
