% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-helpers.R
\name{change_iso_code}
\alias{change_iso_code}
\title{Change a particular ISO code in a table to another code.}
\usage{
change_iso_code(d, from, to, col = "iso")
}
\arguments{
\item{d}{The data to be transformed}

\item{from}{The old ISO code}

\item{to}{The replacement ISO code}

\item{col}{The name of the columns with the ISO codes (default = 'iso')}
}
\value{
Tibble with the old iso code replaced by the new one.
}
\description{
A few countries have changed their ISO codes over time.  Usually this is because countries
split or merged; occasionally it's because they just didn't like the old one.  This function
allows us to insert an ISO code change into a pipeline.
}
\details{
If you're dealing with a split or a merger, then making this change will generally create
duplicate rows, so you should follow it up with an appropriate aggregation.
}
