% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunk_details_description}
\alias{chunk_details_description}
\title{chunk_details_description}
\usage{
chunk_details_description(chunkname)
}
\arguments{
\item{chunkname}{The name of the chunk to extract details from}
}
\value{
Tibble with chunk name, details, and description
}
\description{
Get the description and details from a chunk from its /man*.Rd page.
All non-batch chunk info can be retrieved with the following code:
chunk_names <- find_chunks('^module_[a-zA-Z\\.]*_L.*$')
bind_rows(lapply(chunk_names$name, chunk_details_description))
}
\author{
HN & RLH October 2022
}
