% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunk_inputs}
\alias{chunk_inputs}
\title{chunk_inputs}
\usage{
chunk_inputs(chunks = find_chunks()$name, call_flag = driver.DECLARE_INPUTS)
}
\arguments{
\item{chunks}{A character vector of chunks names}

\item{call_flag}{Flag to look for in inputs}
}
\value{
A tibble with columns 'name' (chunk name), 'input' (name of data),
'file_file' (whether object is read from a file), and 'optional' (whether
the object is optional or not).
}
\description{
chunk_inputs
}
