% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunks_info}
\alias{chunks_info}
\title{chunks_info}
\usage{
chunks_info(chunkname)
}
\arguments{
\item{chunkname}{The name of the chunk to extract details from}
}
\value{
Tibble with chunk name, details, and description
}
\description{
Get the description and details from a chunk from its /man*.Rd page.
Currently setup to return info of all non-batch (xml) chunks.
}
\details{
All non-batch chunk info can be retrieved with the following code:
chunk_names <- find_chunks('^module_[a-zA-Z\\.]*_L.*$')
bind_rows(lapply(chunk_names$name, chunks_info))
}
\examples{
chunks_info("module_water_L1233.Elec_water")
chunk_names <- find_chunks('^module_[a-zA-Z\\\\.]*_L.*$')
bind_rows(lapply(chunk_names$name, chunks_info))

}
\author{
Hassan Niazi, October 2022
}
