% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{cogen_global_tech}
\alias{cogen_global_tech}
\title{cogen_global_tech}
\usage{
cogen_global_tech(df_name, env)
}
\arguments{
\item{df_name}{String name of tibble}
}
\value{
assignment of new db
}
\description{
Helper function to rename cogen techs in global tech db
}
