% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{cogen_stubtech_rename}
\alias{cogen_stubtech_rename}
\title{cogen_stubtech_rename}
\usage{
cogen_stubtech_rename(df_name, env, grid_region_df = grid_regions)
}
\arguments{
\item{df_name}{String name of tibble}
}
\value{
assignment of new db
}
\description{
Helper function to rename cogen techs in global tech db
}
