% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-helpers.R
\name{copy_data_forward_long}
\alias{copy_data_forward_long}
\title{Helper function to extend data which is in long format.}
\usage{
copy_data_forward_long(data, var, all_years, ...)
}
\arguments{
\item{data}{The data to extend}

\item{var}{The column which contains the data to copy}

\item{all_years}{The set of years for which there must be data.}

\item{...}{The result of the columns to "nest" by}
}
\value{
The data extended to include all requested years.
}
\description{
Given a data frame the \code{var} column which contains the data and a set of years
for which all years must have data this helper expands the data and fills using
\code{approx_fun(rule=2)}, in other words will copy data beyond the end years.
Thus making it suitable to extend historical data sets by copying forward.
}
