% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-helpers.R
\name{copy_data_forward_wide}
\alias{copy_data_forward_wide}
\title{Helper function to extend data which is in wide format.}
\usage{
copy_data_forward_wide(data, copy_from_year, copy_to_years)
}
\arguments{
\item{data}{The data to extend}

\item{copy_from_year}{The column which contains the data to copy}

\item{all_years}{The set of years for which there must be data.}
}
\value{
The data extended to include all requested years.
}
\description{
Given a data frame and a \code{copy_from_year} column this helper will copy it
to new columns for each \code{copy_to_years}.
}
