% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{copy_filter_europe}
\alias{copy_filter_europe}
\title{copy_filter_europe}
\usage{
copy_filter_europe(
  all_data,
  df_list,
  regions_to_keep = gcameurope.EUROSTAT_COUNTRIES
)
}
\arguments{
\item{df_list}{Character vector of outputs from "main" datasystem to filter}

\item{regions_to_keep}{character vector of regions to remove}
}
\value{
list of new tibbles
}
\description{
Helper function to copy an object from main data system and filter to Europe
}
