% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{currency_constant}
\alias{currency_constant}
\title{currency_constant}
\usage{
currency_constant(value, conv_year = PRICE_YEAR)
}
\arguments{
\item{value}{The price/cost constant value, with an attribute, ATTR_CURR_YEAR, that contains the dollar year of the value}

\item{conv_year}{The year to convert to, defaults to PRICE_YEAR}
}
\value{
Converted price/cost value in conv_year dollars
}
\description{
Convert any currency/price/cost constants with \code{gdp_deflator}
}
\author{
RLH Dec 2022
}
