% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_solar_reeds_USA_xml.R
\name{disabled_module_gcamusa_batch_solar_reeds_USA_xml}
\alias{disabled_module_gcamusa_batch_solar_reeds_USA_xml}
\title{module_gcamusa_batch_solar_reeds_USA_xml}
\usage{
disabled_module_gcamusa_batch_solar_reeds_USA_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{solar_reeds_USA.xml}.
The corresponding file in the original data system was \code{batch_solar_USA_reeds.xml} (gcamusa XML batch).
}
\description{
Construct XML data structure for \code{solar_reeds_USA.xml}.
}
