% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2234.elec_segments_USA.R
\name{disabled_module_gcamusa_L2234.elec_segments_USA}
\alias{disabled_module_gcamusa_L2234.elec_segments_USA}
\title{module_gcamusa_L2234.elec_segments_USA}
\usage{
disabled_module_gcamusa_L2234.elec_segments_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2234.Supplysector_elecS_USA}, \code{L2234.ElecReserve_elecS_USA}, \code{L2234.SubsectorLogit_elecS_USA},
\code{L2234.SubsectorShrwtInterp_elecS_USA}, \code{L2234.SubsectorShrwtInterpTo_elecS_USA}, \code{L2234.SubsectorShrwt_elecS_USA},
\code{L2234.StubTechEff_elecS_USA}, \code{L2234.StubTechCapFactor_elecS_solar_USA}, \code{L2234.StubTechCapFactor_elecS_wind_USA},
\code{L2234.SubsectorShrwtFllt_elecS_grid_USA}, \code{L2234.SubsectorShrwtInterp_elecS_grid_USA}, \code{L2234.PassThroughSector_elecS_USA},
\code{L2234.PassThroughTech_elecS_grid_USA}, \code{L2234.GlobalTechShrwt_elecS_USA}, \code{L2234.GlobalIntTechShrwt_elecS_USA},
\code{L2234.PrimaryRenewKeyword_elecS_USA},\code{L2234.PrimaryRenewKeywordInt_elecS_USA}, \code{L2234.AvgFossilEffKeyword_elecS_USA},
\code{L2234.GlobalTechCapital_elecS_USA}, \code{L2234.GlobalIntTechCapital_elecS_USA}, \code{L2234.GlobalTechOMfixed_elecS_USA},
\code{L2234.GlobalIntTechOMfixed_elecS_USA}, \code{L2234.GlobalTechOMvar_elecS_USA}, \code{L2234.GlobalIntTechOMvar_elecS_USA},
\code{L2234.GlobalTechCapFac_elecS_USA}, \code{L2234.GlobalTechEff_elecS_USA}, \code{L2234.GlobalIntTechEff_elecS_USA},
\code{L2234.GlobalTechLifetime_elecS_USA}, \code{L2234.GlobalIntTechLifetime_elecS_USA}, \code{L2234.GlobalTechProfitShutdown_elecS_USA},
\code{L2234.GlobalTechSCurve_elecS_USA}, \code{L2234.GlobalTechCapture_elecS_USA}, \code{L2234.GlobalIntTechBackup_elecS_USA},
\code{L2234.StubTechMarket_elecS_USA}, \code{L2234.StubTechMarket_backup_elecS_USA}, \code{L2234.StubTechElecMarket_backup_elecS_USA},
\code{L2234.StubTechProd_elecS_USA}, \code{L2234.StubTechFixOut_elecS_USA}, \code{L2234.StubTechFixOut_hydro_elecS_USA}, \code{L2234.StubTechCost_offshore_wind_elecS_USA},
\code{L2234.TechShrwt_elecS_grid_USA}, \code{L2234.TechCoef_elecS_grid_USA}, \code{L2234.TechProd_elecS_grid_USA}.
The corresponding file in the original data system was \code{L2234.elec_segments_USA.R} (gcam-usa level2).
}
\description{
Generates GCAM-USA model inputs for multiple load segments electricity sector by state.
}
\details{
This chunk generates input files to create an electricity generation sector with multiple load segments
for each state and creates the demand for the state-level electricity sectors in the grid regions.
}
\author{
MTB Aug 2018
}
