% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2242.elec_hydro_USA.R
\name{disabled_module_gcamusa_L2242.elec_hydro_USA}
\alias{disabled_module_gcamusa_L2242.elec_hydro_USA}
\title{module_gcamusa_L2242.elec_hydro_USA}
\usage{
disabled_module_gcamusa_L2242.elec_hydro_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2242.StubTechFixOut_hydro_USA}.
The corresponding file in the original data system was \code{L2242.elec_hydro_USA.R} (gcam-usa level2).
}
\description{
Update hydro-electricity fixed output to match historical data for 2015 and AEO-2018 for the future.
}
\details{
Update state-level hydro-electricity fixed outputs
}
\author{
MTB September 2018
}
