% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2246.coal_vintage_USA.R
\name{disabled_module_gcamusa_L2246.coal_vintage_USA}
\alias{disabled_module_gcamusa_L2246.coal_vintage_USA}
\title{module_gcamusa_L2246.coal_vintage_USA}
\usage{
disabled_module_gcamusa_L2246.coal_vintage_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2246.StubTechProd_coal_vintage_USA}, \code{L2246.StubTechEff_coal_vintage_USA},
\code{L2246.StubTechLifetime_coal_vintage_USA}, \code{L2246.StubTechSCurve_coal_vintage_USA},
\code{L2246.StubTechProfitShutdown_coal_vintage_USA}, \code{L2246.StubTechMarket_coal_vintage_USA},
\code{L2246.GlobalTechShrwt_coal_vintage_USA}, \code{L2246.GlobalTechEff_coal_vintage_USA},
\code{L2246.GlobalTechCapFac_coal_vintage_USA}, \code{L2246.GlobalTechCapital_coal_vintage_USA},
\code{L2246.GlobalTechOMfixed_coal_vintage_USA}, \code{L2246.GlobalTechOMvar_coal_vintage_USA}.
The corresponding file in the original data system was \code{L2246.coal_vintage.R} (gcam-usa level2).
}
\description{
Generates GCAM-USA model inputs to include detailed coal vintages.
}
\details{
This chunk creates input files to include detailed coal vintages. Calibrated outputs in 2010 are allocated based on
generation share by vintage groups in 2015 using EIA unit-level data.
}
\author{
RC Sep 2018
}
