% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2261.regional_biomass_USA.R
\name{disabled_module_gcamusa_L2261.regional_biomass_USA}
\alias{disabled_module_gcamusa_L2261.regional_biomass_USA}
\title{module_gcamusa_L2261.regional_biomass_USA}
\usage{
disabled_module_gcamusa_L2261.regional_biomass_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2261.DeleteSupplysector_bio_USA}, \code{L2261.Supplysector_bio_USA},
\code{L2261.SubsectorShrwtFllt_bio_USA}, \code{L2261.SubsectorInterp_bio_USA}, \code{L2261.SubsectorLogit_bio_USA},
\code{L2261.StubTech_bio_USA}, \code{L2261.StubTechMarket_bio_USA}, \code{L2261.StubTechShrwt_rbO_USA},
\code{L2261.StubTechFractSecOut_bio_USA}, \code{L2261.StubTechFractProd_bio_USA}, \code{L2261.Rsrc_DDGS_USA},
\code{L2261.RsrcPrice_DDGS_USA}, \code{L2261.Tech_rbm_USA}, \code{L2261.TechShrwt_rbm_USA},
\code{L2261.TechCoef_rbm_USA}, \code{L2261.Tech_dbm_USA}, \code{L2261.TechShrwt_dbm_USA},
\code{L2261.TechEff_dbm_USA}, \code{L2261.TechCost_dbm_USA}, \code{L2261.CarbonCoef_bio_USA},
\code{L2261.StubTechMarket_en_USA}, \code{L2261.StubTechMarket_elecS_USA}, \code{L2261.StubTechMarket_ind_USA},
\code{L2261.StubTechMarket_cement_USA}, \code{L2261.StubTechMarket_bld_USA}.
The corresponding file in the original data system was \code{L2261.regional_biomass_USA.R} (gcam-usa level2).
}
\description{
Create biomass supply sectors at the state level.
}
\details{
Create biomass supply sectors at the state level, in order ensure that biomass carbon-tracking is
contained entirely within the consuming region (state).
}
\author{
MTB Aug 2018
}
