% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1236.elec_load_segments_solver_USA.R
\name{disabled_module_gcamusa_LB1236.elec_load_segments_solver_USA}
\alias{disabled_module_gcamusa_LB1236.elec_load_segments_solver_USA}
\title{module_gcamusa_LB1236.elec_load_segments_solver_USA}
\usage{
disabled_module_gcamusa_LB1236.elec_load_segments_solver_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1236.grid_elec_supply_USA}.

The corresponding file in the original data system was \code{LB1236.elec_load_segments_solver_2010.R} (gcam-usa level1).
}
\description{
Calculate the fraction of electricity generation by fuel by horizontal load segment such that the total supply
of electricity in each grid region matches total demand of electricity in that grid region.
}
\details{
Calculates the fraction of electricity generation by fuel, by horizontal load segment, by grid region, in 2010.
}
\author{
MTB August 2018
}
