% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2233.electricity_water.R
\name{disabled_module_water_L2233.electricity_water}
\alias{disabled_module_water_L2233.electricity_water}
\title{module_water_L2233.electricity_water}
\usage{
disabled_module_water_L2233.electricity_water(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2233.StubTech_elecPassthru}, \code{L2233.StubTechProd_elecPassthru},
\code{L2233.GlobalPassThroughTech}, \code{L2233.GlobalTechEff_elecPassthru},
 \code{L2233.GlobalTechShrwt_elecPassthru}, \code{L2233.GlobalIntTechCapital_elec},
 \code{L2233.GlobalTechCapital_elecPassthru}, \code{L2233.GlobalIntTechOMfixed_elec},
\code{L2233.GlobalTechOMfixed_elecPassthru}, \code{L2233.GlobalIntTechOMvar_elec},
 \code{L2233.GlobalTechOMvar_elecPassthru}, \code{L2233.GlobalTechInterp_elecPassthru},
 \code{L2233.PassThroughSector_elec_cool},
\code{L2233.Supplysector_elec_cool}, \code{L2233.ElecReserve_elec_cool},
\code{L2233.SubsectorShrwtFllt_elec_cool}, \code{L2233.SubsectorLogit_elec_cool},
 \code{L2233.StubTech_elec_cool}, \code{L2233.StubTechEff_elec_cool},
\code{L2233.StubTechProd_elec_cool}, \code{L2233.StubTechFixOut_hydro},
  \code{L2233.StubTechShrwt_elec_cool}, \code{L2233.GlobalTechCapital_elec_cool},
 \code{L2233.GlobalIntTechCapital_elec_cool}, \code{L2233.GlobalTechCoef_elec_cool},
\code{L2233.GlobalIntTechCoef_elec_cool}, \code{L2233.InputEmissCoeff_hist_elecPassthru},
 \code{L2233.InputEmissCoeff_fut_elecPassthru}, \code{L2233.AvgFossilEffKeyword_elec_cool},
 \code{L2233.DeleteCreditInput_elec}, \code{L2233.CreditInput_elec}.


The corresponding file in the
original data system was \code{L2233.electricity_water.R} (water level2).
}
\description{
Generates GCAM model inputs for electricity sector with cooling system types disaggregated.
}
\details{
Disaggregates electricity sector for all cooling system types.
}
\author{
ST June 2017
}
