% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakout_helpers.R
\name{downscale_to_breakout_regions}
\alias{downscale_to_breakout_regions}
\title{downscale_to_breakout_regions}
\usage{
downscale_to_breakout_regions(
  data = NULL,
  composite_region = NULL,
  disag_regions = NULL,
  share_data = NULL,
  value.column = NULL,
  share.column = NULL,
  ndigits = energy.DIGITS_CALOUTPUT
)
}
\arguments{
\item{data}{Data frame with a "region" column}

\item{composite_region}{Composite (larger) region that data are being downscaled from (e.g. "Thailand")}

\item{disag_regions}{Disaggregated regions that data are being assigned to (e.g. c("Bangkok", "Rest of Thailand"))}

\item{share_data}{Data frame with the disaggregated regions' shares to use for downscaling}

\item{value.column}{Name of the column in data to be downscaled (e.g. "value")}

\item{share.column}{Name of the column in share_data to use for the downscaling (e.g. "popshare", "gdpshare")}
}
\value{
Table with both regions and values that are properly assigned
}
\description{
Downscale (apportion) Thailand data to disaggregated Bangkok and Rest of Thailand regions
}
\details{
Uses some share proxy to apportion some physical quantity from the composite
region to its constituent disaggregated regions. The default is to use population shares.
}
