% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{driver}
\alias{driver}
\title{driver}
\usage{
driver(all_data = empty_data(), stop_before = NULL,
  stop_after = NULL, return_inputs_of = stop_before,
  return_outputs_of = stop_after,
  return_data_names = union(inputs_of(return_inputs_of),
  outputs_of(return_outputs_of)), return_data_map_only = FALSE,
  write_outputs = !return_data_map_only, write_xml = write_outputs,
  outdir = OUTPUTS_DIR, xmldir = XML_DIR, quiet = FALSE)
}
\arguments{
\item{all_data}{Data to be pre-loaded into data system}

\item{stop_before}{Stop immediately before this chunk (character)}

\item{stop_after}{Stop immediately after this chunk  (character)}

\item{return_inputs_of}{Return the data objects that are inputs for these chunks (character).
If \code{stop_before} is specified, by default that chunk's inputs are returned}

\item{return_outputs_of}{Return the data objects that are output from these chunks (character)
If \code{stop_after} is specified, by default that chunk's outputs are returned}

\item{return_data_names}{Return these data objects (character). By default this is the union of \code{return_inputs_of} and \code{return_outputs_of}}

\item{return_data_map_only}{Return only the precursor information? (logical) This overrides
the other \code{return_*} parameters above}

\item{write_outputs}{Write all chunk outputs to disk?}

\item{write_xml}{Write XML Batch chunk outputs to disk?}

\item{outdir}{Location to write output data (ignored if \code{write_outputs} is \code{FALSE})}

\item{xmldir}{Location to write output XML (ignored if \code{write_outputs} is \code{FALSE})}

\item{quiet}{Suppress output?}
}
\value{
A list of all built data (or a data map tibble if requested).
}
\description{
Run the entire data system.
}
\details{
The driver loads any necessary data from input files,
runs all code chunks in an order dictated by their dependencies,
does error-checking, and writes outputs. For more details, see
the relevant wiki page at \url{ https://github.com/bpbond/gcamdata/wiki/Driver}.
}
\author{
BBL
}
