% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstrace.R
\name{dstrace_chunks}
\alias{dstrace_chunks}
\title{A simplified dstrace which simply finds the chunks which are precursors
recursively.}
\usage{
dstrace_chunks(chunk_names, gcam_data_map, prev_trace_list = c())
}
\arguments{
\item{chunk_names}{A list of chunks to find the precursors for.}

\item{gcam_data_map}{A tibble of metadata information; normally a built-in package dataset.}

\item{prev_trace_list}{The list of chunks already found.}
}
\value{
The unique list of chunk names which are precursors for \code{chunk_names}.
}
\description{
The algorithm is performing a depth first search where chunks are "nodes"
and data objects are "verticies" connecting them.
}
