% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{expand_by_region}
\alias{expand_by_region}
\title{expand_by_region}
\usage{
expand_by_region(data, region_map)
}
\arguments{
\item{data}{data}

\item{region_map}{region_map with 2 columns, a market column and region column}
}
\value{
assignment of new db
}
\description{
Filter policy tables to desired xml
}
