% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{expand_to_segments}
\alias{expand_to_segments}
\title{expand_to_segments}
\usage{
expand_to_segments(df, sector = "supplysector", group_by_cols, segments)
}
\arguments{
\item{sector}{sector name to match}

\item{group_by_cols}{columns to group by for each segment}

\item{segments}{names of segments}
}
\value{
new tibble
}
\description{
Helper function to replace normal electricity sector with segments
}
