% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_header_info}
\alias{extract_header_info}
\title{extract_header_info}
\usage{
extract_header_info(
  header_lines,
  label,
  filename,
  required = FALSE,
  multiline = FALSE
)
}
\arguments{
\item{header_lines}{Character vector holding raw header lines}

\item{label}{Label to search for - character}

\item{filename}{Filename (for error reporting purposes)}

\item{required}{Is this label required? (logical)}

\item{multiline}{Can this label hold multi-line information? (logical)}
}
\value{
Extracted label information, as a character vector
}
\description{
Extract information from CSV headers.
}
\details{
CSV files can have headers, commented lines of the form "# Title: xxxx",
"# Source: xxxx", etc. Extract this information if present. Note that empty headers
are not allowed. This function is called by \code{\link{parse_csv_header}}.
}
