% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolators.R
\name{extrapolate_constant_optionB}
\alias{extrapolate_constant_optionB}
\title{extrapolate_constant_optionB}
\usage{
extrapolate_constant_optionB(x, n = 1)
}
\arguments{
\item{x}{Tibble with column name 'value' that contains NA values to be filled in
via constant extrapolation of the mean of last n non-NA values.}

\item{n}{Number of non-NA values to be averaged to provide the filler value.}
}
\value{
Tibble with column name 'value' in which all NA values have been filled
in with the specified mean value.
}
\description{
computes the mean of last n values and uses this constant value to fill in
 missing years at the end of the time series
}
\details{
Computes the mean of last n non-NA values of column 'value' in an input tibble
(may be grouped)  and uses this constant value to fill in NA values in column 'value'.
}
\author{
ACS June 2019
}
