% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{FAO_AREA_DISAGGREGATE_HIST_DISSOLUTION}
\alias{FAO_AREA_DISAGGREGATE_HIST_DISSOLUTION}
\title{FAO_AREA_DISAGGREGATE_HIST_DISSOLUTION}
\usage{
FAO_AREA_DISAGGREGATE_HIST_DISSOLUTION(
  .DF,
  AFFECTED_AREA_CODE,
  YEAR_DISSOLVE_DONE,
  YEAR_AFTER_DISSOLVE_ACCOUNT = 3
)
}
\arguments{
\item{.DF}{dataframe to disaggregate}

\item{AFFECTED_AREA_CODE}{FAO area codes for regions affected; first one should be pre-dissolved region (e.g., USSR) followed by post-dissolved regions.}

\item{YEAR_DISSOLVE_DONE}{First year after dissolution}

\item{YEAR_AFTER_DISSOLVE_ACCOUNT}{Number of years of data after dissolution used for sharing historical data}
}
\value{
Disaggregated data for the historical period for of the dissolved region
}
\description{
FAO_AREA_DISAGGREGATE_HIST_DISSOLUTION
}
