% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{FAO_AREA_DISAGGREGATE_HIST_DISSOLUTION_ALL}
\alias{FAO_AREA_DISAGGREGATE_HIST_DISSOLUTION_ALL}
\title{FAO_AREA_DISAGGREGATE_HIST_DISSOLUTION_ALL}
\usage{
FAO_AREA_DISAGGREGATE_HIST_DISSOLUTION_ALL(
  .DF,
  SUDAN2012_BREAK = F,
  SUDAN2012_MERGE = T
)
}
\arguments{
\item{SUDAN2012_BREAK}{If T break Sudan before 2012 based on 2013- 2016 data}

\item{SUDAN2012_MERGE}{If T merge South Sudan into Sudan}
}
\value{
data with historical periods of dissolved region disaggregated to small pieces.
}
\description{
FAO_AREA_DISAGGREGATE_HIST_DISSOLUTION_ALL
}
