% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{fill_exp_decay_extrapolate}
\alias{fill_exp_decay_extrapolate}
\title{fill_exp_decay_extrapolate}
\usage{
fill_exp_decay_extrapolate(d, out_years)
}
\arguments{
\item{d}{The wide format tibble with values under year columns that will be filled
to ensure the given years are present using the rules mentioned above.}

\item{out_years}{A vector of integers specifying which years must have values in the
output data.}
}
\value{
The filled out data set in long format.  The years will be in the \code{year}
column and will include all values in \code{out_years} and the filled in values will
be in the \code{value} column.  All extrapolation parameters will be cleaned out.
}
\description{
Takes a wide format tibble with years as columns, coverts to long format, and
ensures values are filled in for all \code{out_years} using the following rules:
  - Linearly interpolated for missing values that have end points
  - Extrapolated using an exponential decay function paramaterized by the columns
    \code{improvement.rate} and \code{improvement.max} using the following formulation
    \code{v_0*max+(v_0-v_0*max)*(1-rate)^(y-y_0)}
  - For rows that specify a char value in the column \code{improvement.shadow.technology}
    exponential decay will be calculated on the difference between the values calculated
    by left joining with itself on the column \code{improvement.shadow.technology} with
    the column \code{technology}.  In other words for shadowing technologies the decay is
    only applied to the difference in the values in the last year in which one was
    specified. This is to allow for instance a Gas CC plant to have cost reductions at a
    moderate pace but a Gas CC+CCS can have rapid cost reductions to the CCS portion of
    the cost.
}
\author{
Pralit Patel
}
