% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{filter_regions_europe}
\alias{filter_regions_europe}
\title{filter_regions_europe}
\usage{
filter_regions_europe(
  df,
  regions_to_keep_name = gcameurope.EUROSTAT_COUNTRIES,
  regions_to_keep_iso = gcameurope.EUROSTAT_ISO,
  region_ID_mapping = NULL,
  inverse = FALSE
)
}
\arguments{
\item{df}{tibble; ideally with a region/market.name columns}

\item{regions_to_keep_name}{character vector of regions to keep}

\item{regions_to_keep_iso}{character vector of regions to keep}

\item{region_ID_mapping}{mapping file between region NAMES and GCAM_region_ID}

\item{inverse}{if TRUE, remove all european regions}
}
\value{
tibble without specified regions
}
\description{
Helper function to filter to regions from a tibble. If you want to filter by ID,
you need to provide the mapping file `region_ID_mapping` yourself, with at least
two columns called `region` and `GCAM_region_ID` (e.g. common/GCAM_region_names).
}
