% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{generate_metadata_csv}
\alias{generate_metadata_csv}
\title{generate_metadata_csv}
\usage{
generate_metadata_csv(
  chunk.info.fn = NULL,
  chunk.info.map.fn = NULL,
  files.info.fn = NULL,
  ALL_CHUNKS = FALSE
)
}
\arguments{
\item{chunk.info.fn}{The file name to write the chunk info CSV to, if not null.}

\item{chunk.info.map.fn}{The file name to write the chunk info map CSV to, if not null.}

\item{files.info.fn}{The file name to write the file info CSV to, if not null.}

\item{ALL_CHUNKS}{If false omits batch XML chunks from the chunk info CSVs.}
}
\description{
A tibble of chunk information and chunk information map and file info. Uses chunk
documentation to generate a tibble of chunk names, module, their dependencies,
and their documentation.
}
\details{
Uses \code{\link{find_chunks}} and \code{\link{chunks_info}} from \code{utils.R}.
\code{GENENRATE_CHUNKS_INFO} option is set to \code{TRUE} and
\code{ALL_CHUNKS} option is set to \code{FALSE} by default. This means that
the tibble will only contain non-batch/non-xml chunks because batch and xml
chunks have straightforward documentation and are not needed in the tibble.
}
\author{
Hassan Niazi, Jan 2024
}
