% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{get_price_year}
\alias{get_price_year}
\title{get_price_year}
\usage{
get_price_year(unit)
}
\arguments{
\item{unit}{String from price_units}

\item{...}{The objects}
}
\value{
String of year
}
\description{
Gets price year from price_units string.
}
