% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_relative_to_workdir}
\alias{get_relative_to_workdir}
\title{Normalize file path to the working directory}
\usage{
get_relative_to_workdir(fqfn)
}
\arguments{
\item{fqfn}{The full path of the file to normalize.}
}
\value{
A relative path to \code{fqfn} from the current working directory
given by \code{getwd()}.
}
\description{
On Windows we may run into the MAX_PATH file path limit when referencing
files using the absolute path.  This method will normalize the file path
to the working directory which _may_ be shorter.
}
