% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{get_ssp_regions}
\alias{get_ssp_regions}
\title{get_ssp_regions}
\usage{
get_ssp_regions(
  pcGDP,
  reg_names,
  income_group,
  ssp_filter = "SSP4",
  year_filter = aglu.PCGDP_YEAR
)
}
\arguments{
\item{pcGDP}{A tibble with per capita GDP estimates, including columns \code{GCAM_region_ID},
\code{scenario}, \code{year}, and \code{value}}

\item{reg_names}{A tibble with columns \code{GCAM_region_ID} and \code{region}}

\item{income_group}{A string indicating which region group (low, medium, high)}

\item{ssp_filter}{A string indicating which SSP to filter to (SSP4 by default)}

\item{year_filter}{An integer indicating which year to use (2010 by default)}
}
\value{
A character vector of region names belonging to the specified income group.
}
\description{
Get regions for different income groups in SSP4 2010 (by default)
}
