% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_xml_names}
\alias{get_xml_names}
\title{get_xml_names}
\usage{
get_xml_names(csv_name, default_xml)
}
\arguments{
\item{csv_name}{CSV file with xml column that contains output xml names}

\item{default_xml}{A default xml to output if no xml names provided for any output}
}
\value{
Vector of xml names
}
\description{
Return list of xml names as output for chunk when they are included in csv file
}
\author{
Russell Horowitz
}
