% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-helpers.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Helper function: select elements not in user defined criteria}
\usage{
a %notin% b
}
\description{
Helper function: select elements not in user defined criteria
}
