% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{GROSS_TRADE_ADJUST}
\alias{GROSS_TRADE_ADJUST}
\title{Balance gross trade}
\usage{
GROSS_TRADE_ADJUST(
  .DF,
  .MIN_TRADE_PROD_RATIO = 0.01,
  .Reg_VAR = "area_code",
  .GROUP_VAR = c("item_code", "year")
)
}
\arguments{
\item{.DF}{An input dataframe with an element col including Import and Export}

\item{.MIN_TRADE_PROD_RATIO}{Trade will be removed if world total export or import over production is smaller than .MIN_TRADE_PROD_RATIO, 0.01 default value}

\item{.Reg_VAR}{Region variable name; default is area_code}

\item{.GROUP_VAR}{Group variable; default is item_code and year}
}
\value{
The same dataframe with balanced world export and import.
}
\description{
Scale gross export and import in all regions to make them equal at the world level.
}
