% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstrace.R
\name{info}
\alias{info}
\title{info}
\usage{
info(
  object_name,
  gcam_data_map = NULL,
  previous_tracelist = NULL,
  upstream = TRUE,
  downstream = TRUE
)
}
\arguments{
\item{object_name}{Name of object to get information about (can be either a data object or a code chunk)}

\item{gcam_data_map}{A tibble of metadata information; normally a built-in package dataset}

\item{previous_tracelist}{Information about previous objects printed (if called from \code{\link{dstrace}})}

\item{upstream}{Print and return upstream precursors? Logical}

\item{downstream}{Print and return downstream dependents? Logical}
}
\value{
If called directly, returns an entry from \code{GCAM_DATA_MAP}; if called from \code{\link{dstrace}}, a two-
element list with linked object and tracelist information. If called and \code{object_name} is the name of a
code chunk (as opposed to that of a data object), the relevant help page will be pulled up.
}
\description{
Print and return information about an object.
}
\examples{
info("L100.FAO_ag_Exp_t")
info("module_aglu_L222.land_input_2")
}
