% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolators.R
\name{last_n}
\alias{last_n}
\title{last_n}
\usage{
last_n(x, n)
}
\arguments{
\item{x}{Vector with some NA values}

\item{n}{The number of non-NA values sought.}
}
\value{
A vector with the last n non-NA values from input
vector x.
}
\description{
finds the last n non-NA values in an input vector.
}
\details{
finds the last n non-NA values in an input vector.
}
\author{
ACS June 2019
}
