% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolators.R
\name{last_n_nonNA}
\alias{last_n_nonNA}
\title{last_n_nonNA}
\usage{
last_n_nonNA(x, n)
}
\arguments{
\item{x}{Vector with some NA values}

\item{n}{The number of non-NA values sought.}
}
\value{
A vector with the last n non-NA values from input
vector x.
}
\description{
finds the last n non-NA values in an input vector.
 A convenience functions for users who wish to customize
 their extrapolations beyond the default or who wish to
 identify NA values in their original (unextrapolated)
 data.
}
\details{
finds the last n non-NA values in an input vector.
}
\author{
ACS June 2019
}
