% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-helpers.R
\name{left_join_error_no_match}
\alias{left_join_error_no_match}
\title{left_join_error_no_match}
\usage{
left_join_error_no_match(d, ..., ignore_columns = NULL)
}
\arguments{
\item{d}{Data frame (typically from pipeline)}

\item{...}{Rest of call to \code{\link{left_join}}}

\item{ignore_columns}{Optional column name(s) to ignore, character vector}
}
\value{
Joined data.
}
\description{
A restrictive version of \code{\link{left_join}}.
}
\details{
Restrictive version of dplyr::left_join meant for replacing `match` calls.
}
