% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-helpers.R
\name{left_join_strict}
\alias{left_join_strict}
\title{left_join_strict}
\usage{
left_join_strict(left_df, right_df, ..., ignore_columns = NULL)
}
\arguments{
\item{left_df}{A data frame. The left dataset in the join.}

\item{right_df}{A data frame. The right dataset in the join.}

\item{...}{Rest of call to \code{\link{left_join}}}

\item{ignore_columns}{Optional column name(s) to ignore, character vector}
}
\value{
A data frame resulting from the left join. If any rows in `left_df` do not have matching keys in `right_df`, an error is thrown.
}
\description{
A restrictive version of \code{\link{left_join}} that ensures that all keys
in the left dataset have corresponding matches in the right dataset.
If any rows in the left dataset do not have matching keys in the right dataset,
the function will throw an error. However, it will NOT through an error if
the number of lines is modified.
}
