% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_csv_files}
\alias{load_csv_files}
\title{load_csv_files}
\usage{
load_csv_files(filenames, optionals, quiet = FALSE, dummy = NULL, ...)
}
\arguments{
\item{filenames}{Character vector of filenames to load}

\item{optionals}{Logical vector, specifying whether corresponding file is optional}

\item{quiet}{Logical - suppress messages?}

\item{dummy}{Not used, here as a hack for drake}

\item{...}{Any other parameter to pass to \code{readr::read_csv}}
}
\value{
A list of data frames (tibbles).
}
\description{
Load one or more internal, i.e. included with the package, csv (or csv.gz) data files.
}
\details{
The data frames read in are marked as inputs, not ones that have
been computed, via \code{\link{add_comments}}. Optional files that are not found
as returned as NA in the list.
}
