% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_from_cache}
\alias{load_from_cache}
\title{load_from_cache}
\usage{
load_from_cache(return_data_names, ...)
}
\arguments{
\item{return_data_names}{The names of the data to return.}

\item{...}{Parameters to pass directly to drake::readd to allow
users to specify a non-default cache for instance.}
}
\value{
The data loaded from cache returned as a list in the
same format as all_data.
}
\description{
Load the given chunk inputs / outputs from the drake cache.  The
names given can come from \code{inputs_of} or \code{outputs_of} for
instance.
}
