% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{map_province_name}
\alias{map_province_name}
\title{map_province_name}
\usage{
map_province_name(data, mapping, map_names, replace = FALSE)
}
\arguments{
\item{data}{Data, tibble}

\item{mapping}{Table of alternative province names, tibble}

\item{map_names}{Name of new province name columns in \code{mapping}, character}

\item{replace}{Replace old province name to new name columns? Logical}
}
\value{
Data with new province name columns added or replaced.
}
\description{
Map in alternative province names, optionally dropping the old name
}
