% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2013_ag_yield_2015.R
\name{module_aglu_ag_yield_2015}
\alias{module_aglu_ag_yield_2015}
\title{module_aglu_ag_yield_2015}
\usage{
module_aglu_ag_yield_2015(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2013.AgProdChange_2015}.
}
\description{
Creates marginal abatement cost curves "MACC", for fossil resources, agriculture, animals, and processing.
}
\details{
Creates marginal abatement cost curves "MACC", for fossil resources, agriculture, animals, and processing.
}
\author{
RH August 2017
}
