% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_L100.FAO_SUA_connection.R
\name{module_aglu_L100.FAO_SUA_connection}
\alias{module_aglu_L100.FAO_SUA_connection}
\title{module_aglu_L100.FAO_SUA_connection}
\usage{
module_aglu_L100.FAO_SUA_connection(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{FAO_SUA_APE_balance}
}
\description{
Pull and further process SUA data needed
}
\details{
Pull and further process SUA data needed. Calculate moving average if needed.
}
\author{
XZ 2022
}
