% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_L100.IMAGE_downscale_ctry_yr.R
\name{module_aglu_L100.IMAGE_downscale_ctry_yr}
\alias{module_aglu_L100.IMAGE_downscale_ctry_yr}
\title{module_aglu_L100.IMAGE_downscale_ctry_yr}
\usage{
module_aglu_L100.IMAGE_downscale_ctry_yr(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.IMAGE_an_Feedfrac_ctry_C_Sys_Fd_Y}, \code{L100.IMAGE_an_FeedIO_ctry_C_Sys_Y}, \code{L100.IMAGE_an_Prodmixfrac_ctry_C_Y}. The corresponding file in the
original data system was \code{LA100.IMAGE_downscale_ctry_yr.R} (aglu level1).
}
\description{
Extrapolate IMAGE data to all AGLU historical years and countries.
}
\details{
Each IMAGE table is extrapolated to all AGLU historical years and then downscaled from IMAGE region to all AGLU countries.
}
\author{
BBL June 2017 XZ 2022
}
