% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_L100.regional_ag_an_for_prices.R
\name{module_aglu_L100.regional_ag_an_for_prices}
\alias{module_aglu_L100.regional_ag_an_for_prices}
\title{module_aglu_L100.regional_ag_an_for_prices}
\usage{
module_aglu_L100.regional_ag_an_for_prices(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L1321.ag_prP_R_C_75USDkg},
  \code{L1321.an_prP_R_C_75USDkg}, \code{L1321.expP_R_F_75USDm3}
}
\description{
Calculate the calibration prices for all GCAM AGLU commodities.
}
\details{
This chunk calculates average prices over calibration years by GCAM commodity and region. Averages across
  years, when applicable, are unweighted; averages over FAO item are weighted by production.
}
\author{
GPK/RC/STW February 2019; XZ 2022
}
