% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_L101.ag_FAO_R_C_Y.R
\name{module_aglu_L101.ag_FAO_R_C_Y}
\alias{module_aglu_L101.ag_FAO_R_C_Y}
\title{module_aglu_L101.ag_FAO_R_C_Y}
\usage{
module_aglu_L101.ag_FAO_R_C_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L101.ag_HA_bm2_R_C_Y}, \code{L101.ag_Prod_Mt_R_C_Y}. The corresponding file in the
original data system was \code{LA101.ag_FAO_R_C_Y.R} (aglu level1).
}
\description{
Aggregate FAO food consumption, ag production, and harvested area
data to GCAM regions and GCAM commodities.
}
\details{
This chunk aggregates FAO food consumption, agricultural production,
and harvested area data up to GCAM commodities and GCAM regions. Data is converted
from FAO units (tons, hectares) to GCAM units (Mt, Pcal, billion km2). Note that
FAO's alfalfa production in the USA is divided by 4 "for consistency with USDA".
Note (August 2018 GPK revision) - The FAO production and harvested area are disaggregated
to basin PRIOR to aggregation by GCAM region. This reduces the bias from using a single
year (around 2000) to disaggregate to basin, in multi-country regions.
XZ 03-2022 Food related processing was moved so this chunk only downscales prod and area to GLU level ----
}
\author{
KVC March 2017 (revised August 2018 by GPK); XZ 2022
}
