% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_L112.ag_prodchange_R_C_Y_GLU.R
\name{module_aglu_L112.ag_prodchange_R_C_Y_GLU}
\alias{module_aglu_L112.ag_prodchange_R_C_Y_GLU}
\title{module_aglu_L112.ag_prodchange_R_C_Y_GLU}
\usage{
module_aglu_L112.ag_prodchange_R_C_Y_GLU(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L112.ag_YieldRatio_R_C_Ysy_GLU}, \code{L112.ag_YieldRate_R_C_Y_GLU}, \code{L112.bio_YieldRate_R_Y_GLU}. The corresponding file in the
original data system was \code{LB112.ag_prodchange_R_C_Y_GLU.R} (aglu level1).
}
\description{
Calculate yield change ratios and yield change rates by GCAM commodity / region / future years.
}
\details{
This chunk first calculates the yield ratios relative to the base year by GCAM commodity / region / future years to 2050,
and calculates the annual productivity change rates by GCAM commodity / region / future years to 2100.
Yield improvement rates are based on FAO estimates up to 2050 and the default agriculture productivity change assumptions beyond 2050.
}
\author{
RC April 2017
}
