% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_L121.Carbon_LT.R
\name{module_aglu_L121.Carbon_LT}
\alias{module_aglu_L121.Carbon_LT}
\title{module_aglu_L121.Carbon_LT}
\usage{
module_aglu_L121.Carbon_LT(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L121.CarbonContent_kgm2_R_LT_GLU}, \code{L121.Yield_kgm2_R_Past_GLU}. The corresponding file in the
original data system was \code{LB121.Carbon_LT.R} (aglu level1).
}
\description{
Compute natural vegetation, managed land, and pasture carbon density, mature age, and yield.
}
\details{
Combine Houghton et al. (1999 and similar) carbon density, mature age, and pasture yield data with
SAGE data to produce tables giving, for each GCAM region and land use type, data on these variables.
}
\author{
BBL April 2017
}
