% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_L123.LC_R_MgdPastFor_Yh_GLU.R
\name{module_aglu_L123.LC_R_MgdPastFor_Yh_GLU}
\alias{module_aglu_L123.LC_R_MgdPastFor_Yh_GLU}
\title{module_aglu_L123.LC_R_MgdPastFor_Yh_GLU}
\usage{
module_aglu_L123.LC_R_MgdPastFor_Yh_GLU(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L123.ag_Prod_Mt_R_Past_Y_GLU}, \code{L123.ag_Yield_kgm2_R_Past_Y_GLU}, \code{L123.LC_bm2_R_MgdPast_Yh_GLU}, \code{L123.For_Prod_bm3_R_Y_GLU}, \code{L123.For_Yield_m3m2_R_GLU}, \code{L123.LC_bm2_R_MgdFor_Yh_GLU}. The corresponding file in the
original data system was \code{LB123.LC_R_MgdPastFor_Yh_GLU.R} (aglu level1).
}
\description{
Calculates pasture and forests production, yields and managed land cover by GCAM region / GLU / year.
}
\details{
This chunk calculates pasture and forests production and yields by GCAM region / GLU / year,
and the managed pasture and forests land cover by GCAM region / GLU / historical year.
Managed land area for both are adjusted below the threshold percentage of total pasture and forest land,
and yields are increased so that production stays unchanged for those regions.
}
\author{
RC August 2017
}
