% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_L125.LC_tot.R
\name{module_aglu_L125.LC_tot}
\alias{module_aglu_L125.LC_tot}
\title{module_aglu_L125.LC_tot}
\usage{
module_aglu_L125.LC_tot(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L125.LC_bm2_R}, \code{L125.LC_bm2_R_GLU}, \code{L125.LC_bm2_R_LT_Yh_GLU}.
The corresponding file in the
original data system was \code{LB125.LC_tot.R} (aglu level1).
}
\description{
Build Total Land Cover by GCAM region, and by GCAM region and GLU, and by GCAM region/GLU/year.
}
\details{
This module builds three total land cover area data from the lower-level raw data:
1) total land area by GCAM region; 2) total land area by GCAM region and GLU; and
3) total land area by GCAM region, GLU and historical year. Units of the data are billion square meters.
The land area changing rates (bm2 per year) are checked to make sure they are within certain tolerances.
}
\author{
MC May 2017
}
