% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_L142.ag_Fert_IO_R_C_Y_GLU.R
\name{module_aglu_L142.ag_Fert_IO_R_C_Y_GLU}
\alias{module_aglu_L142.ag_Fert_IO_R_C_Y_GLU}
\title{module_aglu_L142.ag_Fert_IO_R_C_Y_GLU}
\usage{
module_aglu_L142.ag_Fert_IO_R_C_Y_GLU(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L142.ag_Fert_Prod_MtN_ctry_Y}, \code{L142.ag_Fert_NetExp_MtN_R_Y}, \code{L142.ag_Fert_IO_R_C_Y_GLU}. The corresponding file in the
original data system was \code{LB142.ag_Fert_IO_R_C_Y_GLU.R} (aglu level1).
}
\description{
Calculate the adjusted fertilizer production by country / year, fertilizer net exports by GCAM region / year,
and fertilizer input-output coefficients by GCAM region / commodity / year / GLU.
}
\details{
This chunk calculates fertilizer production by country / year (adjusted to global total consumption),
fertilizer net exports by GCAM region / year as production minus consumption, and fertilizer input-output coefficients
by GCAM region / commodity / year / GLU.
}
\author{
RC June 2017
}
