% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L203.ag_an_demand_input.R
\name{module_aglu_L203.ag_an_demand_input}
\alias{module_aglu_L203.ag_an_demand_input}
\title{module_aglu_L203.ag_an_demand_input}
\usage{
module_aglu_L203.ag_an_demand_input(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L203.Supplysector_demand}, \code{L203.NestingSubsectorAll_demand_food},
  \code{L203.SubsectorAll_demand_food}, \code{L203.SubsectorAll_demand_nonfood},
  \code{L203.StubTech_demand_food}, \code{L203.StubTech_demand_nonfood}, \code{L203.SubregionalShares},
  \code{L203.DemandFunction_food}, \code{L203.DemandStapleParams}, \code{L203.DemandNonStapleParams},
  \code{L203.DemandStapleRegBias}, \code{L203.DemandNonStapleRegBias}, \code{L203.StapleBaseService},
  \code{L203.NonStapleBaseService}, \code{L203.GlobalTechCoef_demand}, \code{L203.GlobalTechShrwt_demand}, \code{L203.GlobalTechInterp_demand},
  \code{L203.StubTechProd_food}, \code{L203.StubTechProd_nonfood_crop}, \code{L203.StubTechProd_nonfood_meat},
  \code{L203.StubTechProd_For}, \code{L203.StubCalorieContent},
  \code{L203.PerCapitaBased}, \code{L203.BaseService}, \code{L203.IncomeElasticity}, \code{L203.PriceElasticity}. The
  corresponding file in the original data system was \code{L203.demand_input.R} (aglu level2).
}
\description{
Builds agriculture demand inputs for the core and each SSP scenario.
}
\details{
This chunk specifies the input tables for agriculture demand: generic information for supply sector, subsector and technology,
food and non-food demand in calibration years, forest product demand, net exports and caloric contents in calibration and future years,
income elasticities for future years in core and SSP scenarios, as well as price elasticities.
}
\author{
RC July 2017
}
