% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2042.resbio_input_irr_mgmt.R
\name{module_aglu_L2042.resbio_input_irr_mgmt}
\alias{module_aglu_L2042.resbio_input_irr_mgmt}
\title{module_aglu_L2042.resbio_input_irr_mgmt}
\usage{
module_aglu_L2042.resbio_input_irr_mgmt(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2042.AgResBio_For}, \code{L2042.AgResBioCurve_For}, \code{L2042.GlobalResBio_Mill}, \code{L2042.StubResBioCurve_Mill}, \code{L2042.AgResBio_ag_irr_mgmt}, \code{L2042.AgResBioCurve_ag_irr_mgmt}. The corresponding file in the
original data system was \code{L2042.resbio_input_irr_mgmt.R} (aglu level2).
}
\description{
Produce a table of global Mill Residue Biomass Paramters by year, a table of regional Forest Residue Biomass Paramters by year, and a table of
Agricultural Residue Biomass Paramters by irrigation-management level-year. This chunk also produces tables by region and year of residue biomass supply
curves for Mill, Forest, and Agriculture.
}
\details{
For Forest Residue Biomass, each GCAM region-commodity-GLU combination in L123.For_Prod_bm3_R_Y_GLU is converted to a
region-supplySector-supplySubsector-ProductionTech combination, and externally set parameters are added to form the table of Forest
residue biomass parameters for each region and year. For each region-supplySector-supplySubsector-ProductionTech combination in
this table, base supply curves are read in from A_resbio_curves and then, in specified calibration years BASE_YEARS, replaced by
fractions in  A_bio_frac_prod_R to form the table of Forest resbio supply curves for each region and year.

For Mill Residue Biomass, sector, subsector, and technology combinations for NonFoodDeman_Forest are pulled from A_demand_technology,
and externally set parameters are added to form the table of global Mill residue biomass parameters in each year. Sector, subsector,
and technology combinations from this table are repeated for each GCAM region and year, base supply curves are read in from
A_resbio_curves and then, in specified calibration years BASE_YEARS, replaced by fractions in  A_bio_frac_prod_R to form the table of
Mill resbio supply curves for each region and year.

For Agriculture Residue Biomass, each GCAM region-commodity-GLU combination in L103.ag_Prod_Mt_R_C_Y_GLU is converted to a
region-supplySector-supplySubsector-ProductionTech combination, and parameters from L111.ag_resbio_R_C are joined and rounded
for each region-supplySector. These parameters are then repeated for model years, irrigation types, and management technology levels.
Finally, for each region-supplySector-supplySubsector-ProductionTech-irrigation-tech combination in this table, base
supply curves are read in from A_resbio_curves to form the table of Agriculture resbio supply curves for each region and year.
}
\author{
ACS July 2017
}
