% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_L2052.ag_prodchange_cost_irr_mgmt.R
\name{module_aglu_L2052.ag_prodchange_cost_irr_mgmt}
\alias{module_aglu_L2052.ag_prodchange_cost_irr_mgmt}
\title{module_aglu_L2052.ag_prodchange_cost_irr_mgmt}
\usage{
module_aglu_L2052.ag_prodchange_cost_irr_mgmt(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2052.AgCost_ag_irr_mgmt}, \code{L2052.AgCost_bio_irr_mgmt}, \code{L2052.AgCost_For}, \code{L2052.AgProdChange_ag_irr_ref}, \code{L2052.AgProdChange_bio_irr_ref}, \code{L2052.AgProdChange_irr_high}, \code{L2052.AgProdChange_irr_low}, \code{L2052.AgProdChange_irr_ssp4}. The corresponding file in the
original data system was \code{L2052.ag_prodchange_cost_irr_mgmt.R} (aglu level2).
}
\description{
Specify production costs and future agricultural productivity changes for all technologies.
}
\details{
This chunk maps the production costs of crops, biomass and forest to all four technologies (irrigated / rainfed; high / low),
and calculates future productivity change of crops and biomass for all technologies along reference, high, low and SSP4 scenarios.
}
\author{
RC July 2017
}
