% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_L2062.ag_Fert_irr_mgmt.R
\name{module_aglu_L2062.ag_Fert_irr_mgmt}
\alias{module_aglu_L2062.ag_Fert_irr_mgmt}
\title{module_aglu_L2062.ag_Fert_irr_mgmt}
\usage{
module_aglu_L2062.ag_Fert_irr_mgmt(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2062.AgCoef_Fert_ag_irr_mgmt}, \code{L2062.AgCoef_Fert_bio_irr_mgmt}, \code{L2062.AgCost_ag_irr_mgmt_adj}, \code{L2062.AgCost_bio_irr_mgmt_adj}. The corresponding file in the
original data system was \code{L2062.ag_Fert_irr_mgmt.R} (aglu level2).
}
\description{
Specifies fertilizer coefficients for all technologies; adjusts nonLandVariableCost to remove fertilizer cost.
}
\details{
This chunk maps the fertilizer coefficients calculated in LB142 to all agricultural technologies.
We assume coefficients (in kgN per kgCrop) are equal for all four technologies (irr v rfd; hi v lo).
Adjust nonLandVariableCost to remove the now explicitly computed fertilizer cost.
}
\author{
KVC June 2017
}
