% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2072.ag_water_irr_mgmt.R
\name{module_aglu_L2072.ag_water_irr_mgmt}
\alias{module_aglu_L2072.ag_water_irr_mgmt}
\title{module_aglu_L2072.ag_water_irr_mgmt}
\usage{
module_aglu_L2072.ag_water_irr_mgmt(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L2071.AgCoef_IrrBphysWater_ag_mgmt},
  \code{L2072.AgCoef_IrrWaterWdraw_ag_mgmt}, \code{L2072.AgCoef_IrrWaterCons_ag_mgmt},
  \code{L2072.AgCoef_RfdBphysWater_ag_mgmt}, \code{L2072.AgCoef_BphysWater_bio_mgmt},
  \code{L2072.AgCoef_IrrWaterWdraw_bio_mgmt}, \code{L2072.AgCoef_IrrWaterCons_bio_mgmt},
  \code{L2072.AgNonEnergyCost_IrrWaterWdraw}. The corresponding file in the original data system was
  \code{L2072.ag_water_irr_mgmt.R} (aglu level2).
}
\description{
Calculate agriculture water Input-Output coefficients by region / crop / year / GLU / technology.
}
\details{
This chunk calculates the Input-Output coefficients of irrigation water withdrawals and consumption, and
  biophysical water consumption of irrigated and rainfed crops, for each primary and dedicated bioenergy crop by
  region / year / GLU / management level.
}
\author{
RC July 2017
}
