% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2082.ag_energy_irr_mgmt.R
\name{module_aglu_L2082.ag_energy_irr_mgmt}
\alias{module_aglu_L2082.ag_energy_irr_mgmt}
\title{module_aglu_L2082.ag_energy_irr_mgmt}
\usage{
module_aglu_L2082.ag_energy_irr_mgmt(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2082.AgCoef_en_ag_irr_mgmt}, \code{L2082.AgCoef_en_bio_irr_mgmt}, \code{L2082.AgCost_ag_irr_mgmt_adj}, \code{L2082.AgCost_bio_irr_mgmt_adj}.
}
\description{
Specifies energy (liquid fuel) demand coefficients for all ag technologies; adjusts nonLandVariableCost to remove liquid fuel cost.
}
\details{
This chunk assigns the energy demand quantities energy/industry (L132) to all agricultural technologies.
Adjust nonLandVariableCost to remove the now explicitly computed energy cost.
}
\author{
GPK March 2019
}
