% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2092.ag_P_irr_mgmt.R
\name{module_aglu_L2092.ag_P_irr_mgmt}
\alias{module_aglu_L2092.ag_P_irr_mgmt}
\title{module_aglu_L2092.ag_P_irr_mgmt}
\usage{
module_aglu_L2092.ag_P_irr_mgmt(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2092.AgCoef_P_ag_irr_mgmt,}, \code{L2092.AgCoef_P_bio_irr_mgmt}. No corresponding file in the
original data system
}
\description{
Specifies fertilizer coefficients for all technologies; adjusts nonLandVariableCost to remove fertilizer cost.
}
\details{
This chunk maps the fertilizer coefficients calculated in LB142 to all agricultural technologies.
We assume coefficients (in kgP per kgCrop) are equal for all four technologies (irr v rfd; hi v lo).
Adjust nonLandVariableCost to remove the now explicitly computed fertilizer cost.
}
\author{
MO July 2020
}
