% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_L2253.unmanaged_land_adjust.R
\name{module_aglu_L2253.unmanaged_land_adjust}
\alias{module_aglu_L2253.unmanaged_land_adjust}
\title{module_aglu_L2253.unmanaged_land_adjust}
\usage{
module_aglu_L2253.unmanaged_land_adjust(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L222.LN2_UnmgdAllocation_adj},  \code{L2231.LN3_UnmgdAllocation_adj}.
}
\description{
Adjust unmanaged land based on total land available after removing cropland
}
\author{
RLH November 2023
}
