% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L240.ag_trade.R
\name{module_aglu_L240.ag_trade}
\alias{module_aglu_L240.ag_trade}
\title{module_aglu_L240.ag_trade}
\usage{
module_aglu_L240.ag_trade(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L240.Supplysector_tra},
  \code{L240.SectorUseTrialMarket_tra}, \code{L240.SubsectorAll_tra}, \code{L240.TechShrwt_tra},
  \code{L240.TechCost_tra}, \code{L240.TechCoef_tra}, \code{L240.Production_tra}, \code{L240.Supplysector_reg},
  \code{L240.SubsectorAll_reg}, \code{L240.TechShrwt_reg}, \code{L240.TechCoef_reg}, \code{L240.Production_reg_imp},
  \code{L240.Production_reg_dom}.
}
\description{
Model input for regional and (globally) traded agricultural crops and livestock commodities
}
\details{
Build datasets for ssp4 agricultural trade: food and nonfood trade coefficients, feed trade
coefficients, restricted agricultural trade, and trade regions.
}
\author{
GPK February 2019  XZ March 2020
}
