% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L243.bio_trade_input.R
\name{module_aglu_L243.bio_trade_input}
\alias{module_aglu_L243.bio_trade_input}
\title{module_aglu_L243.bio_trade_input}
\usage{
module_aglu_L243.bio_trade_input(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L243.DeleteInput_RegBio}, \code{L243.SectorLogitTables[[
  curr_table ]]$data}, \code{L243.TechCoef_RegBio}, \code{L243.Supplysector_Bio},
  \code{L243.SectorUseTrialMarket_Bio}, \code{L243.SubsectorLogitTables[[ curr_table ]]$data},
  \code{L243.SubsectorLogit_Bio}, \code{L243.SubsectorShrwt_TotBio}, \code{L243.SubsectorShrwtFllt_TradedBio},
  \code{L243.GlobalTechCoef_TotBio}, \code{L243.GlobalTechShrwt_TotBio}, \code{L243.StubTech_TotBio},
  \code{L243.StubTechShrwt_TotBio}, \code{L243.StubTechCoef_ImportedBio}, \code{L243.StubTechCoef_DomesticBio},
  \code{L243.TechCoef_TradedBio}, \code{L243.TechShrwt_TradedBio}, \code{L243.SubsectorShrwt_TotBio_SSP4},
  \code{L243.SubsectorShrwtFllt_TradedBio_SSP4}, \code{L243.SubsectorShrwt_TotBio_SSP3},
  \code{L243.StubTechShrwt_TotBio_SSP3}, \code{L243.Supplysector_reg_SSP4}, \code{L243.SubsectorInterpTo_tra_SSP4}.
  The corresponding file in the original data system was \code{L243.bio_trade_input.R} (aglu level2).
}
\description{
Establish regionally differentiated trade structure for bioenergy.
}
\details{
This chunk sets up a structure for regionally differentiated bioenergy trade. Each
region consumes a blend of domestic and international bioenergy. Regions can supply to either
the domestic or international market. Share weights in the default case depend on the amount of
cropland a region has, with the largest region having a share weight of 1. Share weights in SSP3
and SSP4 are adjusted to reflect the trade frictions in the SSP storylines.
}
\author{
KVC July 2017
}
