% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA102.ag_LDS_R_C_GLU.R
\name{module_aglu_LA102.ag_LDS_R_C_GLU}
\alias{module_aglu_LA102.ag_LDS_R_C_GLU}
\title{module_aglu_LA102.ag_LDS_R_C_GLU}
\usage{
module_aglu_LA102.ag_LDS_R_C_GLU(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L102.ag_HA_bm2_R_C_GLU}, \code{L102.ag_Prod_Mt_R_C_GLU}. The corresponding file in the
original data system was \code{LA102.ag_LDS_R_C_GLU.R} (aglu level1).
}
\description{
Initializes harvested area and production values using LDS data for each GCAM Region-Commodity-GLU combination.
}
\details{
This module takes LDS initialization data (sourced from HYDE, SAGE and other places) at the
iso country-GLU-GTAP_crop level for both Harvested Area and Agricultural Production, and aggregates to the
GCAM region and commodity level. This module also converts to GCAM model units of billion meters squared (bm2)
for Harvested Area and of Megatons (Mt) for Production.
}
\author{
ACS April 2017
}
