% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1091.ag_GrossTrade.R
\name{module_aglu_LB1091.ag_GrossTrade}
\alias{module_aglu_LB1091.ag_GrossTrade}
\title{module_aglu_LB1091.ag_GrossTrade}
\usage{
module_aglu_LB1091.ag_GrossTrade(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1091.GrossTrade_Mt_R_C_Y} (aglu level1).
}
\description{
Calculate primary agricultural good and animal product mass balances, by region / commodity / year.
}
\details{
This chunk processes the bi-lateral trade flow data matrix from FAOSTAT, in order to differentiate trade of
  agricultural commodities between GCAM regions.
}
\author{
GPK/RC/STW February 2019
}
