% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1251.Peat_R_LT_GLU_Yh.R
\name{module_aglu_LB1251.PeatFrac_R_LT_GLU_Yh}
\alias{module_aglu_LB1251.PeatFrac_R_LT_GLU_Yh}
\title{module_aglu_LB1251.PeatFrac_R_LT_GLU_Yh}
\usage{
module_aglu_LB1251.PeatFrac_R_LT_GLU_Yh(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: (something).
}
\description{
Peatland versus mineral soil fractions by (selected) GCAM region / aggregate land type / historical year / GLU.
}
\details{
Aggregate the \code{aglu/RTI_peatland_data} dataset to GCAM regions and use this to disaggregate land cover
by LUT into peatland versus mineral soil fractions. Returns the fractions and the soil carbon contents of peatlands.
}
\author{
GPK April 2020
}
