% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1321.regional_ag_prices.R
\name{module_aglu_LB1321.regional_ag_prices}
\alias{module_aglu_LB1321.regional_ag_prices}
\title{module_aglu_LB1321.regional_ag_prices}
\usage{
module_aglu_LB1321.regional_ag_prices(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{

}
\description{
Calculate the calibration prices for all GCAM AGLU commodities.
}
\details{
This chunk calculates average prices over calibration years by GCAM commodity and region. Averages across
  years are unweighted; averages over FAO item are weighted by production.
}
\author{
GPK/RC/STW February 2019
}
