% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB133.ag_Costs_USA_C_2005.R
\name{module_aglu_LB133.ag_Costs_USA_C_2005}
\alias{module_aglu_LB133.ag_Costs_USA_C_2005}
\title{module_aglu_LB133.ag_Costs_USA_C_2005}
\usage{
module_aglu_LB133.ag_Costs_USA_C_2005(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L133.ag_Cost_75USDkg_C}. The corresponding file in the
original data system was \code{LB133.ag_Costs_USA_C_2005.R} (aglu level1).
}
\description{
This module computes costs for each GCAM commodity (used by each region-GLU) using USDA cost
information for USA when available. For commodities without USDA cost data, the average profit
among USDA commodities and LDS harvested area and production information are used to compute
cost.
}
\details{
USDA cost information is scaled to the level of GCAM commodity using LDS harvested area and
production data to compute weighting factors for aggregation and convert cost from USD/square meter
to the more useful USD/kg. Cost information is converted to the GCAM base of 1975 USD before aggregation.
For commodities without USDA cost information, the average profit among USDA commodities is combined
with LDS harvested area and production data to calculate cost.
}
\author{
ACS May 2017
}
