% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB134.Diet_FLX.R
\name{module_aglu_LB134.Diet_FLX}
\alias{module_aglu_LB134.Diet_FLX}
\title{module_aglu_LB134.Diet_FLX}
\usage{
module_aglu_LB134.Diet_FLX(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L134.pcFood_kcald_R_Y_FLX}.
}
\description{
Compute future food demand quantities (Staples, NonStaples) for the FLX dietary transition scenario
}
\details{
estimate the future food demand levels by GCAM class (staples, non-staples) in the
Springmann FLX scenario.
}
\author{
BBL May 2017
}
