% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB134.Diet_Rfao.R
\name{module_aglu_LB134.Diet_Rfao}
\alias{module_aglu_LB134.Diet_Rfao}
\title{module_aglu_LB134.Diet_Rfao}
\usage{
module_aglu_LB134.Diet_Rfao(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L134.pcFood_kcald_R_Dmnd_Y}, \code{L134.pcFood_kcald_R_Dmnd_Y_ssp1}, \code{L134.pcFood_kcald_R_Dmnd_Y_ssp2}, \code{L134.pcFood_kcald_R_Dmnd_Y_ssp3}, \code{L134.pcFood_kcald_R_Dmnd_Y_ssp4}, \code{L134.pcFood_kcald_R_Dmnd_Y_ssp5}. The corresponding file in the
original data system was \code{LB134.Diet_Rfao.R} (aglu level1).
}
\description{
Build historical and future time series of per-capita caloric demands.
}
\details{
Build historical time series of per-capita caloric demands;
divide by population to calculate the historical per-capita food demands;
extrapolate this to future periods based on FAO projections;
calculate the future demand ratios by FAO2050 region for each demand type;
aggregate by GCAM demand and FAO region, and compute future diet ratios by FAO2050 region;
extend the projected diets to all years, assuming convergence year and demand levels;
make SSP-specific projections based on GDP changes.
}
\author{
BBL May 2017
}
