% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB143.ag_Fert_IO_ratio_hi_lo.R
\name{module_aglu_LB143.ag_Fert_IO_ratio_hi_lo}
\alias{module_aglu_LB143.ag_Fert_IO_ratio_hi_lo}
\title{module_aglu_LB143.ag_Fert_IO_ratio_hi_lo}
\usage{
module_aglu_LB143.ag_Fert_IO_ratio_hi_lo(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{asdf}.
}
\description{
Calculate the ratio of fertilizer applied to the "hi" technology divided by the "lo" technology, by region / basin / year
}
\details{
This chunk calculates the ratio in the fertilizer input-output coefficients of the "hi-yield" technology divided by the "lo-yield"
technology for all regions, basins, and crops.
}
\author{
Eva December 2020
}
