% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB143.ag_P_IO_R_C_Y_GLU.R
\name{module_aglu_LB143.ag_P_IO_R_C_Y_GLU}
\alias{module_aglu_LB143.ag_P_IO_R_C_Y_GLU}
\title{module_aglu_LB143.ag_P_IO_R_C_Y_GLU}
\usage{
module_aglu_LB143.ag_P_IO_R_C_Y_GLU(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L143.ag_P_IO_cons_R_C_Y_GLU}. No corresponding file in the
original data system
}
\description{
Calculate the P fertilizer input-output coefficients by GCAM region / commodity / year / GLU.
}
\details{
This chunk calculates P fertilizer input-output coefficients
by GCAM region / commodity / year / GLU.
}
\author{
RC June 2017
}
