% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB152.ag_GTAP_R_C_GLU_irr.R
\name{module_aglu_LB152.ag_GTAP_R_C_GLU_irr}
\alias{module_aglu_LB152.ag_GTAP_R_C_GLU_irr}
\title{module_aglu_LB152.ag_GTAP_R_C_GLU_irr}
\usage{
module_aglu_LB152.ag_GTAP_R_C_GLU_irr(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L152.ag_irrHA_bm2_R_C_GLU}, \code{L152.ag_rfdHA_bm2_R_C_GLU}, \code{L152.ag_irrProd_Mt_R_C_GLU}, \code{L152.ag_rfdProd_Mt_R_C_GLU}. The corresponding file in the
original data system was \code{LB152.ag_GTAP_R_C_GLU_irr.R} (aglu level1).
}
\description{
Aggregate the irrigated/rainfed harvest area and production data by GCAM region / commodity / GLU.
}
\details{
This chunk aggregates the irrigated/rainfed harvest area and production data from country and GTAP crop to GCAM region and commodity by each GLU.
}
\author{
RC May 2017
}
