% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB164.ag_Costs_USA_C_2005_irr.R
\name{module_aglu_LB164.ag_Costs_USA_C_2005_irr}
\alias{module_aglu_LB164.ag_Costs_USA_C_2005_irr}
\title{module_aglu_LB164.ag_Costs_USA_C_2005_irr}
\usage{
module_aglu_LB164.ag_Costs_USA_C_2005_irr(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L164.ag_Cost_75USDkg_C}. The corresponding file in the
original data system was \code{LB164.ag_Costs_USA_C_2005_irr.R} (aglu level1).
}
\description{
This module calculates production costs of GCAM commodities not including purchased irrigation water.
}
\details{
USDA cost data is used to calculate total production costs - purchased irrigation water costs for
GCAM commodities covered in USDA spreadsheets. This produces a commodity level water cost fraction for some
but not all GCAM commodities. L161.ag_irr_HA_frac_R_C_GLU irrigated vs rainfed harvested area data for the
covered GCAM commodities is used to form a linear model describing water cost fraction as a function of the
irrigated fraction of harvested area. This linear model then predicts the water cost fraction based on
irrigated fraction of harvested area for the missing commodities. Finally, cost data from LB133.ag_Cost_75USDkg_C
is adjusted as LB133_cost * (1 - water cost fraction) = production cost - purchased irrigation water for each
commodity.
}
\author{
ACS June 2017
}
