% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB165.ag_water_R_C_Y_GLU_irr.R
\name{module_aglu_LB165.ag_water_R_C_Y_GLU_irr}
\alias{module_aglu_LB165.ag_water_R_C_Y_GLU_irr}
\title{module_aglu_LB165.ag_water_R_C_Y_GLU_irr}
\usage{
module_aglu_LB165.ag_water_R_C_Y_GLU_irr(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L165.BlueIrr_m3kg_R_C_GLU}, \code{L165.TotIrr_m3kg_R_C_GLU},
\code{L165.GreenRfd_m3kg_R_C_GLU}, \code{L165.ag_IrrEff_R}, \code{L165.IrrWithd_km3_R_Y}. The corresponding file in
the original data system was \code{LB165.ag_water_R_C_Y_GLU_irr.R} (aglu level1).
}
\description{
Compute irrigation efficiency (by GCAM region) and blue, green, and total consumption coefficients (by region, commodity, and GLU).
}
\details{
Use inventory estimates of either water demand coefficients by country and crop,
or aggregated gridded volumes of water use by country, GLU, and crop, to calculate average
water consumption coefficients by GCAM region, crop, GLU, and irrigation level.
Blue water is assigned to only irrigated production, but green water applies to both
irrigated and rainfed production, as the two management technologies are not disaggregated
in the inventory data. For rainfed crops, total biophysical = green, and for irrigated crops,
total biophysical = blue + green.
}
\author{
BBL April 2017
}
