% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB166.ag_bio_CCI_R_C_GLU.R
\name{module_aglu_LB166.ag_bio_CCI_R_C_GLU}
\alias{module_aglu_LB166.ag_bio_CCI_R_C_GLU}
\title{module_aglu_LB166.ag_bio_CCI_R_C_GLU}
\usage{
module_aglu_LB166.ag_bio_CCI_R_C_GLU(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L166.YieldRatio_R_C_Y_GLU_irr_CCIscen}.
}
\description{
Calculate base year bioenergy yields by GCAM region and GLU
}
\details{
Calculate global average yields for each FAO crop in the base year;
calculate each region / zone / crop's comparative yield; compute bioenergy yields as
this region/zone-specific index multiplied by a base yield.
}
\author{
BBL June 2017
}
