% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_xml_paper_delete_ag_demand.R
\name{module_aglu_paper_delete_ag_demand_xml}
\alias{module_aglu_paper_delete_ag_demand_xml}
\title{module_aglu_paper_delete_ag_demand_xml}
\usage{
module_aglu_paper_delete_ag_demand_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{paper_delete_ag_demand.xml}. The corresponding file in the
original data system was \code{batch_paper_delete_ag_demand_xml.R} (energy XML).
}
\description{
Construct XML data structure for \code{paper_delete_ag_demand.xml} and \code{paper_delete_ag_demand_USA.xml}.
}
