% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_xml_pasture_ssp34.R
\name{module_aglu_pasture_ssp34_xml}
\alias{module_aglu_pasture_ssp34_xml}
\title{module_aglu_pasture_ssp34_xml}
\usage{
module_aglu_pasture_ssp34_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{pasture_ssp34.xml}. The corresponding file in the
original data system was \code{batch_pasture_ssp34.xml.R} (aglu XML).
}
\description{
Construct XML data structure for \code{pasture_ssp34.xml}.
}
