% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_transportation_INDIA_xml.R
\name{module_batch_transportation_India_xml}
\alias{module_batch_transportation_India_xml}
\title{module_batch_transportation_India_xml}
\usage{
module_batch_transportation_India_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{transportation_India_new.xml}. The corresponding file in the
original data system was \code{batch_transportation_India_xml.R} (gcamIndia XML).
}
\description{
Construct XML data structure for \code{transportation_India_new.xml}.
}
\author{
Malyan_Ankur_CEEW @updated Rishabh_Varma_CEEW 2025
}
