% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/zchunk_Xbatch_socioeconomics_xml_Bangkok_Thailand.R
\name{module_breakout_Xbatch_socioeconomics_xml_Bangkok_Thailand}
\alias{module_breakout_Xbatch_socioeconomics_xml_Bangkok_Thailand}
\title{module_breakout_Xbatch_socioeconomics_xml_Bangkok_Thailand}
\usage{
module_breakout_Xbatch_socioeconomics_xml_Bangkok_Thailand(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{socioeconomics_USA.xml}. The corresponding file in the
original data system was \code{batch_socioeconomics_USA.xml} (gcamusa XML).
}
\description{
Construct XML data structure for \code{socioeconomics_USA.xml}.
}
