% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1091.ag_GrossTrade.R
\name{module_data_FAO_BilateralTrade}
\alias{module_data_FAO_BilateralTrade}
\title{module_data_FAO_BilateralTrade}
\usage{
module_data_FAO_BilateralTrade(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{FAO_BilateralTrade}.
}
\description{
Read the FAO_BilateralTrade trade from file and reduce / summarize as soon as possible to
keep its memory footprint small.  NOTE: we do not market the FAO_BilateralTrade as
INPUT in order to have more control of when it gets loaded and minimize the amount of
time it is loaded in memory.
}
\author{
PLP
}
