% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_all_aglu_emissions_IRR_MGMT_xml.R
\name{module_emissions_batch_all_aglu_emissions_IRR_MGMT_xml}
\alias{module_emissions_batch_all_aglu_emissions_IRR_MGMT_xml}
\title{module_emissions_batch_all_aglu_emissions_IRR_MGMT_xml}
\usage{
module_emissions_batch_all_aglu_emissions_IRR_MGMT_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{all_aglu_emissions_IRR_MGMT.xml}, \code{all_aglu_emissions_IRR_MGMT_MAC.xml}.
The corresponding file in the original data system was
\code{batch_all_aglu_emissions_IRR_MGMT.xml} (emissions XML).
}
\description{
Construct XML data structure for \code{all_aglu_emissions_IRR_MGMT.xml}.
}
