% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L102.nonco2_ceds_R_S_Y.R
\name{module_emissions_L102.nonco2_ceds_R_S_Y}
\alias{module_emissions_L102.nonco2_ceds_R_S_Y}
\title{module_emissions_L102.nonco2_ceds_R_S_Y}
\usage{
module_emissions_L102.nonco2_ceds_R_S_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L102.ceds_nonco2_tg_R_S_F}, \code{L102.ceds_GFED_nonco2_tg_C_S_F}, \code{L102.ceds_int_shipping_nonco2_tg_S_F}.
}
\description{
Calculates emissions using CEDS and CMIP emissions data for all sectors and fuels and aggregates to GCAM regions. Note that the outputs of this chunk are a part of the prebuilt data.
To change the ouputs of this chunk, add CEDS data to the CEDS folder under emissions and rebuild prebuilt data.
}
\details{
Calculates emissions using CEDS and CMIP data.
}
\author{
CWR May 2019, KBN June 2020
}
