% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L103.ghg_an_USA_S_T_Y.R
\name{module_emissions_L103.ghg_an_USA_S_T_Y}
\alias{module_emissions_L103.ghg_an_USA_S_T_Y}
\title{module_emissions_L103.ghg_an_USA_S_T_Y}
\usage{
module_emissions_L103.ghg_an_USA_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L103.ghg_tgmt_USA_an_Sepa_F_2005}. The corresponding file in the
original data system was \code{L103.ghg_an_USA_S_T_Y.R} (emissions level1).
}
\description{
Calculates methane emissions factors for animals by GCAM technology,
computed from EPA emissions data and FAO animal data for the US, 2005
}
\details{
Calculated methane emissions factors for animal production by GCAM technology (animal type) from EPA
emissions estimates and FAO production data for the US in 2005.
}
\author{
RH April 2017
}
