% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L112.ceds_ghg_en_R_S_T_Y.R
\name{module_emissions_L112.ceds_ghg_en_R_S_T_Y}
\alias{module_emissions_L112.ceds_ghg_en_R_S_T_Y}
\title{module_emissions_L112.ceds_ghg_en_R_S_T_Y}
\usage{
module_emissions_L112.ceds_ghg_en_R_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L112.cedsghg_tg_R_en_S_F_Yh}, \code{L112.cedsghg_tgej_R_en_S_F_Yh}. The corresponding file in the
original data system was \code{L112.ghg_en_R_S_T_Y.R} (emissions level1).
}
\description{
Calculates emissions and emissions factors using CEDS and GFED emissions data for GCAM sectors and technologies. The chunk contains processing for the following,
Part 1 : Distributing CEDS road transport emissions into different modes using the GAINS emissions data
Part 2 : Calculation of emission factors, input emissions for all combustion emission types.
Part 3 : Calculation of input emissions for process emission sectors
Part 4 : Calculation of animal sector emissions
Part 5 : Calculation of cropland emissions.
Part 6 : Calculation of agricultural waste burning emissions.
Part 7 : Calculation of input emissions, deforestation co-efs for unmanaged land emissions sectors.
Part 8 : Scale CH4 and N2O emissions to 2019 EPA nonCO2 report
}
\details{
Calculates emissions using CEDS emissions and GCAM sectors.
}
\author{
CWR Oct. 2018 , YO Mar. 2020, KBN 2020
}
