% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L123.bcoc_awb_R_S_T_Y.R
\name{module_emissions_L123.bcoc_awb_R_S_T_Y}
\alias{module_emissions_L123.bcoc_awb_R_S_T_Y}
\title{module_emissions_L123.bcoc_awb_R_S_T_Y}
\usage{
module_emissions_L123.bcoc_awb_R_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L123.bcoc_tgmt_R_awb_2000}. The corresponding file in the
original data system was \code{L123.bcoc_awb_R_S_T_Y.R} (emissions level1).
}
\description{
Produces BC and OC AWB emission factors for each Ag production technology.
}
\details{
Use RCP year 2000 BC and OC emissions from agricultural waste burning on fields (AWB) to calculate emission factors for each ag production technology. Allocates regional AWB emissions to GCAM technology according to above-ground non-harvested biomass.
}
\author{
SJS May 2017
}
