% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L152.MACC.R
\name{module_emissions_L152.MACC}
\alias{module_emissions_L152.MACC}
\title{module_emissions_L152.MACC}
\usage{
module_emissions_L152.MACC(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L152.MAC_pct_R_S_Proc_EPA}. The corresponding file in the
original data system was \code{L152.MACC.R} (emissions level1).
}
\description{
Create Marginal Abatement Cost Curves, in percent reduction by 1990 USD abatement costs from EPA cost curves.
}
\details{
Create Marginal abatement cost curves, in percent reduction by 1990 USD costs from EPA cost curves.
}
\author{
RMH May 2017 / YO Mar 2020
}
