% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zemissions_L161.nonghg_en_ssp_R_S_T_Y.R
\name{module_emissions_L161.nonghg_en_ssp_R_S_T_Y}
\alias{module_emissions_L161.nonghg_en_ssp_R_S_T_Y}
\title{module_emissions_L161.nonghg_en_ssp_R_S_T_Y}
\usage{
module_emissions_L161.nonghg_en_ssp_R_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L161.SSP15_EF}, \code{L161.SSP2_EF}, \code{L161.SSP34_EF}. The corresponding file in the
original data system was \code{L161.nonghg_en_ssp_R_S_T_Y.R} (emissions level1).
}
\description{
Produce future non-GHG emissions factors by SSP scenario.
}
\details{
Scales future GAINS, Greenhouse Gas - Air Pollution Interactions and Synergies model, non-GHG emissions factors to L111/L114 base year emissions factors,
then applies future emissions factors to some GCAM years based on SSP-specific rules.
}
\author{
RLH July 2017
}
