% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L201.en_nonco2.R
\name{module_emissions_L201.en_nonco2}
\alias{module_emissions_L201.en_nonco2}
\title{module_emissions_L201.en_nonco2}
\usage{
module_emissions_L201.en_nonco2(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L201.en_pol_emissions}, \code{L201.en_ghg_emissions}, \code{L201.en_bcoc_emissions}, \code{L201.OutputEmissions_elec}, \code{L201.nonghg_max_reduction}, \code{L201.nonghg_steepness}, \code{L201.nonghg_max_reduction_res}, \code{L201.nonghg_steepness_res}, \code{L201.nonghg_res}, \code{L201.ghg_res}, \code{L201.ResReadInControl_nonghg_res}, \code{L201.ResReadInControl_ghg_res}. The corresponding file in the
original data system was \code{L201.en_nonco2.R} (emissions level2).
}
\description{
Generate non-CO2 emissions: pollutants, GHGs, non-CO2, BC/OCs, and reduction data.
}
\details{
Set up all of the inputs needed for the energy system non-CO2 emissions in GCAM.
This includes historical emissions, drivers (input or output), and pollution controls.
}
\author{
BBL July 2017
}
