% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zemissions_L212.unmgd_nonco2.R
\name{module_emissions_L212.unmgd_nonco2}
\alias{module_emissions_L212.unmgd_nonco2}
\title{module_emissions_L212.unmgd_nonco2}
\usage{
module_emissions_L212.unmgd_nonco2(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs:\code{L212.AgSupplySector}, \code{L212.AgSupplySubsector}, \code{L212.ItemName},
\code{L212.GRASSEmissions}, \code{L212.FORESTEmissions_FF}, \code{L212.FORESTEmissions_D}, \code{L212.GRASSEmissionsFactors_BCOC},
\code{L212.FORESTEmissionsFactors_BCOC_FF}, \code{L212.FORESTEmissionsFactors_BCOC_D}, \code{L212.FORESTEmissionsFactors_future},
\code{L212.ItemName_prot}, \code{L212.GRASSEmissions_prot}, \code{L212.GRASSEmissions_noprot}, \code{L212.FORESTEmissions_FF_prot},
\code{L212.FORESTEmissions_FF_noprot}, \code{L212.FORESTEmissions_D_prot}, \code{L212.FORESTEmissions_D_noprot}, \code{L212.GRASSEmissionsFactors_BCOC_prot},
\code{L212.GRASSEmissionsFactors_BCOC_noprot}, \code{L212.FORESTEmissionsFactors_BCOC_FF_prot}, \code{L212.FORESTEmissionsFactors_BCOC_FF_noprot},
\code{L212.FORESTEmissionsFactors_BCOC_D_prot}, \code{L212.FORESTEmissionsFactors_BCOC_D_noprot}. The corresponding file in the
original data system was \code{L212.unmgd_nonco2.R} (emissions level2).
}
\description{
Outputs unmanaged land emissions and emissions coefficients for forest fires, deforestation, and grassland fires.
}
\details{
Outputs unmanaged land emissions and emissions coefficients for forest fires, deforestation, and grassland fires.
}
\author{
RLH August 2017
}
