% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L232.prc_nonco2.R
\name{module_emissions_L232.prc_nonco2}
\alias{module_emissions_L232.prc_nonco2}
\title{module_emissions_L232.prc_nonco2}
\usage{
module_emissions_L232.prc_nonco2(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L232.nonco2_prc}, \code{L232.nonco2_max_reduction}, \code{L232.nonco2_steepness}. The corresponding file in the
original data system was \code{L232.prc_nonco2.R} (emissions level2).
}
\description{
Generates input emissions by energy technology, GHG, and base historical year.
Writes out max emissions reductions and steepness to all energy technologies and regions.
}
\details{
Generates input emissions by energy technology, GHG, and base historical year. Writes out max emissions reductions and steepness to all energy technologies and regions.
}
\author{
RH July 2017
}
