% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2521.MACC_IRR.R
\name{module_emissions_L2521.MACC_IRR}
\alias{module_emissions_L2521.MACC_IRR}
\title{module_emissions_L2521.MACC_IRR}
\usage{
module_emissions_L2521.MACC_IRR(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2521.AgMAC}, \code{L2521.MAC_Ag_TC_SSP1}, \code{L2521.MAC_An_TC_SSP1}, \code{L2521.MAC_Ag_TC_SSP2}, \code{L2521.MAC_An_TC_SSP2}, \code{L2521.MAC_Ag_TC_SSP5}, \code{L2521.MAC_An_TC_SSP5}. The corresponding file in the
original data system was \code{L2521.MACC_IRR.R} (emissions level2).
}
\description{
Adds technologies and tech changes to marginal abatement cost curves, MACC, for animals and agriculture.
}
\details{
Adds agricultural technology to agriculture marginal abatement cost curves, MACC. Adds SSP-specific tech changes to animal and agriculture MAC curves.
}
\author{
RH July 2017
}
