% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L253.emission_controls.R
\name{module_emissions_L253.emission_controls}
\alias{module_emissions_L253.emission_controls}
\title{module_emissions_L253.emission_controls}
\usage{
module_emissions_L253.emission_controls(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L253.EF_retrofit}, \code{L253.EF_NSPS_new_vintage},
\code{L253.Retrofit_off}, \code{L253.delete_gdp_control},
\code{L253.user_EF_retrofit}, \code{L253.user_EF_NSPS_new_vintage},
\code{L253.user_Retrofit_off}, \code{L253.user_delete_gdp_control},
\code{L253.EF_retrofit_USA}, \code{L253.EF_NSPS_new_vintage_USA},
\code{L253.Retrofit_off_USA}, \code{L253.delete_gdp_control_USA},
\code{L253.user_EF_retrofit_USA}, \code{L253.user_EF_NSPS_new_vintage_USA},
\code{L253.user_Retrofit_off_USA}, \code{L253.user_delete_gdp_control_USA},
}
\description{
Allow user to specify NSPS and existing vintage retrofits for regions
}
\author{
ENL August 2021
}
