% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_heat_CHINA_xml.R
\name{module_energy_batch_heat_china_xml}
\alias{module_energy_batch_heat_china_xml}
\title{module_energy_batch_heat_china_xml}
\usage{
module_energy_batch_heat_china_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{heat_china.xml}. The corresponding file in the
original data system was \code{batch_heat.xml.R} (energy XML).
}
\description{
Construct XML data structure for \code{heat.xml}.
}
