% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_negative_emissions_budget_xml.R
\name{module_energy_batch_negative_emissions_budget_xml}
\alias{module_energy_batch_negative_emissions_budget_xml}
\title{module_energy_batch_negative_emissions_budget_xml}
\usage{
module_energy_batch_negative_emissions_budget_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{paste0( "negative_emissions_budget_", c("GCAM3", paste0("SSP", 1:5), paste0("gSSP", 1:5), paste0("spa", 1:5)), ".xml")}. The corresponding file in the
original data system was \code{L270.limits.R} (energy XML).
}
\description{
Construct XML data structure for \code{paste0( "negative_emissions_budget_", c("GCAM3", paste0("SSP", 1:5), paste0("gSSP", 1:5)), paste0("spa", 1:5), ".xml")}.
}
