% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_Off_road_SSP_xml.R
\name{module_energy_batch_Off_road_incelas_SSP_xml}
\alias{module_energy_batch_Off_road_incelas_SSP_xml}
\title{module_energy_batch_Off_road_incelas_SSP_xml}
\usage{
module_energy_batch_Off_road_incelas_SSP_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{Off_road_incelas_gcam3.xml}, \code{Off_road_incelas_ssp1.xml}, \code{Off_road_incelas_ssp2.xml}, \code{Off_road_incelas_ssp3.xml},
\code{Off_road_incelas_ssp4.xml}, \code{Off_road_incelas_ssp5.xml}, \code{Off_road_incelas_gssp1.xml}, \code{Off_road_incelas_gssp2.xml},
\code{Off_road_incelas_gssp3.xml}, \code{Off_road_incelas_gssp4.xml}, and \code{Off_road_incelas_gssp5.xml}.
}
\description{
Construct XML data structures for all the \code{Off_road_incelas_SSP.xml} files.
}
