% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_industry_SSP_CHINA_xml.R
\name{module_energy_batch_other_industry_incelas_SSP_CHINA_xml}
\alias{module_energy_batch_other_industry_incelas_SSP_CHINA_xml}
\title{module_energy_batch_other_industry_incelas_SSP_CHINA_xml}
\usage{
module_energy_batch_other_industry_incelas_SSP_CHINA_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{other_industry_incelas_gcam3_china.xml}, \code{other_industry_incelas_ssp1_china.xml}, \code{other_industry_incelas_ssp2_china.xml}, \code{other_industry_incelas_ssp3_china.xml},
\code{other_industry_incelas_ssp4_china.xml}, \code{other_industry_incelas_ssp5_china.xml}, \code{other_industry_incelas_gssp1_china.xml}, \code{other_industry_incelas_gssp2_china.xml},
\code{other_industry_incelas_gssp3_china.xml}, \code{other_industry_incelas_gssp4_china.xml}, and \code{other_industry_incelas_gssp5_china.xml}.
}
\description{
Construct XML data structures for all the \code{other_industry_incelas_SSP_CHINA.xml} files.
}
