% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_refining_USA_BlendWall_xml.R
\name{module_energy_batch_refining_USA_BlendWall_xml}
\alias{module_energy_batch_refining_USA_BlendWall_xml}
\title{module_energy_batch_refining_USA_BlendWall_xml}
\usage{
module_energy_batch_refining_USA_BlendWall_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{refining_USA_BlendWall.xml}.
}
\description{
Construct XML data structure for \code{refining_USA_BlendWall.xml}.
}
