% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_xml_cement.R
\name{module_energy_cement_xml}
\alias{module_energy_cement_xml}
\title{module_energy_cement_xml}
\usage{
module_energy_cement_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{cement.xml}. The corresponding file in the
original data system was \code{batch_cement_xml.R} (energy XML).
}
\description{
Construct XML data structure for \code{cement.xml}.
}
