% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_xml_geo_tech_adv.R
\name{module_energy_geo_tech_adv_xml}
\alias{module_energy_geo_tech_adv_xml}
\title{module_energy_geo_tech_adv_xml}
\usage{
module_energy_geo_tech_adv_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{geo_tech_adv.xml}. The corresponding file in the
original data system was \code{batch_geo_tech_adv_xml.R} (energy XML).
}
\description{
Construct XML data structure for \code{geo_tech_adv.xml}.
}
