% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_L1011.fossilFuel_GrossTrade_EJ_R_C_Y.R
\name{module_energy_L1011.ff_GrossTrade}
\alias{module_energy_L1011.ff_GrossTrade}
\title{module_energy_L1011.ff_GrossTrade}
\usage{
module_energy_L1011.ff_GrossTrade(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1011.ff_GrossTrade_EJ_R_C_Y}, \code{L1011.ff_GrossTrade_EJ_R_Y_LNG},
\code{L1011.ff_GrossTrade_EJ_R_Y_NG_pipe}, \code{L1011.ff_BilatTrade_EJ_R_Y_NG_pipe} (energy level1).
}
\description{
Calculate primary fossil fuel (coal, natural gas, crude oil) product balances, by region / commodity / year.
}
\details{
This chunk processes the bi-lateral trade flow data matrix from UN Comtrade, in order to identify trade of
  fossil fuels between GCAM regions.
}
\author{
GPK/JEH April 2019
}
