% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_L1092.iron_steel_GrossTrade.R
\name{module_energy_L1092.iron_steel_GrossTrade}
\alias{module_energy_L1092.iron_steel_GrossTrade}
\title{module_energy_L1092.iron_steel_GrossTrade}
\usage{
module_energy_L1092.iron_steel_GrossTrade(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{LB1092.Tradebalance_iron_steel_Mt_R_Y}.
}
\description{
Reads country-level steel production, consumption, imports and exports data from World Steel Association data set.
Adjusts regional steel exports and imports by a scaling factor to fit the trade balance equation (consumption = production - exports + imports).
Aggregates trade data by GCAM region, Calculates and removes intra-regional trade using bilateral trade data from resource trade database.
Scales the regional trade data to ensure global imports minus exports and global production minus consumption is equal to zero.
}
\author{
Siddarth Durga July 2022
}
