% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_L111.rsrc_fos_Prod.R
\name{module_energy_L111.rsrc_fos_Prod}
\alias{module_energy_L111.rsrc_fos_Prod}
\title{module_energy_L111.rsrc_fos_Prod}
\usage{
module_energy_L111.rsrc_fos_Prod(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L111.Prod_EJ_R_F_Yh}, \code{L111.RsrcCurves_EJ_R_Ffos}. The corresponding file in the
original data system was \code{LA111.rsrc_fos_Prod.R} (energy level1).
}
\description{
Calculate historical fossil energy production and fossil resource supply curves.
}
\details{
For historical fossil energy production, determine regional shares of production for each primary fuel,
interpolate unconventional oil production to all historical years, deduct unconventional oil from total oil,
and include it in calibrated production table.
For fossil resource supply curves, downscale GCAM3.0 supply curves to the country level (on the basis of
resource production) and aggregate by the new GCAM regions, using crude oil production shares as a proxy
for unconventional oil resources.
}
\author{
BBL August 2017
}
