% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_L112.U.R
\name{module_energy_L112.U}
\alias{module_energy_L112.U}
\title{module_energy_L112.U}
\usage{
module_energy_L112.U(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L112.RsrcCurves_Mt_R_U}. The corresponding file in the
original data system was \code{LA112.U.R} (energy level1).
}
\description{
Assigning global uranium supply curve to GCAM_region_ID 1.
}
\details{
Describe in detail what this chunk does.
}
\author{
HM April 2017
}
