% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_L120.offshore_wind.R
\name{module_energy_L120.offshore_wind}
\alias{module_energy_L120.offshore_wind}
\title{module_energy_L120.offshore_wind}
\usage{
module_energy_L120.offshore_wind(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L120.RsrcCurves_EJ_R_offshore_wind}, \code{L120.TechChange_offshore_wind},
\code{L120.GridCost_offshore_wind}, \code{L120.RegCapFactor_offshore_wind}. The corresponding file in the
original data system was \code{LA120.offshore_wind.R} (energy level1).
}
\description{
Takes in data on country-level offshore wind energy potential and global offshore wind capital
cost assumptions and generates tables containing region-level offshore wind data.
}
\details{
Takes in data on country-level offshore wind energy potential and global offshore wind
capital cost assumptions and generates tables containing region-level offshore wind data.
}
\author{
MB GI AJS March 2019
}
