% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_L121.liquids.R
\name{module_energy_L121.liquids}
\alias{module_energy_L121.liquids}
\title{module_energy_L121.liquids}
\usage{
module_energy_L121.liquids(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L121.in_EJ_R_unoil_F_Yh},
  \code{L121.in_EJ_R_TPES_crude_Yh}, \code{L121.in_EJ_R_TPES_unoil_Yh}, \code{L121.share_R_TPES_biofuel_tech}. The
  corresponding file in the original data system was \code{LA121.oil.R} (energy level1).
}
\description{
Process historical oil data and separate into unconventional oil, crude oil, and energy inputs to oil.
}
\details{
This chunk uses energy production data to determine the energy inputs to the oil sector. It uses the IEA
  energy balance to separate out unconventional oil production and crude oil (total liquids - unconventional oil).
  It also uses data from IIASA to downscale ethanol and biodiesel consumption to modeled technologies and feedstocks.
}
\note{
If the (proprietary) IEA data aren't available, pre-built summaries are used.
}
\author{
JDH June 2017, ed GPK April 2019
}
