% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/zchunk_LA1262.refinedLiquids_GrossTrade_EJ_R_C_Y.R
\name{module_energy_L1262.refinedOil_GrossTrade}
\alias{module_energy_L1262.refinedOil_GrossTrade}
\title{module_energy_L1262.refinedOil_GrossTrade}
\usage{
module_energy_L1262.refinedOil_GrossTrade(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1262.refinedOil_GrossTrade_EJ_R_C_Y} (energy level1)
}
\description{
Calculate primary regional liquids product mass balances, by region / commodity / year.
}
\details{
This chunk processes the bi-lateral trade flow data matrix from UN Comtrade, in order to differentiate trade of
  regional liquids between GCAM regions.
}
\author{
GPK/JEH April 2019
}
