% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_L142.building_agg.R
\name{module_energy_L142.building_agg}
\alias{module_energy_L142.building_agg}
\title{module_energy_L142.building_agg}
\usage{
module_energy_L142.building_agg(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L142.in_EJ_R_bld_F_Yh}. The corresponding file in the
original data system was \code{LA142.building_agg.R} (energy level1).
}
\description{
Calculate building sector energy consumption, producing the following output table: Building energy consumption by GCAM region / fuel / historical year
}
\details{
Building sector energy consumption was obtained from end use energy consumption data

Fuel inputs to heat were added to building energy use in regions where heat is not modeled as a final fuel
}
\author{
AS May 2017
}
