% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_L152.transportation.R
\name{module_energy_L152.transportation}
\alias{module_energy_L152.transportation}
\title{module_energy_L152.transportation}
\usage{
module_energy_L152.transportation(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L152.in_EJ_R_trn_F_Yh}. The corresponding file in the
original data system was \code{LA152.transportation.R} (energy level1).
}
\description{
Calculate transportation sector energy consumption.
}
\details{
Calculate transportation sector energy consumption by GCAM region, sector, fuel, and year.
}
\author{
BBL April 2017
}
