% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L202.Ccoef.R
\name{module_energy_L202.Ccoef}
\alias{module_energy_L202.Ccoef}
\title{module_energy_L202.Ccoef}
\usage{
module_energy_L202.Ccoef(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L202.CarbonCoef}. The corresponding file in the
original data system was \code{L202.Ccoef.R} (energy level2).
}
\description{
Create the file of CO2 coefficients (carbon emissions/GJ energy consumed) for primary energy sources.
}
\details{
Based on user-defined switches (in constants.R), this code produces the CO2 coefficients for primary energy sources
# based on 1. GCAM 3.0 assumptions 2. global average emissions coefficients based on CDIAC inventory and IEA energy balances
# or 3. region-specific average emissions coefficients based on CDIAC inventory and IEA energy balances.
}
\author{
JDH Sept 2017
}
