% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2221.biochar_demand_cropland.R
\name{module_energy_L2221.biochar_demand_cropland}
\alias{module_energy_L2221.biochar_demand_cropland}
\title{module_energy_L2221.biochar_demand_cropland}
\usage{
module_energy_L2221.biochar_demand_cropland(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2221.LN5_HistMgdAllocation_crop_biochar}, \code{L2221.LN5_HistMgdAllocation_bio_biochar},
\code{L2221.LN5_MgdAllocation_crop_biochar}, \code{L2221.LN5_MgdAllocation_bio_biochar},
\code{L2221.LN5_MgdCarbon_crop_biochar}, \code{L2221.LN5_MgdCarbon_bio_biochar},
\code{L2221.LN5_LeafGhostShare}, \code{L2221.LN5_LeafGhostShare_bio},
\code{L2221.AgProduction_ag_irr_mgmt}, \code{L2221.AgHAtoCL_irr_mgmt},
\code{L2221.AgYield_biochar_ref}, \code{L2221.AgYield_bio_biochar_ref},
\code{L2221.AgCost_ag_irr_mgmt}, \code{L2221.AgCost_bio_irr_mgmt},
\code{L2221.AgProdChange_ag_irr_ref}, \code{L2221.AgProdChange_bio_irr_ref},
\code{L2221.AgCoef_Fert_ag_irr_mgmt}, \code{L2221.AgCoef_Fert_bio_irr_mgmt},
\code{L2221.AgCost_ag_irr_mgmt_adj}, \code{L2221.AgCost_bio_irr_mgmt_adj}, \code{L2221.AgResBio_ag_irr_mgmt},
\code{L2221.AgResBioCurve_ag_irr_mgmt}, \code{L2221.AgCoef_Biochar_ag_irr_mgmt}, \code{L2221.AgCost_Biochar_irr_mgmt_adj},
\code{L2221.AgCoef_Biochar_bio_irr_mgmt}, \code{L2221.AgCost_Biochar_irr_mgmt_adj_bio},
\code{L2221.CarbonCoef}, \code{L2221.AgCoef_IrrBphysWater_ag_mgmt}, \code{L2221.AgCoef_BphysWater_bio_mgmt},
\code{L2221.AgCoef_IrrWaterWdraw_ag_mgmt}, \code{L2221.AgCoef_IrrWaterWdraw_bio_mgmt},
\code{L2221.AgCoef_IrrWaterCons_ag_mgmt}, \code{L2221.AgCoef_IrrWaterCons_bio_mgmt},
\code{L2221.AgCoef_RfdBphysWater_ag_mgmt}, \code{L2221.AgNonEnergyCost_IrrWaterWdraw}.
}
\description{
Prepare the assumptions for teh demand of biochar in croppland; similar to fertilizer demand
}
\details{
This chunk sets up the biochar demand in agreiculture. We assume an applicaiton rate of 20 tons of
biochar per hectare. We follow similar approach to fertilizer to calculate IO coefficients.
Applies only to future years
}
\author{
MCB March 2021
}
