% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2221.en_transformation_biochar.R
\name{module_energy_L2221.en_transformation_biochar}
\alias{module_energy_L2221.en_transformation_biochar}
\title{module_energy_L2221.en_transformation_biochar}
\usage{
module_energy_L2221.en_transformation_biochar(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{

}
\description{
Prepare the assumptions and calibrated outputs for energy transformation supplysectors, subsectors, and technologies.
}
\details{
This chunk sets up the biochar sector as well as writing out assumptions to all regions for shareweights and logits.
Applies only to future years
}
\author{
MCB March 2021
}
