% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L225.hydrogen.R
\name{module_energy_L225.hydrogen}
\alias{module_energy_L225.hydrogen}
\title{module_energy_L225.hydrogen}
\usage{
module_energy_L225.hydrogen(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L225.Supplysector_h2}, \code{L225.SubsectorLogit_h2}, \code{L225.SubsectorShrwtFllt_h2}, \code{L225.StubTech_h2}, \code{L225.GlobalTechCoef_h2}, \code{L225.GlobalTechCost_h2}, \code{L225.GlobalTechShrwt_h2}, \code{L225.PrimaryRenewKeyword_h2}, \code{L225.GlobalTechCapture_h2}, \code{L225.StubTechCost_h2}, \code{L225.GlobalTechProfitShutdown_h2}, \code{L225.GlobalTechSCurve_h2}. The corresponding file in the
original data system was \code{L225.hydrogen.R} (energy level2).
}
\description{
Provides supply sector information, subsector information, technology information for hydrogen sectors.
}
\details{
Provides supply sector information, subsector information, technology information for hydrogen sectors.
}
\author{
LF Augest 2017
}
