% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2261.gasoline.R
\name{module_energy_L2261.gasoline}
\alias{module_energy_L2261.gasoline}
\title{module_energy_L2261.gasoline}
\usage{
module_energy_L2261.gasoline(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2222.SubsectorLogit_en}, \code{L2222.SubsectorShrwt_en},
\code{L2222.SubsectorShrwtFllt_en}, \code{L2222.StubTechInterp_en}, \code{L2222.StubTechCoef_en},
 \code{L2222.StubTechCost_en}, \code{L2222.StubTechShrwt_en}, \code{L2222.StubTechSecOut_en},
 \code{L2222.StubTechSCurve_en}, \code{L2222.StubTechProfitShutdown_en}, \code{L2222.StubTechCoef_BlendWall}.
}
\description{
Prepare the assumptions and calibrated outputs for motor gasoline ("refined liquids gasoline pool")
}
\details{
This chunk modifies the parameterization of subsectors and technologies in the USA refining sector as per
  data provided by OTAQ
}
\author{
GPK March 2019
}
