% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2261.liquid_fuels_distribution.R
\name{module_energy_L2261.liquid_fuels_distribution}
\alias{module_energy_L2261.liquid_fuels_distribution}
\title{module_energy_L2261.liquid_fuels_distribution}
\usage{
module_energy_L2261.liquid_fuels_distribution(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2261.Supplysector_en},
\code{L2261.SubsectorLogit_en}, \code{L2261.SubsectorShrwtFllt_en}, \code{L2261.SubsectorInterp_en},
 \code{L2261.StubTech_en}, \code{L2261.GlobalTechEff_en}, \code{L2261.GlobalTechCost_en},
 \code{L2261.GlobalTechShrwt_en}, \code{L2261.StubTechProd_gsln}.
}
\description{
Prepare the assumptions and calibrated outputs for motor gasoline ("refined liquids gasoline pool")
}
\details{
This chunk modifies the parameterization of subsectors and technologies in the USA refining sector as per
  data provided by OTAQ
}
\author{
GPK March 2019
}
