% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2262.refined_liquids_trade.R
\name{module_energy_L2262.refined_liquids_trade}
\alias{module_energy_L2262.refined_liquids_trade}
\title{module_energy_L2262.refined_liquids_trade}
\usage{
module_energy_L2262.refined_liquids_trade(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L2262.Supplysector_refinedLiquids_tra},
  \code{L2262.SectorUseTrialMarket_refinedLiquids_tra}, \code{L2262.SubsectorAll_refinedLiquids_tra}, \code{L2262.TechShrwt_refinedLiquids_tra},
  \code{L2262.TechCost_refinedLiquids_tra}, \code{L2262.TechCoef_refinedLiquids_tra}, \code{L2262.Production_refinedLiquids_tra}, \code{L2262.Supplysector_refinedLiquids_reg},
  \code{L2262.SubsectorAll_refinedLiquids_reg}, \code{L2262.TechShrwt_refinedLiquids_reg}, \code{L2262.TechCoef_refinedLiquids_reg}, \code{L2262.Production_refinedLiquids_reg_imp},
  \code{L2262.Production_refinedLiquids_reg_dom}.
}
\description{
Model input for regional and (globally) traded refined liquids
}
\details{
Build datasets for ssp4 agricultural trade: food and nonfood trade coefficients, feed trade
coefficients, restricted agricultural trade, and trade regions.
}
\author{
GPK/JEH March 2019
}
