% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2323.iron_steel.R
\name{module_energy_L2323.iron_steel}
\alias{module_energy_L2323.iron_steel}
\title{module_energy_L2323.iron_steel}
\usage{
module_energy_L2323.iron_steel(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2323.SectorLogitTables[[ curr_table ]]$data}, \code{L2323.Supplysector_iron_steel}, \code{L2323.FinalEnergyKeyword_iron_steel},
\code{L2323.SubsectorLogitTables[[ curr_table ]]$data}, \code{L2323.SubsectorLogit_iron_steel}, \code{L2323.SubsectorShrwtFllt_iron_steel},
\code{L2323.SubsectorInterp_iron_steel}, \code{L2323.StubTech_iron_steel}, \code{L2323.GlobalTechShrwt_iron_steel}, \code{L2323.GlobalTechCoef_iron_steel},
\code{L2323.GlobalTechCost_iron_steel}, \code{L2323.GlobalTechCapture_iron_steel}, \code{L2323.StubTechProd_iron_steel}, \code{L2323.StubTechCalInput_iron_steel},
\code{L2323.StubTechCoef_iron_steel}, \code{L2323.PerCapitaBased_iron_steel}, \code{L2323.BaseService_iron_steel}, \code{L2323.PriceElasticity_iron_steel},
\code{object}. The corresponding file in the
original data system was \code{L2323.iron_steel.R} (energy level2).
}
\description{
Compute a variety of final energy keyword, sector, share weight, and technology information for iron and steel-related GCAM inputs.
}
\details{
The chunk provides final energy keyword, supplysector/subsector information, supplysector/subsector interpolation information, global technology share weight, global technology efficiency, global technology coefficients, global technology cost, price elasticity, stub technology information, stub technology interpolation information, stub technology calibrated inputs, and etc for iron and steel sector.
}
\author{
Yang Liu Sep 2019
}
