% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2324.Off_road.R
\name{module_energy_L2324.Off_road}
\alias{module_energy_L2324.Off_road}
\title{module_energy_L2324.Off_road}
\usage{
module_energy_L2324.Off_road(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2324.SectorLogitTables[[ curr_table ]]$data}, \code{L2324.Supplysector_Off_road}, \code{L2324.FinalEnergyKeyword_Off_road},
\code{L2324.SubsectorLogitTables[[ curr_table ]]$data}, \code{L2324.SubsectorLogit_Off_road}, \code{L2324.SubsectorShrwtFllt_Off_road},
\code{L2324.SubsectorInterp_Off_road}, \code{L2324.StubTech_Off_road}, \code{L2324.GlobalTechShrwt_Off_road}, \code{L2324.GlobalTechCoef_Off_road},
\code{L2324.GlobalTechCost_Off_road}, \code{L2324.GlobalTechCapture_Off_road}, \code{L2324.StubTechProd_Off_road}, \code{L2324.StubTechCalInput_Off_road},
\code{L2324.StubTechCoef_Off_road}, \code{L2324.PerCapitaBased_Off_road}, \code{L2324.BaseService_Off_road}, \code{L2324.PriceElasticity_Off_road},\code{L2324.GlobalTechCSeq_ind},
\code{object}. The corresponding file in the
original data system was \code{L2324.Off_road.R} (energy level2).
}
\description{
Compute a variety of final energy keyword, sector, share weight, and technology information for Off_road-related GCAM inputs.
}
\details{
The chunk provides final energy keyword, supplysector/subsector information, supplysector/subsector interpolation information, global technology share weight, global technology efficiency, global technology coefficients, global technology cost, price elasticity, stub technology information, stub technology interpolation information, stub technology calibrated inputs, and etc for Off_road sector.
}
\author{
Yang Liu Sep 2019, Molly Charles 2020-21, 2022 modifications from Jay Fuhrman, Siddarth Durga, Page Kyle
}
