% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2325.chemical.R
\name{module_energy_L2325.chemical}
\alias{module_energy_L2325.chemical}
\title{module_energy_L2325.chemical}
\usage{
module_energy_L2325.chemical(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2325.SectorLogitTables[[ curr_table ]]$data}, \code{L2325.Supplysector_chemical}, \code{L2325.FinalEnergyKeyword_chemical},
\code{L2325.SubsectorLogitTables[[ curr_table ]]$data}, \code{L2325.SubsectorLogit_chemical}, \code{L2325.SubsectorShrwtFllt_chemical},
\code{L2325.SubsectorInterp_chemical}, \code{L2325.StubTech_chemical}, \code{L2325.GlobalTechShrwt_chemical}, \code{L2325.GlobalTechCoef_chemical},
\code{L2325.GlobalTechCost_chemical}, \code{L2325.GlobalTechCapture_chemical}, \code{L2325.StubTechProd_chemical}, \code{L2325.StubTechCalInput_chemical},
\code{L2325.StubTechCoef_chemical}, \code{L2325.PerCapitaBased_chemical}, \code{L2325.BaseService_chemical}, \code{L2325.PriceElasticity_chemical},
\code{L2325.GlobalTechCapture_chemical}, \code{L2325.GlobalTechEff_chemical},\code{L2325.GlobalTechCSeq_ind},
\code{object}. The corresponding file in the
}
\description{
Compute a variety of final energy keyword, sector, share weight, and technology information for chemical-related GCAM inputs.
}
\details{
The chunk provides final energy keyword, supplysector/subsector information, supplysector/subsector interpolation information, global technology share weight, global technology efficiency, global technology coefficients, global technology cost, price elasticity, stub technology information, stub technology interpolation information, stub technology calibrated inputs, and etc for chemical sector.
}
\author{
Yang Liu Dec 2019
}
