% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_L2328.food_processing.R
\name{module_energy_L2328.food_processing}
\alias{module_energy_L2328.food_processing}
\title{module_energy_L2328.food_processing}
\usage{
module_energy_L2328.food_processing(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2328.SectorLogitTables[[ curr_table ]]$data}, \code{L2328.Supplysector_food}, \code{L2328.FinalEnergyKeyword_food},
\code{L2328.SubsectorLogitTables[[ curr_table ]]$data}, \code{L2328.SubsectorLogit_food}, \code{L2328.SubsectorShrwtFllt_food},
\code{L2328.SubsectorInterp_food}, \code{L2328.StubTech_food}, \code{L2328.GlobalTechShrwt_food}, \code{L2328.GlobalTechCoef_food},
\code{L2328.GlobalTechCost_food}, \code{L2328.StubTechCost_food}, \code{L2328.StubTechProd_food}, \code{L2328.StubTechCalInput_food_heat},
\code{L2328.StubTechCoef_food}, \code{L2328.StubCalorieContent}, \code{L2328.StubCaloriePriceConv},
\code{L2328.GlobalTechSecOut_food}, \code{L2328.GlobalTechTrackCapital_food},
\code{object}.
}
\description{
Compute a variety of sector, share weight, and technology information for food processing-related GCAM inputs.
}
\details{
The chunk provides supplysector/subsector information, supplysector/subsector interpolation information,
global technology share weight, global technology efficiency, global technology coefficients, global technology cost,
stub technology information, stub technology interpolation information, stub technology calibrated inputs, and etc. for the food processing sector.
}
\author{
SAS Dec 2022
}
