% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L239.ff_trade.R
\name{module_energy_L239.ff_trade}
\alias{module_energy_L239.ff_trade}
\title{module_energy_L239.ff_trade}
\usage{
module_energy_L239.ff_trade(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L239.PrimaryConsKeyword_en}, \code{L239.Supplysector_tra},
  \code{L239.SectorUseTrialMarket_tra}, \code{L239.SubsectorAll_tra}, \code{L239.TechShrwt_tra},
  \code{L239.TechCost_tra}, \code{L239.TechCoef_tra}, \code{L239.Production_tra}, \code{L239.Supplysector_reg},
  \code{L239.SubsectorAll_reg}, \code{L239.TechShrwt_reg}, \code{L239.TechCoef_reg}, \code{L239.Production_reg_imp},
  \code{L239.Production_reg_dom}, \code{L239.Consumption_intraregional}, \code{L239.CarbonCoef}.
}
\description{
Model input for regionally and globally traded fossil fuel commodities
}
\author{
GPK February 2019
}
