% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_L2391.gas_trade_flows.R
\name{module_energy_L2391.gas_trade_flows}
\alias{module_energy_L2391.gas_trade_flows}
\title{module_energy_L2391.gas_trade_flows}
\usage{
module_energy_L2391.gas_trade_flows(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
 \code{command} is "MAKE") all the generated outputs: \code{L2391.NG_export_calOutput_LNG},
 \code{L2391.NG_export_calOutput_pipeline}, \code{L2391.NG_import_calOutput_LNG},
 \code{L2391.NG_import_calOutput_pipeline},  \code{L2391.NG_import_calOutput_statdiff},
 \code{L2391.NG_export_calOutput_statdiff}.
}
\description{
Model input for natural gas trade by LNG and regional pipeline networks.
}
\author{
MTB August 2021
}
