% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_L2392.gas_trade.R
\name{module_energy_L2392.gas_trade}
\alias{module_energy_L2392.gas_trade}
\title{module_energy_L2392.gas_trade}
\usage{
module_energy_L2392.gas_trade(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs:\code{L2392.Delete_Supplysector_tra_NG},
  \code{L2392.Delete_Supplysector_reg_NG}, \code{L2392.PrimaryConsKeyword_en_NG}, \code{L2392.CarbonCoef_NG},
  \code{L2392.Supplysector_tra_NG},\code{L2392.SectorUseTrialMarket_tra_NG}, \code{L2392.SubsectorAll_tra_NG},
  \code{L2392.TechShrwt_tra_NG}, \code{L2392.TechCost_tra_NG},\code{L2392.TechLifetime_tra_NG},\code{L2392.TechSCurve_tra_NG},
  \code{L2392.TechCoef_tra_NG}, \code{L2392.ProfitShutdown_tra_NG},
  \code{L2392.Supplysector_reg_NG}, \code{2392.NestingSubsectorAll_reg_NG}, \code{L2392.SubsectorAll_reg_NG},
  \code{L2392.TechShrwt_reg_NG},\code{L2392.TechCoef_reg_NG},\code{L2392.TechCost_reg_NG}, \code{L2392.TechLifetime_reg_NG},
  \code{L2392.TechSCurve_reg_NG},\code{L2392.ProfitShutdown_reg_NG}, \code{L2392.TechInterp_reg_NG},
  \code{L2392.Production_tra_NG}, \code{L2392.Production_reg_imp_NG}, \code{L2392.Production_reg_dom_NG}
}
\description{
Model input for regionally and globally traded gas commodities. Traded gas includes LNG and pipeline gas.
}
\author{
BY August 2021
}
