% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_L254.transportation_UCD.R
\name{module_energy_L254.transportation_UCD}
\alias{module_energy_L254.transportation_UCD}
\title{module_energy_L254.transportation_UCD}
\usage{
module_energy_L254.transportation_UCD(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L254.Supplysector_trn}, \code{L254.FinalEnergyKeyword_trn}, \code{L254.tranSubsectorLogit},
\code{L254.tranSubsectorShrwt}, \code{L254.tranSubsectorShrwtFllt}, \code{L254.tranSubsectorInterp},
\code{L254.tranSubsectorInterpTo}, \code{L254.tranSubsectorSpeed}, \code{L254.tranSubsectorSpeed_passthru},
\code{L254.tranSubsectorSpeed_noVOTT}, \code{L254.tranSubsectorSpeed_nonmotor}, \code{L254.tranSubsectorVOTT},
\code{L254.tranSubsectorFuelPref}, \code{L254.StubTranTech}, \code{L254.StubTech_passthru}, \code{L254.StubTech_nonmotor},
\code{L254.GlobalTechShrwt_passthru}, \code{L254.GlobalTechShrwt_nonmotor}, \code{L254.GlobalTechCoef_passthru},
\code{L254.GlobalRenewTech_nonmotor}, \code{L254.GlobalTranTechInterp}, \code{L254.GlobalTranTechShrwt},
\code{L254.GlobalTranTechSCurve}, \code{L254.StubTranTechCalInput}, \code{L254.StubTranTechLoadFactor},
\code{L254.StubTranTechCost}, \code{L254.StubTranTechCoef}, \code{L254.StubTechCalInput_passthru},
\code{L254.StubTechProd_nonmotor}, \code{L254.PerCapitaBased_trn}, \code{L254.PriceElasticity_trn},
\code{L254.IncomeElasticity_trn}, \code{L254.BaseService_trn}. The corresponding file in the
original data system was \code{L254.transportation_UCD.R} (energy level2).
}
\description{
Calculate transportation data using information from the global UCD transportation technology database.
}
\details{
Due to the asymmetrical nature of the transportation sectors in the various regions, we can't simply write
generic information to all regions. Instead, technology information is read from the global UCD transportation
technology database, and supplysector and subsector attributes are matched in from lookup tables.
}
\author{
AJS September 2017
}
