% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L270.limits.R
\name{module_energy_L270.limits}
\alias{module_energy_L270.limits}
\title{module_energy_L270.limits}
\usage{
module_energy_L270.limits(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L270.CreditOutput}, \code{L270.CreditInput_elec}, \code{L270.CreditInput_feedstocks}, \code{L270.CreditMkt}, \code{L270.CTaxInput}, \code{L270.LandRootNegEmissMkt}, \code{L270.NegEmissBudgetMaxPrice}, \code{paste0("L270.NegEmissBudget_", c("GCAM3", paste0("SSP", 1:5), paste0("gSSP", 1:5))) )}. The corresponding file in the
original data system was \code{L270.limits.R} (energy level2).
}
\description{
Generate GCAM policy constraints which limit model behavior in some way.  In
particular limit the fraction of liquid feedstocks and inputs to electricity
generation which can come from sources other than crude oil.  Constrain the
total amount of subsidy as a fraction of GDP which an economy is will to give
to have net negative emissions.
}
\details{
Generate GCAM policy constraints which enforce limits to liquid feedstocks
and the amount of subsidies given for net negative emissions.
}
\author{
PLP March 2018
}
