% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L271.ag_en_freight_inputs.R
\name{module_energy_L271.ag_en_freight_inputs}
\alias{module_energy_L271.ag_en_freight_inputs}
\title{module_energy_L271.ag_en_freight_inputs}
\usage{
module_energy_L271.ag_en_freight_inputs(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L271.TechCoef_freight}, \code{L271.TechPMult_freight}, \code{L271.TechCost_freight}, \code{L271.BaseService_freightNetEnAg}.
}
\description{
Generate inputs of freight transportation to energy and agricultural commodities, and modify the
final demands of freight transportation services accordingly
}
\details{
The freight input-output coefficients are exogenous, as are cost adjustments and price unit conversions.
  They are multiplied by commodity flow volumes to compute the trn_freight quantities that are no longer part of the
  final demands, and this deduction is performed here.
}
\author{
GPK March 2019
}
