% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA1012.en_bal_EFW.R
\name{module_energy_LA1012.en_bal_EFW}
\alias{module_energy_LA1012.en_bal_EFW}
\title{module_energy_LA1012.en_bal_EFW}
\usage{
module_energy_LA1012.en_bal_EFW(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1012.en_bal_EJ_R_Si_Fi_Yh}.
}
\description{
Adjustments to the IEA energy balances to account for energy-for-water processes explicitly modeled.
}
\details{
This chunk deducts bottom-up energy-for-water (EFW) estimates from the IEA Energy Balances estimates of
  energy consumption in the commercial and industrial sectors, applying a set of rules that ensure that resulting
  energy demand levels are reasonable. The result is the full energy balance table with an additional EFW
}
\author{
GPK January 2019
}
