% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEMO_chunk_LA112.U.R
\name{module_energy_LA112.U_DEMO}
\alias{module_energy_LA112.U_DEMO}
\title{module_energy_LA112.U_DEMO}
\usage{
module_energy_LA112.U_DEMO(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L112.RsrcCurves_Mt_R_U}. The corresponding file in the
original data system was \code{LA112.U.R} (energy level1).
}
\description{
Build the GCAM uranium energy resource curves.
}
\details{
Currently not built up from inventory data; just using GCAM 3.0 values
These were not disaggregated to regions in GCAM 3.0.
}
\author{
BBL March 2017
}

