% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA113.atb_cost.R
\name{module_energy_LA113.atb_cost}
\alias{module_energy_LA113.atb_cost}
\title{module_energy_LA113.atb_cost}
\usage{
module_energy_LA113.atb_cost(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L113.globaltech_capital_ATB}, \code{L113.globaltech_capital_ATB_adv},
\code{L113.globaltech_capital_ATB_low}, \code{L113.elecS_globaltech_capital_battery_ATB},
\code{L113.globaltech_OMfixed_ATB}, \code{L113.globaltech_OMvar_ATB}.
There was no corresponding file in the original data system.
}
\description{
Generates power sector cost input files based on 2019 ATB data.
}
\details{
Includes ATB capital cost data as starting point, improvement rate and improvement max data generated based on cost pathway.
}
\author{
AJS March 2019
}
