% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA116.geo.R
\name{module_energy_LA116.geo}
\alias{module_energy_LA116.geo}
\title{module_energy_LA116.geo}
\usage{
module_energy_LA116.geo(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L116.RsrcCurves_EJ_R_geo}, \code{L116.RsrcCurves_EJ_R_EGS}. The corresponding file in the
original data system was \code{LA116.geo.R} (energy level1).
}
\description{
Generate geothermal (hydrothermal and engineered geothermal systems (EGS)) supply curves by GCAM region
Supply curves developed for the 14 GCAM 3.0 regions are downscaled to the country level on the basis
of land area, and aggregated to the current GCAM regions. The data source for the supply curves is
Hannam et al. 2009: http://www.pnl.gov/main/publications/external/technical_reports/PNNL-19231.pdf
}
\details{
This code chunk assigns geothermal electric resources to GCAM regions. Because the underlying
source data were developed for the 14 regions of GCAM 3.0 and prior, they are first
downscaled to the nation level on the basis of land area, and aggregated to the current GCAM regions.
}
\author{
GPK April 2017
}
