% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA1232.desal_from_elec.R
\name{module_energy_LA1232.desal_from_elec}
\alias{module_energy_LA1232.desal_from_elec}
\title{module_energy_LA1232.desal_from_elec}
\usage{
module_energy_LA1232.desal_from_elec(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1232.desalsecout_R_elec_F_tech}.
}
\description{
Output ratio of desalinated seawater produced as secondary output from electric power plants.
}
\details{
This chunk computes the output-ratio of desalinated water from electric power plants, indicated in m3 of
  desalinated water per GJ of electricity produced. It is only computed in region(s) with combined electric and desal
  plants.
}
\author{
GPK January 2019
}
