% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA125.hydrogen.R
\name{module_energy_LA125.hydrogen}
\alias{module_energy_LA125.hydrogen}
\title{module_energy_LA125.hydrogen}
\usage{
module_energy_LA125.hydrogen(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L125.globaltech_coef},  \code{L125.globaltech_cost},  \code{L125.Electrolyzer_IdleRatio_Params}.
}
\description{
Provides supply sector information, subsector information, technology information for hydrogen sectors.
}
\details{
Takes inputs from H2A and generates GCAM's assumptions by technology and year
}
\author{
GPK/JF/PW July 2021


# Note:  NREL H2A v2018 did not include the following H2 production technologies:
}
