% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA1323.iron_steel.R
\name{module_energy_LA1323.iron_steel}
\alias{module_energy_LA1323.iron_steel}
\title{module_energy_LA1323.iron_steel}
\usage{
module_energy_LA1323.iron_steel(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1323.out_Mt_R_iron_steel_Yh}, \code{L1323.IO_GJkg_R_iron_steel_F_Yh}, \code{L1323.in_EJ_R_iron_steel_F_Y}, \code{L1323.in_EJ_R_indenergy_F_Yh}. The corresponding file in the
original data system was \code{LA1323.iron_steel.R} (energy level1).
}
\description{
Sets up input, output, and IO coefficients for iron and steel and subtracts input energy from industry energy use
}
\details{
This chunk generates input, output, and IO coefficients for the iron and steel sector. It begins by downscaling Worrell regional data from 1994
to set up process emissions factors that are multiplied by country emissions from CDIAC to determine production. Limestone consumption is calculated from the same downscaling.
IEA fuelshares and heat and electricity are used to determine energy use by fuel. Energy inputs are then subtracted from industrial energy use and any resulting negative values
are dealt with by moving their accounting to the iron and steel sector.
}
\author{
Yang Liu Sep 2019
}
