% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA161.Cstorage.R
\name{module_energy_LA161.Cstorage}
\alias{module_energy_LA161.Cstorage}
\title{module_energy_LA161.Cstorage}
\usage{
module_energy_LA161.Cstorage(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L161.RsrcCurves_MtC_R}. The corresponding file in the
original data system was \code{LA161.Cstorage.R} (energy level1).
}
\description{
Build carbon storage supply curves by region.
}
\details{
Build carbon storage supply curves by region.
}
\author{
RLH November 2017
}
