% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA162.dac.R
\name{module_energy_LA162.dac}
\alias{module_energy_LA162.dac}
\title{module_energy_LA162.dac}
\usage{
module_energy_LA162.dac(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L162.out_Mt_R_dac_Yh},
}
\description{
Sets up input, output, and IO coefficients for dac
}
\details{
This chunk defines historical demand for the climate engineering services sector, setting up structure for future competition.
For all historical years, is met entirely by "no DAC" technology
}
\author{
JF March 2021
}
