% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB2011.ff_ALL_R_C_Y.R
\name{module_energy_LB2011.ff_ALL_R_C_Y}
\alias{module_energy_LB2011.ff_ALL_R_C_Y}
\title{module_energy_LB2011.ff_ALL_R_C_Y}
\usage{
module_energy_LB2011.ff_ALL_R_C_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2011.ff_ALL_EJ_R_C_Y} \code{L2011.ff_GrossTrade_EJ_R_C_Y}.
}
\description{
Calculate fossil fuel energy balances, by region / commodity / year and harmonize GCAM's data to Comtrade's.
}
\details{
This chunk combines fossil fuel production and consumption to calculate energy balances by
GCAM region, commodity and year. After calculating GCAM's (assumed) net trade
Comtrade's gross trade is adjusted to match.
}
\author{
JEH Nov 2019
}
