% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_xml_other_industry.R
\name{module_energy_other_industry_xml}
\alias{module_energy_other_industry_xml}
\title{module_energy_other_industry_xml}
\usage{
module_energy_other_industry_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{other_industry.xml}. The corresponding file in the
original data system was \code{batch_industry_xml.R} (energy XML).
}
\description{
Construct XML data structure for \code{other_industry.xml}.
}
