% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_xml_resources_SSP.R
\name{module_energy_resources_SSP_xml}
\alias{module_energy_resources_SSP_xml}
\title{module_energy_resources_SSP_xml}
\usage{
module_energy_resources_SSP_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{resources_SSP1.xml}, \code{resources_SSP2.xml}, \code{resources_SSP3.xml}, \code{resources_SSP4.xml}, \code{resources_SSP5.xml}.
}
\description{
Construct XML data structure for all the \code{resources_SSP.xml} files.
}
