% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenergy_xml_turn_off_ccs.R
\name{module_energy_turn_off_ccs_xml}
\alias{module_energy_turn_off_ccs_xml}
\title{module_energy_turn_off_ccs_xml}
\usage{
module_energy_turn_off_ccs_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{turn_off_ccs.xml}.
}
\description{
Construct XML data structure for \code{turn_off_ccs.xml}.
}
\author{
EL 2022
}
