% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_X232.industry_Subregions_Thailand.R
\name{module_energy_X232.industry_Subregions_Thailand}
\alias{module_energy_X232.industry_Subregions_Thailand}
\title{module_energy_X232.industry_Subregions_Thailand}
\usage{
module_energy_X232.industry_Subregions_Thailand(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: check them out.
}
\description{
Creates level2 data for the industry sector.
}
\details{
Creates level2 data for the industry sector.
}
\author{
ZK 4 June 2021
}
