% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/zchunk_X254.transportation_KualaLumpur_Malaysia.R
\name{module_energy_X254.transportation_KualaLumpur_Malaysia}
\alias{module_energy_X254.transportation_KualaLumpur_Malaysia}
\title{module_energy_X254.transportation}
\usage{
module_energy_X254.transportation_KualaLumpur_Malaysia(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: check them out.
The corresponding file in the original data system was \code{L244.building_det.R} (energy level2).
}
\description{
Creates level2 data for the building sector.
}
\details{
Creates level2 data for the building sector.
}
\author{
ZK 11 June 2021
}
