% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/zchunk_Xbatch_building_xml_KualaLumpur_Malaysia.R
\name{module_energy_Xbatch_building_xml_KualaLumpur_Malaysia}
\alias{module_energy_Xbatch_building_xml_KualaLumpur_Malaysia}
\title{module_energy_Xbatch_building_xml_KualaLumpur_Malaysia}
\usage{
module_energy_Xbatch_building_xml_KualaLumpur_Malaysia(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{building_KualaLumpur_Malaysia.xml}. The corresponding file in the
original data system was \code{batch_building_KualaLumpur_Malaysia.xml} (energy XML).
}
\description{
Construct XML data structure for \code{building_KualaLumpur_Malaysia.xml}.
}
