% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/zchunk_Xbatch_industry_xml_Subregions_Malaysia.R
\name{module_energy_Xbatch_industry_xml_Subregions_Malaysia}
\alias{module_energy_Xbatch_industry_xml_Subregions_Malaysia}
\title{module_energy_Xbatch_industry_xml_Subregions_Malaysia}
\usage{
module_energy_Xbatch_industry_xml_Subregions_Malaysia(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{industry_Subregions_Malaysia}. The corresponding file in the
original data system was \code{batch_industry_Subregions_Malaysia} (energy XML).
}
\description{
Construct XML data structure for \code{industry_Subregions_Malaysia}.
}
