% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/zchunk_Xbatch_transportation_xml_Bangkok_Thailand.R
\name{module_energy_Xbatch_transportation_xml_Bangkok_Thailand}
\alias{module_energy_Xbatch_transportation_xml_Bangkok_Thailand}
\title{module_energy_Xbatch_transportation_xml_Bangkok_Thailand}
\usage{
module_energy_Xbatch_transportation_xml_Bangkok_Thailand(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{transportation_Bangkok_Thailand.xml}. The corresponding file in the
original data system was \code{batch_transportation_Bangkok_Thailand} (energy XML).
}
\description{
Construct XML data structure for \code{transportation_Bangkok_Thailand}.
}
