% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L100.Socioeconomics_preliminary_CHINA.R
\name{module_gcam.china_L100.Socioeconomics_preliminary}
\alias{module_gcam.china_L100.Socioeconomics_preliminary}
\title{module_gcam.china_L100.Socioeconomics_preliminary}
\usage{
module_gcam.china_L100.Socioeconomics_preliminary(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.population_province_SSP}, \code{L100.gdp_province_SSP},
}
\description{
This chunk generates historical and future GDP and Population data by province
}
\details{
This chunk generated historical and future GDP and population data by province
}
\author{
MAW September 2020
}
