% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L101.Socioeconomics_china.R
\name{module_gcam.china_L101.Socioeconomics}
\alias{module_gcam.china_L101.Socioeconomics}
\title{module_gcam.china_L101.Socioeconomics}
\usage{
module_gcam.china_L101.Socioeconomics(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L101.pcGDP_thous90usd_province}, \code{L101.GDP_mil90usd_province}, \code{L101.Pop_thous_province}. The corresponding file in the
original data system was \code{LA100.Socioeconomics.R} (gcam-china level1).
}
\description{
This chunk generates the historical and future GDP, population and per-capita GDP by province.
}
\details{
This chunk processes historical and future GDP and population data by province,
fill missing values and covert units; and then calculates per capita GDP by province as
GDP divided by population.
}
\author{
RC Jul 2018
}
