% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L201.socioeconomics_china.R
\name{module_gcam.china_L201.socioeconomics}
\alias{module_gcam.china_L201.socioeconomics}
\title{module_gcam.china_L201.socioeconomics}
\usage{
module_gcam.china_L201.socioeconomics(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L201.InterestRate_CHINA}, \code{L201.Pop_GCAMCHINA}, \code{L201.BaseGDP_GCAMCHINA}, \code{L201.LaborForceFillout_CHINA}, \code{L201.LaborProductivity_GCAMCHINA}. The corresponding file in the
original data system was \code{L201.socioeconomics_CHINA.R} (gcam-China level2).
}
\description{
Interest rate, population, and GDP for GCAM-China.
}
\details{
Interest rate, population, and GDP for GCAM-China.
}
\author{
YangLiu Aug 2018
}
