% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L210.Resources_china.R
\name{module_gcam.china_L210.Resources_china}
\alias{module_gcam.china_L210.Resources_china}
\title{module_gcam.china_L210.Resources_china}
\usage{
module_gcam.china_L210.Resources_china(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L210.RenewRsrc_CHINA},\code{L210.UnlimitRsrc_CHINA},\code{L210.UnlimitRsrc_limestone_CHINA},
\code{L210.UnlimitRsrcPrice_CHINA}, \code{L210.UnlimitRsrcPrice_limestone_CHINA},\code{L210.SmthRenewRsrcCurves_wind_CHINA},
The corresponding file in the
original data system was \code{L210.resources_CHINA.R} (gcam-china level2).
}
\description{
GCAM-CHINA resource market information, prices, TechChange parameters, and supply curves.
}
\details{
GCAM-China resource market information, prices, TechChange parameters, and supply curves.
}
\author{
BY June 2019
}
