% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L223.electricity_CHINA.R
\name{module_gcam.china_L223.electricity_China}
\alias{module_gcam.china_L223.electricity_China}
\title{module_gcam.china_L223.electricity_China}
\usage{
module_gcam.china_L223.electricity_China(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L223.DeleteSubsector_CHINAelec},
\code{L223.Supplysector_CHINAelec}, \code{L223.SubsectorShrwtFllt_CHINAelec}, \code{L223.SubsectorInterp_CHINAelec},
\code{L223.SubsectorLogit_CHINAelec}, \code{L223.TechShrwt_CHINAelec}, \code{L223.TechCoef_CHINAelec},
\code{L223.Production_CHINAelec},
\code{L223.Supplysector_elec_GRIDR}, \code{L223.SubsectorShrwtFllt_elec_GRIDR}, \code{L223.SubsectorInterp_elec_GRIDR},
\code{L223.SubsectorLogit_elec_GRIDR}, \code{L223.TechShrwt_elec_GRIDR}, \code{L223.TechCoef_elec_GRIDR},
\code{L223.Production_elec_GRIDR}, \code{L223.InterestRate_GRIDR}, \code{L223.Pop_GRIDR}, \code{L223.BaseGDP_GRIDR},
\code{L223.LaborForceFillout_GRIDR},\code{L223.Supplysector_elec_CHINA}, \code{L223.ElecReserve_CHINA},
\code{L223.SubsectorLogit_elec_CHINA}, \code{L223.SubsectorShrwtFllt_elec_CHINA}, \code{L223.SubsectorShrwt_nuc_CHINA},
\code{L223.SubsectorShrwt_renew_CHINA}, \code{L223.SubsectorInterp_elec_CHINA}, \code{L223.SubsectorInterpTo_elec_CHINA},
\code{L223.StubTech_elec_CHINA}, \code{L223.StubTechEff_elec_CHINA}, \code{L223.StubTechCapFactor_elec_CHINA},
\code{L223.StubTechFixOut_elec_CHINA}, \code{L223.StubTechFixOut_hydro_CHINA}, \code{L223.StubTechProd_elec_CHINA},
\code{L223.StubTechMarket_elec_CHINA}, \code{L223.StubTechMarket_backup_CHINA}, \code{L223.StubTechElecMarket_backup_CHINA},
\code{L223.StubTechCapFactor_elec_wind_CHINA}, \code{L223.StubTechCapFactor_elec_solar_CHINA}. The corresponding file in the
original data system was \code{L223.electricity_CHINA.R} (gcam-CHINA level2).
}
\description{
Generates GCAM-China model inputs for electrcity sector by grid regions and provinces.
}
\details{
This chunk generates input files to create an annualized electricity generation sector for each province
and creates the demand for the province-level electricity sectors in the grid regions.
}
\author{
Yisheng SUN Jan 2020
}
