% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lu-L2231.electricity_nuclear_CHINA.R
\name{module_gcam.china_L2231.electricity_nuclear_CHINA}
\alias{module_gcam.china_L2231.electricity_nuclear_CHINA}
\title{module_gcam.china_L2231.electricity_nuclear_CHINA}
\usage{
module_gcam.china_L2231.electricity_nuclear_CHINA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2231.StubTechFixOut_elec_nuclear_CHINA}. The corresponding file in the
original data system was \code{L2231.electricity_nuclear_CHINA.R} (gcam-china level2).
}
\description{
Fixoutput of Electricity generation from nuclear by China province
}
\details{
Fixoutput of Electricity generation from nuclear by China province from 2015 to 2030, adding StubTechnologies
}
\author{
LuRen July 2019
}
