% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L225.hydrogen_CHINA.R
\name{module_gcam.china_L225.hydrogen_china}
\alias{module_gcam.china_L225.hydrogen_china}
\title{module_gcam.china_L225.hydrogen_china}
\usage{
module_gcam.china_L225.hydrogen_china(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L225.DeleteSubsector_h2_CHINA}. The corresponding file in the
original data system was \code{L225.hydrogen_CHINA.R} (gcam-china level2).
}
\description{
Selects the subsectors to be removed from the hydrogen sectors in GCAM-China
}
\details{
This chunk selects the subsectors to be removed from the hydrogen sectors in GCAM-China at the national level.
}
\author{
YangLiu September 2018
}
