% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L226.en_distribution_CHINA.R
\name{module_gcam.china_L226.en_distribution_CHINA}
\alias{module_gcam.china_L226.en_distribution_CHINA}
\title{module_gcam.china_L226.en_distribution_CHINA}
\usage{
module_gcam.china_L226.en_distribution_CHINA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs:
\itemize{
\item{\code{L226.DeleteSupplysector_CHINAelec}: Removing the electricity T&D sectors of the CHINA region.}
\item{\code{L226.StubTechCoef_electd_CHINA}: Stub technology coefficients elec T&D when using national elec markets. State elect_td sectors are treated as stub technologies.}
\item{\code{L226.TechShrwt_electd_CHINA}: tech share weights for elec T&D when using regional electricity markets. The elect_td sectors can not use the global tech database as their input is different.}
\item{\code{L226.TechCost_electd_CHINA}: Tech costs for elec T&D when using regional electricity markets.}
\item{\code{L226.Supplysector_electd_CHINA}: CHINA supply sector input, output, and logit info for elec T&D by province.}
\item{\code{L226.SubsectorShrwtFllt_electd_CHINA}: CHINA subsector shareweight fillout for elec T&D by state.}
\item{\code{L226.SubsectorLogit_electd_CHINA}: CHINA subsector logit info for elec T&D by grid.region.}
\item{\code{L226.SubsectorInterp_electd_CHINA}: CHINA interpolation info for elec T&D by province.}
\item{\code{L226.TechCoef_electd_CHINA}: Tech coeff for elec T&D when using regional electricity markets.}
\item{\code{L226.TechShrwt_en_CHINA}: Technology shareweights of energy handling and delivery. Can't use stub technologies because these would inherit the wrong energy-inputs.}
\item{\code{L226.TechCost_en_CHINA}: Regional price adjustments/cost adders for CHINA energy.}
\item{\code{L226.TechCoef_en_CHINA}: Technology coefficients and market names of energy handling and delivery.}
\item{\code{L226.Supplysector_en_CHINA}: Supply sector information for energy handling and delivery sectors for CHINA grid regions.}
\item{\code{L226.SubsectorShrwtFllt_en_CHINA}: Subsector shareweights of energy handling and delivery.}
\item{\code{L226.SubsectorLogit_en_CHINA}: Logit info for energy subsectors.}
\item{\code{L226.Ccoef_CHINA}: Carbon coef for CHINA cost adder sectors.}
}
The corresponding file in the original data system was \code{L226.en_distribution_CHINA.R} (gcam-china level2).
}
\description{
Create a variety of energy and electricity outputs for CHINA at the province and/or grid.region level.
}
\details{
Create a variety of energy and electricity outputs for CHINA at the province and/or grid.region level.
}
\author{
LuRen Dec 2019, edited by BY Feb 2020
}
