% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L232.industry_CHINA.R
\name{module_gcam.china_L232.industry_CHINA}
\alias{module_gcam.china_L232.industry_CHINA}
\title{module_gcam.china_L232.industry_CHINA}
\usage{
module_gcam.china_L232.industry_CHINA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L232.StubTechCalInput_indenergy_CHINA}, \code{L232.StubTechCalInput_indfeed_CHINA},
\code{L232.L232.StubTechProd_industry_CHINA}, \code{L232.StubTechCoef_industry_CHINA}, \code{L232.StubTechMarket_ind_CHINA},
\code{L232.BaseService_ind_CHINA}, \code{L232.DeleteSubsector_ind_CHINA}, \code{L232.Supplysector_ind_CHINA},
\code{L232.FinalEnergyKeyword_ind_CHINA}, \code{L232.SubsectorLogit_ind_CHINA},
\code{L232.SubsectorShrwtFllt_ind_CHINA}, \code{L232.SubsectorInterp_ind_CHINA},
\code{L232.StubTech_ind_CHINA}, \code{L232.StubTechInterp_ind_CHINA},
\code{L232.PerCapitaBased_ind_CHINA}, \code{L232.PriceElasticity_ind_CHINA}, \code{L232.IncomeElasticity_ind_gcam3_CHINA}.
The corresponding file in the original data system was \code{L232.industry_CHINA.R} (gcam-china level2).
}
\description{
Prepare level 2 industry sector files for CHINA.
}
\details{
Prepare level 2 industry sector files for CHINA.
}
\author{
YangLiu December 2019
}
