% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L244.building_CHINA.R
\name{module_gcam.china_L244.building_CHINA}
\alias{module_gcam.china_L244.building_CHINA}
\title{module_gcam.china_L244.building_CHINA}
\usage{
module_gcam.china_L244.building_CHINA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\description{
Creates GCAM-CHINA building output files for writing to xml.
}
\details{
Creates GCAM-CHINA building output files for writing to xml.
}
\author{
BY January 2020
}
