% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L261.carbon_storage_CHINA.R
\name{module_gcam.china_L261.carbon_storage_CHINA}
\alias{module_gcam.china_L261.carbon_storage_CHINA}
\title{module_gcam.china_L261.carbon_storage_CHINA}
\usage{
module_gcam.china_L261.carbon_storage_CHINA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L261.DeleteSubsector_CHINAC}, \code{L261.Rsrc_province}, \code{L261.RsrcCurves_province},
\code{L261.Supplysector_C_CHINA},\code{L261.SubsectorLogit_C_CHINA},\code{L261.SubsectorShrwtFllt_C_CHINA},
\code{L261.StubTech_C_CHINA},\code{L261.StubTechMarket_C_CHINA},
\code{L261.DeleteRsrc_CHINAC},\code{L261.UnlimitRsrc_CHINA}, \code{L261.GlobalTechCost_C_CHINA},\code{L261.ResTechShrwt_C_CHINA}
The corresponding file in the
original data system was \code{L261_carbon_storage_CHINA.R} (gcam-china level2).
}
\description{
China carbon storage
}
\details{
China carbon storage model input for GCAM-China energy electricity sectors
}
\author{
BY Feb 2020
}
