% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2999.ccs_shrwt_CHINA.R
\name{module_gcam.china_L2999.ccs_shrwt_CHINA}
\alias{module_gcam.china_L2999.ccs_shrwt_CHINA}
\title{module_gcam.china_L2999.ccs_shrwt_CHINA}
\usage{
module_gcam.china_L2999.ccs_shrwt_CHINA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs:  \code{L9999.StubTechShrwt_CHINA}. The corresponding file in the
original data system was \code{L2999.ccs_shrwt_CHINA.R} (gcam-china level2).
}
\description{
Assign shareweight assumptions of CCS to subtechnologies by China province
}
\details{
Assign shareweight assumptions of CCS to subtechnologies by China province. Remove N fertilizer sector from Tibet(XZ) province.
}
\author{
YishengSun July 2019
}
