% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA100.Socioeconomics_china.R
\name{module_gcam.china_LA100.Socioeconomics}
\alias{module_gcam.china_LA100.Socioeconomics}
\title{module_gcam.china_LA100.Socioeconomics}
\usage{
module_gcam.china_LA100.Socioeconomics(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.pcGDP_thous90usd_province}, \code{L100.GDP_mil90usd_province}, \code{L100.Pop_thous_province}. The corresponding file in the
original data system was \code{LA100.Socioeconomics.R} (gcam-china level1).
}
\description{
This chunk calculates the historical GDP, per-capita GDP and population by province.
}
\details{
Describe in detail what this chunk does.
}
\author{
RC Jul 2018
}
