% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA101.Energy_Balance_china.R
\name{module_gcam.china_LA101.Energy_Balance}
\alias{module_gcam.china_LA101.Energy_Balance}
\title{module_gcam.china_LA101.Energy_Balance}
\usage{
module_gcam.china_LA101.Energy_Balance(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L101.NBS_use_all_Mtce}, \code{L101.inNBS_Mtce_province_S_F}. The corresponding file in the
original data system was \code{LA101.Energy_Balance.R} (gcam-china level1).
}
\description{
This chunk generates the historical energy data by province
}
\details{
This chunk processes historical energy data by province, interpolate missinge value
and then estimates of Tibet shares of energy use by sector.  A rough estimate to give scale
Tibet accounts for 1% of the national energy
}
\author{
Yang July 2018
}
