% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA114.wind_china.R
\name{module_gcam.china_LA114.Wind}
\alias{module_gcam.china_LA114.Wind}
\title{module_gcam.china_LA114.Wind}
\usage{
module_gcam.china_LA114.Wind(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L114.CapacityFactor_wind_province}. The corresponding file in the
original data system was \code{LA114.Wind.R} (gcam-usa level1).
}
\description{
Compute capacity factors for wind by China province.
}
\details{
Computes capacity factors for wind by US state.
}
\author{
ST September 2017
}
