% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA119.solar_china.R
\name{module_gcam.china_LA119.Solar}
\alias{module_gcam.china_LA119.Solar}
\title{module_gcam.china_LA119.Solar}
\usage{
module_gcam.china_LA119.Solar(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L119.CapFacScaler_PV_province}, \code{L119.CapFacScaler_CSP_province}. The corresponding file in the
original data system was \code{LA114.Solar.R} (gcam-china level1).
}
\description{
Compute capacity factors for solar capacity factors by province.
}
\details{
Computes solar capacity factor scalars for PV and CSP technologies by province, divided by the national average.
}
\author{
Liu July 2018
}
