% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA122.Refining_china.R
\name{module_gcam.china_LA122.Refining}
\alias{module_gcam.china_LA122.Refining}
\title{module_gcam.china_LA122.Refining}
\usage{
module_gcam.china_LA122.Refining(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L122.in_EJ_province_refining_F}, \code{L122.out_EJ_province_refining_F}. The corresponding file in the
original data system was \code{LA122.Refining.R} (gcam-china level1).
}
\description{
Downscales crude oil, corn ethanol, and biodiesel refining inputs and outputs to province-level.
}
\details{
Downscales crude oil, corn ethanol, and biodiesel refining inputs and outputs to province-level.
}
\author{
YangLiu Jul 2018
}
