% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA1322.Fert_china.R
\name{module_gcam.china_LA1322.Fert}
\alias{module_gcam.china_LA1322.Fert}
\title{module_gcam.china_LA1322.Fert}
\usage{
module_gcam.china_LA1322.Fert(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1322.out_Mt_province_Fert_Yh}, \code{L1322.IO_GJkg_province_Fert_F_Yh}, \code{L1322.in_EJ_province_Fert_Yh}. The corresponding file in the
original data system was \code{LA1322.Fert.R} (gcam-china level1).
}
\description{
Calculate input-output intensity coefficients and input energy for province fertilizer production
}
\details{
Calculate input-output intensity coefficients and input energy for province fertilizer production from province shares of national values.
}
\author{
YangLiu Aug 2018
}
