% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA144.building_CHINA.R
\name{module_gcam.china_LA144.Building_CHINA}
\alias{module_gcam.china_LA144.Building_CHINA}
\title{module_gcam.china_LA144.Building_CHINA}
\usage{
module_gcam.china_LA144.Building_CHINA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L144.in_EJ_province_bld_F_U} \code{L144.flsp_bm2_province_bld} The corresponding file in the
original data system was \code{LA142.Building.R} (gcam-china level1).
}
\description{
Downscaling each province and sector's shares of  building energy use by fuel
}
\details{
Buildings sector energy consumption
}
\author{
BY January 2020
}
