% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB123.Electricity_china.R
\name{module_gcam.china_LB123.Electricity}
\alias{module_gcam.china_LB123.Electricity}
\title{module_gcam.china_LB123.Electricity}
\usage{
module_gcam.china_LB123.Electricity(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L123.in_EJ_province_elec_F}, \code{L123.out_EJ_province_elec_F}, \code{L123.in_EJ_province_ownuse_elec}, \code{L123.out_EJ_province_ownuse_elec}. The corresponding file in the
original data system was \code{LB123.Electricity.R} (gcam-china level1).
}
\description{
Electricity sector inputs and outputs, and electricity ownuse.
}
\details{
Compute electricity sector inputs and outputs, and electricity ownuse.
}
\author{
Liu August 2018
}
