% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1231.Elec_tech_china.R
\name{module_gcam.china_LB1231.Elec_tech_china}
\alias{module_gcam.china_LB1231.Elec_tech_china}
\title{module_gcam.china_LB1231.Elec_tech_china}
\usage{
module_gcam.china_LB1231.Elec_tech_china(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1231.in_EJ_province_elec_F_tech}, \code{L1231.out_EJ_province_elec_F_tech}. The corresponding file in the
original data system was \code{LB1231.Elec_tech.R} (gcam-china level1).
}
\description{
Downscaling electricity by province/fuel to province/fuel/technology
}
\details{
Downscaling electricity by province/fuel to province/fuel/technology
}
\author{
YangLiu Sep 2018
}
