% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1232.Elec_subregions_china.R
\name{module_gcam.china_LB1232.Elec_subregions_china}
\alias{module_gcam.china_LB1232.Elec_subregions_china}
\title{module_gcam.china_LB1232.Elec_subregions_china}
\usage{
module_gcam.china_LB1232.Elec_subregions_china(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1232.out_EJ_sR_elec_CHINA}. The corresponding file in the
original data system was \code{LB1232.Elec_subregions.R} (gcam-china level1).
}
\description{
Aggregates China province electricity generation to electricity subregions.
}
\details{
Aggregates CHINA province electricity generation to electricity subregions.
}
\author{
YangLiu September 2018
}
