% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_socioeconomics_India_xml.R
\name{module_gcam.india_batch_socioeconomics_INDIA_xml}
\alias{module_gcam.india_batch_socioeconomics_INDIA_xml}
\title{module_gcam.india_batch_socioeconomics_INDIA_xml}
\usage{
module_gcam.india_batch_socioeconomics_INDIA_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{socioeconomics_INDIA.xml}.
}
\description{
Construct XML data structure for \code{socioeconomics_INDIA.xml}.
}
\author{
Malyan_Ankur__CEEW
}
