% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L201.socioeconomics_INDIA.R
\name{module_gcam.india_L201.socioeconomics_INDIA}
\alias{module_gcam.india_L201.socioeconomics_INDIA}
\title{module_gcam.india_L201.socioeconomics_INDIA}
\usage{
module_gcam.india_L201.socioeconomics_INDIA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L201.InterestRate_GCAMINDIA}, \code{L201.Pop_GCAMINDIA}, \code{L201.BaseGDP_GCAMINDIA}, \code{L201.LaborForceFillout_GCAMINDIA}, \code{L201.LaborProductivity_GCAMINDIA}.
}
\description{
Interest rate, population, and GDP for GCAM-INDIA.
}
\details{
Interest rate, population, and GDP for GCAM-INDIA.
}
\author{
Malyan_Ankur_CEEW
}
