% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA101.EB_India.R
\name{module_gcam.india_LA101.EB_ASI_MOSPI}
\alias{module_gcam.india_LA101.EB_ASI_MOSPI}
\title{module_gcam.india_LA101.EB_ASI_MOSPI}
\usage{
module_gcam.india_LA101.EB_ASI_MOSPI(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L101.in_EJ_state_S_F_India}, \code{L101.use_all_Mtoe_India}.
}
\description{
Produce two ouput tables from the GHG Platform (ASI-MOSPI combined data) state energy database:
\itemize{
 \item{L101.in_EJ_state_S_F_India: Energy data by GCAM sector and fuel, state, and year; energy units in EJ, years from 1971-2010, includes only rows that have a defined sector and fuel}
 \item{L101.use_all_Mtoe_India: Energy data by ASI-MOSPI sector and fuel, GCAM sector and fuel, state, and year; energy units in Million tonnes of oil equivalent (Mtoe), years from 1960-2011, includes all original data}
}
}
\details{
See above
}
\author{
Malyan_Ankur_CEEW
}
