% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA119.solar_INDIA.R
\name{module_gcam.india_LA119.solar}
\alias{module_gcam.india_LA119.solar}
\title{module_gcam.india_LA119.solar}
\usage{
module_gcam.india_LA119.solar(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L119.CapFacScaler_PV_state_India}, \code{L119.CapFacScaler_CSP_state_India}.
}
\description{
Compute scalars by state to vary capacity factors by state for India.
}
\details{
This chunk computes scalars by state to vary capacity factors for central station PV and CSP technologies by state.
}
\author{
Malyan_Ankur_CEEW
}
