% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2011.socioeconomics_update_USA.R
\name{module_gcam.usa_L2011.socioeconomics_update_USA}
\alias{module_gcam.usa_L2011.socioeconomics_update_USA}
\title{module_gcam.usa_L2011.socioeconomics_update_USA}
\usage{
module_gcam.usa_L2011.socioeconomics_update_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2011.Pop_updated_USA}, \code{L2011.BaseGDP_updated_USA}, \code{L2011.LaborProductivity_updated_USA},
\code{L2011.Pop_national_updated_USA}, \code{L2011.BaseGDP_national_updated_USA}, \code{L2011.LaborProductivity_national_updated_USA}.
The corresponding file in the original data system was \code{L2011.socioeconomics_USA_update.R} (gcam-usa level2).
}
\description{
Population and GDP updates for GCAM-USA.
}
\details{
Interest rate, population, and GDP for GCAM-USA.
}
\author{
MTB August 2018
}
