% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2323.industry_process_USA.R
\name{module_gcam.usa_L2323.industry_process_USA}
\alias{module_gcam.usa_L2323.industry_process_USA}
\title{module_gcam.usa_L2323.industry_process_USA}
\usage{
module_gcam.usa_L2323.industry_process_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2324.StubTechCoef_indproc_USA}.
The corresponding file in the original data system was \code{L2323.StubTechCoef_indproc_USA} (gcam-usa level2).
}
\description{
Creates industrial processes input for state-level industry sectors in order to track industrial processes emissions.
}
\details{
This chunk selects the subsectors to be removed from the hydrogen sectors in GCAM USA on the national level.
}
\author{
MTB September 2018
}
