% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L271.trn_nonco2_USA.R
\name{module_gcam.usa_L271.trn_nonco2_USA}
\alias{module_gcam.usa_L271.trn_nonco2_USA}
\title{module_gcam.usa_L271.trn_nonco2_USA}
\usage{
module_gcam.usa_L271.trn_nonco2_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L271.nonco2_trn_tech_coeff_USA} and \code{L271.nonco2_trn_emiss_control_USA}.
The corresponding file in the original data system was \code{L271.trn_nonco2_USA.R} (gcam-usa level2).
}
\description{
U.S. non CO2 emission coeff from the transportation sector
}
\details{
Estimate future emission coeff for the HDV and LDVs via linear interpolation.
}
\author{
KD September 2018
}
