% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/zchunk_LB1238.elec_load_segments_solver_1990_USA.R
\name{module_gcam.usa_LB1238.elec_load_segments_solver_1990_USA}
\alias{module_gcam.usa_LB1238.elec_load_segments_solver_1990_USA}
\title{module_gcam.usa_LB1238.elec_load_segments_solver_1990_USA}
\usage{
module_gcam.usa_LB1238.elec_load_segments_solver_1990_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1238.grid_elec_supply_USA}, \code{L1238.grid_elec_demand_USA}, \code{L1238.grid_check_USA}.

The corresponding file in the original data system was \code{LB1238.elec_load_segments_solver_1990.R} (gcam-usa level1).
}
\description{
Calculate the fraction of electricity generation by fuel by horizontal load segment in 1990 such that
the total supply of electricity in each grid region matches total demand of electricity in that grid.
}
\details{
Calculates the fraction of electricity generation by fuel, by horizontal load segment, by grid region, in 1990.
}
\author{
MTB August 2018
}
