% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yisheng-zchunk_batch_CCS_shrwt_CHINA_xml.R
\name{module_gcamchina_batch_CCS_shrwt_CHINA_xml}
\alias{module_gcamchina_batch_CCS_shrwt_CHINA_xml}
\title{module_gcam.china_batch_CCS_shrwt_CHINA_xml}
\usage{
module_gcamchina_batch_CCS_shrwt_CHINA_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{CCS_shrwt_CHINA.xml}. The corresponding file in the
original data system was \code{batch_CCS_shrwt_CHINA.xml} (gcam-china processing code-xml batch).
}
\description{
Construct XML data structure for \code{CCS_shrwt_CHINA.xml}.
}
