% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_dac_china_xml.R
\name{module_gcamchina_batch_dac_china_xml}
\alias{module_gcamchina_batch_dac_china_xml}
\title{module_gcamchina_batch_dac_china_xml}
\usage{
module_gcamchina_batch_dac_china_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{dac_china.xml}. The corresponding file in the
original data system was \code{batch_dac_china_xml.R} (gcamchina XML).
}
\description{
Construct XML data structure for \code{dac_china.xml}.
}
\author{
JF March 2021
}
