% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_interest_rate_CHINA_xml.R
\name{module_gcamchina_batch_interest_rate_CHINA_xml}
\alias{module_gcamchina_batch_interest_rate_CHINA_xml}
\title{module_gcamchina_batch_interest_rate_CHINA_xml}
\usage{
module_gcamchina_batch_interest_rate_CHINA_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{interest_rate_CHINA.xml}. The corresponding file in the
original data system was \code{batch_interest_rate_CHINA.xml} (gcamchina XML).
}
\description{
Construct XML data structure for \code{interest_rate_CHINA.xml}.
}
