% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_liquids_limits_CHINA_xml.R
\name{module_gcamchina_batch_liquids_limits_china_xml}
\alias{module_gcamchina_batch_liquids_limits_china_xml}
\title{module_gcamchina_batch_liquids_limits_china_xml}
\usage{
module_gcamchina_batch_liquids_limits_china_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{liquids_limits_CHINA.xml}.
}
\description{
Construct XML data structure for \code{liquids_limits_CHINA.xml}.
}
