% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2235.elec_segments_GRID_CHINA.R
\name{module_gcamchina_L2235.elec_segments_GRID_CHINA}
\alias{module_gcamchina_L2235.elec_segments_GRID_CHINA}
\title{module_gcamchina_L2235.elec_segments_GRID_CHINA}
\usage{
module_gcamchina_L2235.elec_segments_GRID_CHINA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2235.DeleteSupplysector_elec_CHINA}, \code{L2235.InterestRate_GRID_CHINA},
\code{L2235.Pop_GRID_CHINA}, \code{L2235.BaseGDP_GRID_CHINA}, \code{L2235.LaborForceFillout_GRID_CHINA},
\code{L2235.Supplysector_elec_CHINA}, \code{L2235.ElecReserve_elecS_grid_vertical_CHINA}, \code{L2235.SubsectorLogit_elec_CHINA},
\code{L2235.SubsectorShrwtFllt_elec_CHINA}, \code{L2235.SubsectorInterp_elec_CHINA},
\code{L2235.SubsectorShrwtFllt_elecS_grid_vertical_CHINA},\code{L2235.SubsectorShrwtInterp_elecS_grid_vertical_CHINA},
\code{L2235.TechShrwt_elec_CHINA}, \code{L2235.TechCoef_elec_CHINA}, \code{L2235.Production_exports_elec_CHINA},
\code{L2235.TechShrwt_elecS_grid_vertical_CHINA}, \code{L2235.TechCoef_elecS_grid_vertical_CHINA},
\code{L2235.Supplysector_elec_GRID_CHINA}, \code{L2235.SubsectorLogit_elec_GRID_CHINA}, \code{L2235.SubsectorShrwtFllt_elec_GRID_CHINA},
\code{L2235.SubsectorInterp_elec_GRID_CHINA}, \code{L2235.TechShrwt_elec_GRID_CHINA}, \code{L2235.TechCoef_elec_GRID_CHINA},
\code{L2235.TechCoef_elecownuse_GRID_CHINA}, \code{L2235.Production_imports_GRID_CHINA},\code{L2235.Production_elec_gen_GRID_CHINA}.
The corresponding file in the original data system was \code{L2235.elec_segments_FERC.R} (gcam-china level2).
}
\description{
Creates the vertical segment supplysectors in the grid regions and also the domestic supply and
electricity trade sectors.
}
\details{
This chunk generates input files to create the vertical segment supplysectors in the grid regions
and also the domestic supply and electricity trade sectors for the grid regions.
}
\author{
MTB Aug 2018 / YangOu Jul 2023
}
