% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2321.cement_CHINA.R
\name{module_gcamchina_L2321.cement_CHINA}
\alias{module_gcamchina_L2321.cement_CHINA}
\title{module_gcam.china_L2321.cement_CHINA}
\usage{
module_gcamchina_L2321.cement_CHINA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2321.DeleteSupplysector_CHINAcement}, \code{L2321.DeleteFinalDemand_CHINAcement}, \code{L2321.StubTechProd_cement_CHINA},
\code{L2321.StubTechCoef_cement_CHINA}, \code{L2321.StubTechCalInput_cement_heat_CHINA}, \code{L2321.StubTechMarket_cement_CHINA}, \code{L2321.BaseService_cement_CHINA},
\code{L2321.Supplysector_cement_CHINA}, \code{L2321.FinalEnergyKeyword_cement_CHINA}, \code{L2321.SubsectorLogit_cement_CHINA},
\code{L2321.SubsectorShrwtFllt_cement_CHINA}, \code{L2321.StubTech_cement_CHINA}, \code{L2321.PerCapitaBased_cement_CHINA},
 \code{L2321.PriceElasticity_cement_CHINA}, \code{L2321.IncomeElasticity_cement_gcam3_CHINA} .
 The corresponding file in the original data system was \code{L2321.cement_CHINA.R} (gcam-china level2).
}
\description{
Make the logit and input tables for the cement sector in gcam-china
}
\details{
Make the logit and input tables for the cement sector in gcam-china
}
\author{
YangLiu  December 2019
}
