% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2322.Fert_CHINA.R
\name{module_gcamchina_L2322.Fert_CHINA}
\alias{module_gcamchina_L2322.Fert_CHINA}
\title{module_gcamchina_L2322.Fert_CHINA}
\usage{
module_gcamchina_L2322.Fert_CHINA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2322.DeleteSubsector_CHINAFert}, \code{L2322.FinalEnergyKeyword_CHINAFert}, \code{L2322.FinalEnergyKeyword_Fert_CHINA}, \code{L2322.StubTech_Fert_CHINA}, \code{L2322.SubsectorLogit_CHINAFert}, \code{L2322.SubsectorShrwtFllt_CHINAFert}, \code{L2322.TechShrwt_CHINAFert}, \code{L2322.Production_CHINAFert}, \code{L2322.TechCoef_CHINAFer}, \code{L2322.StubTechProd_Fert_CHINA}, \code{L2322.StubTechCoef_Fert_CHINA}, \code{L2322.StubTechMarket_Fert_CHINA}, \code{L2322.SubsectorLogit_Fert_CHINA}, \code{L2322.Supplysector_Fert_CHINA}, \code{L2322.SubsectorShrwtFllt_Fert_CHINA}, \code{L2322.SubsectorInterp_Fert_CHINA},\code{L2322.SubsectorInterp_CHINAFert}
 The corresponding file in the original data system was \code{L2322.Fert_CHINA.R} (gcam-china level2).
}
\description{
Produce tables to create the N fertilizer sector in GCAM-CHINA.
}
\details{
This chunk produces tables to create the N fertilizer sector in GCAM-CHINA.
}
\author{
YangLiu December 2019
}
