% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2323.Detailed_industry_CHINA.R
\name{module_gcamchina_L2323.detailed_industry_CHINA}
\alias{module_gcamchina_L2323.detailed_industry_CHINA}
\title{module_gcamchina_L2323.detailed_industry_CHINA}
\usage{
module_gcamchina_L2323.detailed_industry_CHINA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2323.StubTechProd_detailed_industry},
\code{L2323.StubTechCoef_detailed_industry}, \code{L2323.StubTechCalInput_detailed_industry}, \code{L2323.StubTechMarket_detailed_industry}, \code{L2323.BaseService_detailed_industry},
\code{L2323.Supplysector_detailed_industry_CHINA}, \code{L2323.FinalEnergyKeyword_detailed_industry_CHINA}, \code{L2323.SubsectorLogit_detailed_industry_CHINA},
\code{L2323.SubsectorShrwtFllt_detailed_industry_CHINA}, \code{L2323.StubTech_detailed_industry_CHINA}, \code{L2323.PerCapitaBased_detailed_industry_CHINA},
\code{L2323.PriceElasticity_detailed_industry_CHINA}, \code{L2323.IncomeElasticity_detailed_industry_gcam3_CHINA} , \code{L2323.GlobalTechCSeq_ind}.
The corresponding file in the original data system was \code{L2323.detailed_industry_CHINA.R} (gcam-china level2).
}
\description{
Make the logit and input tables for the detailed_industry sector in gcam-china
}
\details{
Make the logit and input tables for the detailed_industry sector in gcam-china
}
\author{
YangLiu  September 2020
}
