% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L270.limits_CHINA.R
\name{module_gcamchina_L270.limits_CHINA}
\alias{module_gcamchina_L270.limits_CHINA}
\title{module_gcamchina_L270.limits_CHINA}
\usage{
module_gcamchina_L270.limits_CHINA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L270.CreditMkt_CHINA}, \code{L270.CreditInput_elecS_CHINA}, \code{L270.NegEmissBudgetMaxPrice_CHINA},
\code{paste0( "L270.NegEmissBudget_CHINA_", c("GCAM3", paste0("SSP", 1:5), paste0("gSSP", 1:5)) )}.
}
\description{
Add the provinces to the China market in each of the L270 limits polices.  In
particular to limit the fraction of liquid feedstocks and inputs to electricity
generation which can come from sources other than crude oil.  Constrain the
total amount of subsidy as a fraction of GDP which an economy is will to give
to have net negative emissions.
}
\details{
Add provinces to China market for GCAM policy constraints which enforce limits
to liquid feedstocks and the amount of subsidies given for net negative emissions.
}
\author{
BY Jun 2020
}
