% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1234.elec_gridregions_china.R
\name{module_gcamchina_LB1234.elec_gridregions_china}
\alias{module_gcamchina_LB1234.elec_gridregions_china}
\title{module_gcamchina_LB1234.elec_gridregions_china}
\usage{
module_gcamchina_LB1234.elec_gridregions_china(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1234.in_EJ_grid_elec_F_CHINA}, \code{L1234.out_EJ_grid_elec_F_CHINA}.
The corresponding file in the original data system was \code{LB1234.elec_gridregions.R} (gcam-china level1).
}
\description{
Calculate electricity fuel consumption and electricity generation by grid region.
}
\details{
By grid region, calculates electricity fuel consumption and electricity generation.
}
\author{
YangOu Jun 2023
}
