% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1235.elec_load_segments_china.R
\name{module_gcamchina_LB1235.elec_load_segments_china}
\alias{module_gcamchina_LB1235.elec_load_segments_china}
\title{module_gcamchina_LB1235.elec_load_segments_china}
\usage{
module_gcamchina_LB1235.elec_load_segments_china(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1235.grid_elec_supply_CHINA}, \code{L1235.elecS_horizontal_vertical_CHINA},
\code{L1235.elecS_horizontal_vertical_GCAM_coeff_CHINA}, \code{L1235.elecS_demand_fraction_CHINA}.

The corresponding file in the original data system was \code{LB1235.elec_load_segments.R} (gcam-usa level1).
}
\description{
Read in demand fraction and time fraction and compute load curve related parameters
Calculate an initial estimate of generation by fuel (EJ) in the horizontal segments
}
\details{
Compute load curve related parameters and initial estimate of generation by fuel in the horizontal segments.
}
\author{
YangOu Jun 2023
}
