% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1239.elec_province_fractions_china.R
\name{module_gcamchina_LB1239.elec_province_fractions_CHINA}
\alias{module_gcamchina_LB1239.elec_province_fractions_CHINA}
\title{module_gcamchina_LB1239.elec_province_fractions_CHINA}
\usage{
module_gcamchina_LB1239.elec_province_fractions_CHINA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1239.province_elec_supply_CHINA}.

The corresponding file in the original data system was \code{LB1239.elec_state_fractions.R} (gcam-china level1).
}
\description{
Map electricity generation by fuel | grid region | horizontal segment to generation by fuel | province | segment.
The fraction of generation by fuel by horizontal segment is assumed to be equal for all provinces within a grid region.
}
\details{
Calculates the fraction of electricity generation by fuel, by horizontal load segment, by province.
}
\author{
MTB August 2018 / YangOu July 2023
}
