% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB126.Gas_ElecTD_CHINA.R
\name{module_gcamchina_LB126.Gas_ElecTD}
\alias{module_gcamchina_LB126.Gas_ElecTD}
\title{module_gcamchina_LB126.Gas_ElecTD}
\usage{
module_gcamchina_LB126.Gas_ElecTD(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L126.in_EJ_province_td_elec}. The corresponding file in the
original data system was \code{LB126.Gas_ElecTD.R} (gcam-usa level1).
}
\description{
Calculates inputs and outputs of transmission and distribution of electricity by province.
}
\details{
Calculates inputs and outputs of transmission and distribution of electricity by province.
}
\author{
RLH September 2017 / YangOu 2023
}
