% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_xml_bio_trade.R
\name{module_gcameurope_bio_trade_xml}
\alias{module_gcameurope_bio_trade_xml}
\title{module_gcameurope_bio_trade_xml
This xml should fully replace the normal bio trade xml}
\usage{
module_gcameurope_bio_trade_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{bio_trade.xml}. The corresponding file in the
original data system was \code{batch_bio_trade.xml.R} (aglu XML).
}
\description{
Construct XML data structure for \code{bio_trade.xml}.
}
