% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_xml_elecS_ghg_emissions_water.R
\name{module_gcameurope_elecS_ghg_emissions_water_xml}
\alias{module_gcameurope_elecS_ghg_emissions_water_xml}
\title{module_gcameurope_elecS_ghg_emissions_water_xml}
\usage{
module_gcameurope_elecS_ghg_emissions_water_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs
}
\description{
Construct XML data structure for \code{elecS_ghg_emissions_water_EUR.xml}.
}
