% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_xml_en_supply.R
\name{module_gcameurope_en_supply_xml}
\alias{module_gcameurope_en_supply_xml}
\title{module_gcameurope_en_supply_xml}
\usage{
module_gcameurope_en_supply_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{en_supply_EUR.xml}. The corresponding file in the
original data system was \code{batch_en_supply_xml.R} (energy XML).
}
\description{
Construct XML data structure for \code{en_supply_EUR.xml}.
}
