% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_xml_HDDCDD.R
\name{module_gcameurope_HDDCDD_xml}
\alias{module_gcameurope_HDDCDD_xml}
\title{module_gcameurope_HDDCDD_xml}
\usage{
module_gcameurope_HDDCDD_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{HDDCDD_A2_CCSM3x_EUR.xml}, \code{HDDCDD_A2_HadCM3_EUR.xml}, \code{HDDCDD_B1_CCSM3x_EUR.xml},
\code{HDDCDD_B1_HadCM3_EUR.xml}, and \code{HDDCDD_constdd_no_GCM_EUR.xml}.
}
\description{
Construct XML data structure for all the \code{HDDCDD.xml} files.
}
