% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_xml_ind_urb_processing_sectors.R
\name{module_gcameurope_ind_urb_processing_sectors_xml}
\alias{module_gcameurope_ind_urb_processing_sectors_xml}
\title{module_gcameurope_ind_urb_processing_sectors_xml}
\usage{
module_gcameurope_ind_urb_processing_sectors_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{ind_urb_processing_sectors_EUR.xml}, \code{ind_urb_processing_sectors_MAC_EUR.xml}.
The corresponding file in the original data system was
\code{batch_ind_urb_processing_sectors.xml} (emissions XML).
}
\description{
Construct XML data structure for \code{ind_urb_processing_sectors_EUR.xml}.
}
