% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L101.en_bal_Eurostat.R
\name{module_gcameurope_L101.en_bal_Eurostat}
\alias{module_gcameurope_L101.en_bal_Eurostat}
\title{module_gcameurope_L101.en_bal_Eurostat}
\usage{
module_gcameurope_L101.en_bal_Eurostat(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L101.en_bal_EJ_R_Si_Fi_Yh_EUR}, \code{L101.in_EJ_R_trn_Fi_Yh_EUR}, \code{L101.in_EJ_R_bld_Fi_Yh_EUR}, \code{L101.GCAM_EUR_regions}..
}
\description{
Rename Eurostat products and flows to intermediate fuels and sectors used for constructing GCAM's fuel and sector calibration.
}
\details{
Assign Eurostat product and flow data to nomenclature used in GCAM (fuel and sector, respectively), summarizing
by (generally) iso and/or region, sector, fuel, and year.
}
\author{
RLH December 2023
}
