% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L101.nonghg_en_S_T_Y.R
\name{module_gcameurope_L101.nonghg_en_S_T_Y}
\alias{module_gcameurope_L101.nonghg_en_S_T_Y}
\title{module_gcameurope_L101.nonghg_en_S_T_Y}
\usage{
module_gcameurope_L101.nonghg_en_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L101.so2_tgej_EUR_en_Sepa_F_Yh_EUR}, \code{L101.co_tgej_EUR_en_Sepa_F_Yh_EUR}, \code{L101.nox_tgej_EUR_en_Sepa_F_Yh_EUR}, \code{L101.voc_tgej_EUR_en_Sepa_F_Yh_EUR}, \code{L101.nh3_tgej_EUR_en_Sepa_F_Yh_EUR}, \code{L101.in_EJ_R_en_Si_F_Yh_EUR}. The corresponding file in the
original data system was \code{L101.nonghg_en_EUR_S_T_Y.R} (emissions level1).
}
\description{
Compute historical emissions factors for energy by GCAM technology, from EPA emissions data and IEA energy balances.
}
\details{
Compute historical emissions factors for energy (electricity, independent energy, buildings,
transportation, fertilizer production, cement, heating, fossil production) by GCAM technology, from EPA
emissions data and IEA energy balances.
For NH3, 1990 data are used for the 1971-1989 period because these data don't go back
as far as 1971, so we extrapolate in order to retain the capacity to run the model from
any historical year (i.e., any year 1971-2010), representing subsequent historical years
as future model time periods in order to check model performance against the observed data.
}
\author{
CRB May 2024
}
