% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L1011.fossilFuel_GrossTrade_EUR.R
\name{module_gcameurope_L1011.ff_GrossTrade_EUR}
\alias{module_gcameurope_L1011.ff_GrossTrade_EUR}
\title{module_gcameurope_L1011.ff_GrossTrade_EUR}
\usage{
module_gcameurope_L1011.ff_GrossTrade_EUR(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs:
}
\description{
Calculate primary fossil fuel (only coal and crude oil) product balances, by region / commodity / year.
Natural gas trade is not adjusted, since it already has the pipeline structure in place.
}
\details{
This chunk processes the bi-lateral trade flow data matrix from Eurostat
}
\author{
RLH October 2024
}
