% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L110.water_demand_primary.R
\name{module_gcameurope_L110.water_demand_primary}
\alias{module_gcameurope_L110.water_demand_primary}
\title{module_water_L110.water_demand_primary}
\usage{
module_gcameurope_L110.water_demand_primary(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L110.water_demand_primary_R_S_W_m3_GJ_EUR}. The corresponding file in the
original data system was \code{L110.water_demand_primary.R} (water level1).
}
\description{
Gets water use coefficients for primary energy production.
}
\details{
Gets water use coefficients for primary energy production using regional
fractions of saline and freshwater shares.
}
\author{
SWDT April 2017
}
