% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L111.rsrc_fos_Prod.R
\name{module_gcameurope_L111.rsrc_fos_Prod}
\alias{module_gcameurope_L111.rsrc_fos_Prod}
\title{module_gcameurope_L111.rsrc_fos_Prod}
\usage{
module_gcameurope_L111.rsrc_fos_Prod(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L111.Prod_EJ_R_F_Yh_EUR}, \code{L111.RsrcCurves_EJ_R_Ffos_EUR}. \code{L210.DeleteRsrcTradBio_EUR}.
}
\description{
Calculate historical fossil energy production and fossil resource supply curves.
}
\details{
For historical fossil energy production, determine regional shares of production for each primary fuel,
interpolate unconventional oil production to all historical years, deduct unconventional oil from total oil,
and include it in calibrated production table.
}
\author{
RH January 2024
}
