% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L114.bcoc_en_R_S_T_Y.R
\name{module_gcameurope_L114.bcoc_en_R_S_T_Y}
\alias{module_gcameurope_L114.bcoc_en_R_S_T_Y}
\title{module_gcameurope_L114.bcoc_en_R_S_T_Y}
\usage{
module_gcameurope_L114.bcoc_en_R_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L114.bcoc_tgej_R_en_S_F_2000_EUR}. The corresponding file in the
original data system was \code{L114.bcoc_en_R_S_T_Y.R} (emissions level1).
}
\description{
Compute emissions coefficients for black carbon (BC) and organic carbon (OC) for a single representative base year
}
\details{
Computes black carbon (BC) and organic carbon (OC) emissions by RCP emissions dataset sector/source, and
divides by the corresponding drivers for the ~2000 base year in order to compute emissions coefficients. The coefs
are in units of Tg per EJ (or kg per GJ) of fuel input. Data are generated by GCAM region, sector (energy system only),
and fuel, for the year 2000.
}
\author{
GPK May 2017
}
