% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L121.liquids.R
\name{module_gcameurope_L121.liquids}
\alias{module_gcameurope_L121.liquids}
\title{module_gcameurope_L121.liquids}
\usage{
module_gcameurope_L121.liquids(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L121.in_EJ_R_unoil_F_Yh},
  \code{L121.in_EJ_R_TPES_crude_Yh}, \code{L121.in_EJ_R_TPES_unoil_Yh}, \code{L121.share_R_TPES_biofuel_tech}.
}
\description{
Downscale ethanol and biodiesel consumption to feedstock. Sets consumption of unconventional oil using GCAM3 assumptions.
}
\details{
This chunk uses data from IIASA to downscale ethanol and biodiesel consumption to modeled technologies and feedstocks. Also adjusts
unconventional oil and crude oil consumption, but assumes zero unconventional oil production in all of EU (could add for Estonia in future).
}
\author{
RH January 2024
}
