% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L122.gasproc_refining.R
\name{module_gcameurope_L122.gasproc_refining}
\alias{module_gcameurope_L122.gasproc_refining}
\title{module_gcameurope_L122.gasproc_refining}
\usage{
module_gcameurope_L122.gasproc_refining(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L122.out_EJ_R_gasproc_F_Yh}, \code{L122.in_EJ_R_gasproc_F_Yh}, \code{L122.IO_R_oilrefining_F_Yh}, \code{L122.out_EJ_R_refining_F_Yh}, \code{L122.in_EJ_R_refining_F_Yh}, \code{L122.in_Mt_R_C_Yh}, \code{L122.FeedOut_Mt_R_C_Yh}. The corresponding file in the
original data system was \code{LA122.gasproc_refining.R} (energy level1).
}
\description{
Create gasproc, oil refining and crops inputs, outputs and IO "input-output" coefficients for refining.
}
\details{
This chunk creates gasproc, oil refining and crops inputs, outputs and IO "input-output" coefficients for refining.
}
\author{
RH, February 2024
}
