% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L123.electricity.R
\name{module_gcameurope_L123.electricity}
\alias{module_gcameurope_L123.electricity}
\title{module_gcameurope_L123.electricity}
\usage{
module_gcameurope_L123.electricity(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L123.out_EJ_R_elec_F_Yh_EUR}, \code{L123.in_EJ_R_elec_F_Yh_EUR}, \code{L123.eff_R_elec_F_Yh_EUR}, \code{L123.out_EJ_R_indchp_F_Yh_EUR}, \code{L123.in_EJ_R_indchp_F_Yh_EUR}. The corresponding file in the
original data system was \code{LA123.electricity.R} (energy level1).
}
\description{
This script creates electricity generation and inputs by fuel, region and historical year. Estimates are adjusted by efficiency factors.
}
\details{
This script creates electricity generation and inputs by fuel, region and historical year. Estimates are adjusted by efficiency factors.
}
\author{
RH February 2024
}
