% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L1231.elec_tech.R
\name{module_gcameurope_L1231.elec_tech}
\alias{module_gcameurope_L1231.elec_tech}
\title{module_gcameurope_L1231.elec_tech}
\usage{
module_gcameurope_L1231.elec_tech(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1231.in_EJ_R_elec_F_tech_Yh_EUR}, \code{L1231.out_EJ_R_elec_F_tech_Yh_EUR}, \code{L1231.eff_R_elec_F_tech_Yh_EUR}. The corresponding file in the
original data system was \code{LA1231.elec_tech.R} (energy level1).
}
\description{
Obtain inputs, outputs and efficiencies in the electricity sector for all technologies by region, sector, and year.
}
\details{
This chunk obtaines inputs, outputs and efficiencies in the electricity sector for all technologies by region, sector, and year.
}
\author{
RH, February 2024
}
