% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L1232.Elec_subregions.R
\name{module_gcameurope_L1232.Elec_subregions}
\alias{module_gcameurope_L1232.Elec_subregions}
\title{module_gcameurope_L1232.Elec_subregions}
\usage{
module_gcameurope_L1232.Elec_subregions(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1232.out_EJ_sR_elec}. The corresponding file in the
original data system was \code{LB1232.Elec_subregions.R} (gcam-usa level1).
}
\description{
Aggregates European country electricity generation to electricity subregions.
}
\details{
Aggregates USA state electricity generation to electricity subregions.
}
\author{
RLH September 2017
}
