% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L1233.elec_water_grid.R
\name{module_gcameurope_L1233.elec_water_grid}
\alias{module_gcameurope_L1233.elec_water_grid}
\title{module_gcameurope_L1233.elec_water_grid}
\usage{
module_gcameurope_L1233.elec_water_grid(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1233.out_EJ_ctry_elec_F_tech_cool_EUR}, \code{L1233.share_sR_elec_F_tech_cool_EUR}. The corresponding file in the
original data system was \code{LB1233.Elec_water.R} (gcam-usa level1).
}
\description{
Compute water withdrawals/consumption by state, fuel, technology, and cooling system type.
}
\details{
Compute water withdrawals/consumption by state, fuel, technology, and cooling system type.
}
\author{
ST October 2017, NTG May 2020, RLH June 2024
}
