% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L1235.elec_load_segments.R
\name{module_gcameurope_L1235.elec_load_segments}
\alias{module_gcameurope_L1235.elec_load_segments}
\title{module_gcameurope_L1235.elec_load_segments}
\usage{
module_gcameurope_L1235.elec_load_segments(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs
}
\description{
Read in demand fraction and time fraction and compute load curve related parameters
Calculate an initial estimate of generation by fuel (EJ) in the horizontal segments
}
\details{
Compute load curve related parameters and nitial estimate of generation by fuel in the horizontal segments.
}
\author{
MTB August 2018
}
