% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L124.heat.R
\name{module_gcameurope_L124.heat}
\alias{module_gcameurope_L124.heat}
\title{module_gcameurope_L124.heat}
\usage{
module_gcameurope_L124.heat(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L124.in_EJ_R_heat_F_Yh_EUR}, \code{L124.out_EJ_R_heat_F_Yh_EUR}, \code{L124.out_EJ_R_heatfromelec_F_Yh_EUR}, \code{L124.heatoutratio_R_elec_F_tech_Yh_EUR}.
}
\description{
Process historical heat data into input, output from district heat, output from CHP, CHP ratio
}
\details{
This chunk takes in the energy balance, electricity output and mapping files on technology and fuel aggregations.
From this, it creates tables of heat input and output and uses the electricity gen. to determine heat from CHP.
}
\author{
RH February 2024
}
