% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L126.distribution.R
\name{module_gcameurope_L126.distribution}
\alias{module_gcameurope_L126.distribution}
\title{module_gcameurope_L126.distribution}
\usage{
module_gcameurope_L126.distribution(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L126.in_EJ_R_elecownuse_F_Yh_EUR}, \code{L126.out_EJ_R_elecownuse_F_Yh_EUR}, \code{L126.IO_R_elecownuse_F_Yh_EUR}, \code{L126.in_EJ_R_electd_F_Yh_EUR}, \code{L126.out_EJ_R_electd_F_Yh_EUR}, \code{L126.IO_R_electd_F_Yh_EUR}, \code{L126.in_EJ_R_gaspipe_F_Yh_EUR}, \code{L126.out_EJ_R_gaspipe_F_Yh_EUR}, \code{L126.IO_R_gaspipe_F_Yh_EUR}. The corresponding file in the
original data system was \code{LA126.distribution.R} (energy level1).
}
\description{
Adjust for onsite, transmission, and distribution losses for electricity and gas. Energy input, output, and input/output ratio are calculated for the following three processes:
\itemize{
 \item{Electricity net ownuse, i.e., electricity consumed onsite prior to any transmission and distribution losses}
 \item{Electricity transmission and distribution}
 \item{Gas pipeline}
}
Nine tables are created in total, 3x3=9. Energy is reported by GCAM_region, sector, fuel, and from 1971-2010.
}
\details{
Describe in detail what this chunk does.
}
\author{
RH February 2024
}
