% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L126.distribution_grid_regions.R
\name{module_gcameurope_L126.distribution_grid_regions}
\alias{module_gcameurope_L126.distribution_grid_regions}
\title{module_gcameurope_L126.distribution_grid_regions}
\usage{
module_gcameurope_L126.distribution_grid_regions(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs
}
\description{
Adjust distribution losses for electricity for regions in grid regions. These losses do not need to balance because of trade
}
\details{
Describe in detail what this chunk does.
}
\author{
RH February 2024
}
