% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L1261.elec_trade.R
\name{module_gcameurope_L1261.elec_trade}
\alias{module_gcameurope_L1261.elec_trade}
\title{module_gcameurope_L1261.elec_trade}
\usage{
module_gcameurope_L1261.elec_trade(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs
}
\description{
Adjust electricity ownuse by net trade with regions outside of grid. Unadjusted trade balances.
}
\details{
Describe in detail what this chunk does.
}
\author{
RH June 2024
}
