% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L131.enduse.R
\name{module_gcameurope_L131.enduse}
\alias{module_gcameurope_L131.enduse}
\title{module_gcameurope_L131.enduse}
\usage{
module_gcameurope_L131.enduse(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L131.in_EJ_R_Senduse_F_Yh_EUR}, \code{L131.share_R_Senduse_heat_Yh_EUR}. The corresponding file in the
original data system was \code{LA131.enduse.R} (energy level1).
}
\description{
Generate the following two tables:
\itemize{
 \item{Final scaled energy input by GCAM region / end-use sector, incl CHP / fuel / historical year; and}
 \item{Share of heat consumption by end-use sector within GCAM region / historical year}
}
}
\details{
Describe in detail what this chunk does.
}
\author{
RH February 2024
}
