% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L132.industry.R
\name{module_gcameurope_L132.industry}
\alias{module_gcameurope_L132.industry}
\title{module_gcameurope_L132.industry}
\usage{
module_gcameurope_L132.industry(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L132.in_EJ_R_indenergy_F_Yh_EUR}, \code{L132.in_EJ_R_indfeed_F_Yh_EUR}. The corresponding file in the
original data system was \code{LA132.industry.R} (energy level1).
}
\description{
Provides industrial energy consumption (not including CHP) and industrial feedstock consumption by GCAM region/fuel/historical year.
}
\details{
The chunk calculates industrial feedstock consumption directly separated from end-use energy consumption and
         industrial energy consumption by deducting net energy use by unconventional oil production, gas processing, refining, and CHP
}
\author{
LF September 2017
}
