% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L1322.Fert.R
\name{module_gcameurope_L1322.Fert}
\alias{module_gcameurope_L1322.Fert}
\title{module_gcameurope_L1322.Fert}
\usage{
module_gcameurope_L1322.Fert(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1322.Fert_Prod_MtNH3_R_F_Y_EUR}, \code{L1322.IO_R_Fert_F_Yh_EUR}, \code{L1322.in_EJ_R_indenergy_F_Yh_EUR}, \code{L1322.in_EJ_R_indfeed_F_Yh_EUR}, \code{L1322.Fert_NEcost_75USDkgNH3_F_EUR}. The corresponding file in the
original data system was \code{LB1322.Fert.R} (energy level1).
}
\description{
Compute ammonia fertilizer production and energy inputs by technology.
}
\details{
Fertilizer production and energy information was calculated using country-level fuel share and energy intensity data
to allocate generic fertilizer production data across technologies. Re-allocation of energy use between energy and feedstock
quantities for the industrial sector was done to avoid negative values at the country level, before aggregating
to the regional level. Also, a final cost table was calculated using DEU market fertilizer price data and H2A characteristics
of hydrogen production technologies.
}
\author{
AJS July 2017
}
