% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L1325.chemical.R
\name{module_gcameurope_L1325.chemical}
\alias{module_gcameurope_L1325.chemical}
\title{module_gcameurope_L1325.chemical}
\usage{
module_gcameurope_L1325.chemical(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1325.in_EJ_R_chemical_F_Y_EUR}, \code{L1325.in_EJ_R_indenergy_F_Yh_EUR}, \code{L1324.in_EJ_R_indfeed_F_Yh_EUR}. The corresponding file in the
}
\description{
Sets up input, output, and IO coefficients for chemical and subtracts input energy from industry energy use
}
\details{
This chunk generates input, Energy inputs are then subtracted from industrial energy use, feedstock and any resulting negative values
are dealt with by moving their accounting to the chemical sector.
}
\author{
Yang Liu Sep 2019
}
