% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L1326.aluminium.R
\name{module_gcameurope_L1326.aluminum}
\alias{module_gcameurope_L1326.aluminum}
\title{module_gcameurope_L1326.aluminum}
\usage{
module_gcameurope_L1326.aluminum(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1326.in_EJ_R_aluminum_Yh_EUR}, \code{L1326.IO_GJkg_R_aluminum_F_Yh_EUR},\code{L1326.out_Mt_R_aluminum_Yh_EUR}, \code{L1326.in_EJ_R_indenergy_F_Yh_EUR}. The corresponding file in the
}
\description{
Sets up input, output, and IO coefficients for aluminum and subtracts input energy from industry energy use
}
\details{
This chunk generates input, Energy inputs are then subtracted from industrial energy use, feedstock and any resulting negative values
are dealt with by moving their accounting to the aluminum sector.
}
\author{
Yang Liu Sep 2019
}
