% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L1327.paper.R
\name{module_gcameurope_L1327.paper}
\alias{module_gcameurope_L1327.paper}
\title{module_gcameurope_L1327.paper}
\usage{
module_gcameurope_L1327.paper(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1327.in_EJ_R_paper_Yh}, \code{L1327.in_EJ_R_paper_Yh_tech}, \code{L1327.IO_GJkg_R_paper_F_Yh_EUR},\code{L1327.out_Mt_R_paper_Yh_EUR}, \code{L1327.in_EJ_R_indenergy_F_Yh_EUR}, \code{L127.in_EJ_R_indchp_F_Yh_EUR}. The corresponding file in the
}
\description{
Sets up input, output, and IO coefficients for paper and subtracts input energy from industry energy use
}
\details{
This chunk generates input, Energy inputs are then subtracted from industrial energy use, feedstock and any resulting negative values
are dealt with by moving their accounting to the paper  sector.
}
\author{
MMC July 2022
}
