% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L1328.food_processing.R
\name{module_gcameurope_L1328.food_processing}
\alias{module_gcameurope_L1328.food_processing}
\title{module_gcameurope_L1328.food_processing}
\usage{
module_gcameurope_L1328.food_processing(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1328.in_EJ_R_food_F_Yh_EUR}, \code{L1328.out_Pcal_R_food_Yh_EUR},
\code{L1328.IO_EJPcal_R_food_F_Yh_EUR},\code{L1328.in_EJ_R_indenergy_F_Yh_EUR}, \code{L1328.in_EJ_R_indenergy_infilled_for_food_F_Yh_EUR},
\code{L1328.en_bal_frac_industry_food_inonspec_R_Yh_EUR}.
}
\description{
Sets up input, output, and IO coefficients for food processing and subtracts input energy from industry energy use
}
\details{
This chunk obtains input and output and coefficients for the food processing sector and subtracts input energy from industry
energy use. Also infills some historical food processing energy use data for regions where data are limited.
}
\author{
SAS Dec 2022
}
