% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L143.HDDCDD.R
\name{module_gcameurope_L143.HDDCDD}
\alias{module_gcameurope_L143.HDDCDD}
\title{module_gcameurope_L143.HDDCDD}
\usage{
module_gcameurope_L143.HDDCDD(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L143.HDDCDD_scen_R_Y_EUR}, \code{L143.HDDCDD_scen_RG3_Y_EUR}, \code{L143.HDDCDD_scen_ctry_Y_EUR}. The corresponding file in the
original data system was \code{LA143.HDDCDD.R} (energy level1).
}
\description{
Reads in country level heating and cooling degree day data and returns GCAM region degree days via population weighting
}
\details{
Population weights HDDCDD from country level to GCAM region
}
\author{
RH April 2017
}
