% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L154.transportation_UCD.R
\name{module_gcameurope_L154.transportation_UCD}
\alias{module_gcameurope_L154.transportation_UCD}
\title{module_gcameurope_L154.transportation_UCD}
\usage{
module_gcameurope_L154.transportation_UCD(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L154.in_EJ_R_trn_m_sz_tech_F_Yh_EUR},
 \code{L154.intensity_MJvkm_R_trn_m_sz_tech_F_Y_EUR}, \code{L154.loadfactor_R_trn_m_sz_tech_F_Y_EUR},
 \code{L154.cost_usdvkm_R_trn_m_sz_tech_F_Y_EUR}, \code{L154.speed_kmhr_R_trn_m_sz_tech_F_Y_EUR},
 The corresponding file in the original data system was \code{LA154.transportation_UCD.R} (energy level1).
}
\description{
Generates transportation energy and other data using UCD transportation database and Eurostat data.
}
\details{
L154.in_EJ_R_trn_m_sz_tech_F_Yh and L154.in_EJ_ctry_trn_m_sz_tech_F generated by aggregating Eurostat data UCD
transportation technologies, then scaling to transportation enduse data. Other outputs generated by aggregating UCD data
to Europe regions.
}
\author{
CRB January 2024
}
