% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L171.desalination.R
\name{module_gcameurope_L171.desalination}
\alias{module_gcameurope_L171.desalination}
\title{module_gcameurope_L171.desalination}
\usage{
module_gcameurope_L171.desalination(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs:  \code{L171.out_km3_R_desal_F_tech_Yh_EUR}, \code{L171.in_EJ_R_desal_F_Yh_EUR}.
}
\description{
Desalinated water production by region, basin, and technology
}
\details{
Generate estimates of desalinated water production for municipal and industrial purposes, by various cuts:
  by country, by GCAM region, by technology, by basin, and also estimate the desalinated water production at combined
  electric power and desalination facilities.
}
\author{
RH February 2024
}
