% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L173.EFW_manufacturing.R
\name{module_gcameurope_L173.EFW_manufacturing}
\alias{module_gcameurope_L173.EFW_manufacturing}
\title{module_gcameurope_L173.EFW_manufacturing}
\usage{
module_gcameurope_L173.EFW_manufacturing(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L173.in_EJ_R_indEFWtot_F_Yh_EUR}
}
\description{
Energy requirements for manufacturing water use
}
\details{
Generate estimates of energy consumption and energy input-output coefficients  for
  manufacturing water abstraction, treatment, and wastewater treatment, by GCAM region, fuel, and historical year.
}
\author{
RH February 2024
}
