% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L201.water_resources_constrained.R
\name{module_gcameurope_L201.water_resources_constrained}
\alias{module_gcameurope_L201.water_resources_constrained}
\title{module_gcameurope_L201.water_resources_constrained}
\usage{
module_gcameurope_L201.water_resources_constrained(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L201.RenewRsrcCurves_calib_EUR},
\code{201.GrdRenewRsrcMax_runoff}, \code{L201.DepRsrcCurves_ground_EUR}. The corresponding file in the
original data system was \code{L102.water_supply_unlimited.R} (water level1).
}
\description{
Constrained surface and groundwater.
}
\details{
Generates water resource input files for region + basin which includes runoff and groundwater.
}
\author{
RH Feb 2024
}
