% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L2011.ff_ALL_R_C_Y.R
\name{module_gcameurope_L2011.ff_ALL_R_C_Y}
\alias{module_gcameurope_L2011.ff_ALL_R_C_Y}
\title{module_gcameurope_L2011.ff_ALL_R_C_Y}
\usage{
module_gcameurope_L2011.ff_ALL_R_C_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2011.ff_ALL_EJ_R_C_Y_EUR} \code{L2011.ff_GrossTrade_EJ_R_C_Y_EUR}.
}
\description{
Calculate fossil fuel energy balances, by region / commodity / year and harmonize GCAM's data to Comtrade's.
Need to combine Eurostat and IEA data since the production values differ and therefore the calibrated net trade data differs.
}
\details{
This chunk combines fossil fuel production and consumption to calculate energy balances by
GCAM region, commodity and year. After calculating GCAM's (assumed) net trade
Comtrade's gross trade is adjusted to match.
}
\author{
RH Feb 2024
}
