% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L202.an_input.R
\name{module_gcameurope_L202.an_input}
\alias{module_gcameurope_L202.an_input}
\title{module_gcameurope_L202.an_input}
\usage{
module_gcameurope_L202.an_input(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L202.RenewRsrc_EUR}, \code{L202.RenewRsrcPrice_EUR}, \code{L202.maxSubResource_EUR},
\code{L202.RenewRsrcCurves_EUR}, \code{L202.ResTechShrwt_EUR}, \code{L202.UnlimitedRenewRsrcCurves_EUR}, \code{L202.UnlimitedRenewRsrcPrice_EUR},
\code{L202.Supplysector_in_EUR}, \code{L202.SubsectorAll_in_EUR}, \code{L202.StubTech_in_EUR}, \code{L202.StubTechInterp_in_EUR},
\code{L202.StubTechProd_in_EUR}, \code{L202.Supplysector_an_EUR}, \code{L202.SubsectorAll_an_EUR}, \code{L202.StubTechInterp_an_EUR}
\code{L202.StubTechProd_an_EUR}, \code{L202.StubTechCoef_an_EUR}, \code{L202.StubTechCost_an_EUR},
. The corresponding file in the
original data system was \code{L202.an_input.R} (aglu level2).
}
\description{
Produce a wide range of animal-related resource tables: production, import, resource curves.
}
\details{
This chunk produces 22 animal-related resource tables: production, import, resource curves.
}
\author{
CRB July 2024
}
