% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L203.water_td.R
\name{module_gcameurope_L203.water_td}
\alias{module_gcameurope_L203.water_td}
\title{module_gcameurope_L203.water_td}
\usage{
module_gcameurope_L203.water_td(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L203.Supplysector_watertd},
  \code{L203.SubsectorLogit_watertd}, \code{L203.SubsectorShrwtFllt_watertd}, \code{L203.SubsectorInterp_watertd},
  \code{L203.TechShrwt_watertd}, \code{L203.TechCoef_watertd}, \code{L203.Production_watertd_EUR}. The corresponding file
  in the original data system was \code{L203.water.mapping.R} (water level2).
}
\description{
Water distribution sectors.
}
\details{
Generates water mapping sector input files to group demands by sectors.
}
\author{
RH Feb 2024
}
