% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L211.ag_nonco2.R
\name{module_gcameurope_L211.ag_nonco2}
\alias{module_gcameurope_L211.ag_nonco2}
\title{module_gcameurope_L211.ag_nonco2}
\usage{
module_gcameurope_L211.ag_nonco2(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L211.AWBEmissions_EUR}, \code{L211.AGREmissions_EUR}, \code{L211.AnEmissions_EUR}, \code{L211.AnNH3Emissions_EUR}, \code{L211.AGRBio_EUR}, \code{L211.AWB_BCOC_EmissCoeff_EUR}, \code{L211.nonghg_max_reduction_EUR}, \code{L211.nonghg_steepness_EUR}. The corresponding file in the
original data system was \code{L211.ag_nonco2.R} (emissions level2).
}
\description{
Processes agriculture, agricultural waste burning, and animal emissions, adding regions and sectors/technologies.
By region and ag sector/technology, writes out N2O emissions coefficients for biomass and BC/OC emissions coefficients for ag waste burning.
Writes out non-GHG maximum emissions coefficient reduction and steepness,
a shape parameter that reduces emissions coefficient as function of per-capita GDP, by region and ag sector/technology.
}
\details{
Processes agriculture, agricultural waste burning, and animal emissions, adding regions and sectors/technologies.
By region and ag sector/technology, writes out N2O emissions coefficients for biomass and BC/OC emissions coefficients for ag waste burning.
Writes out non-GHG maximum emissions coefficient reduction and steepness,
a shape parameter that reduces emissions coefficient as function of per-capita GDP, by region and ag sector/technology.
}
\author{
RH July 2017
}
