% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L222.en_transformation.R
\name{module_gcameurope_L222.en_transformation}
\alias{module_gcameurope_L222.en_transformation}
\title{module_gcameurope_L222.en_transformation}
\usage{
module_gcameurope_L222.en_transformation(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L222.StubTechProd_gasproc_EUR}, \code{L222.StubTechProd_refining_EUR}, \code{L222.StubTechCoef_refining_EUR},
}
\description{
Prepare the calibrated outputs for energy transformation supplysectors, subsectors, and technologies.
}
\details{
Calibrated outputs for GCAM-Europe gas processing and oil refining as well as I:O coefficients are interpolated from historical values to base model years.
}
\author{
RH Feb 2024
}
