% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L2231.electricity.R
\name{module_gcameurope_L2231.electricity_grid}
\alias{module_gcameurope_L2231.electricity_grid}
\title{module_gcameurope_L2231.electricity}
\usage{
module_gcameurope_L2231.electricity_grid(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs:
}
\description{
Generates GCAM-Europe model inputs for electrcity sector by grid regions and states.
}
\details{
This chunk generates input files to create an annualized electricity generation sector for each region
and creates the demand for the region-level electricity sectors in the grid regions.
}
\author{
RC Oct 2017
}
