% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L2231.wind_update.R
\name{module_gcameurope_L2231.wind_update}
\alias{module_gcameurope_L2231.wind_update}
\title{module_gcameurope_L2231.wind_update}
\usage{
module_gcameurope_L2231.wind_update(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2231.SmthRenewRsrcCurves_onshore_wind_EUR}, \code{L2231.StubTechCapFactor_onshore_wind_EUR}, \code{L2231.SmthRenewRsrcTechChange_onshore_wind_EUR},
\code{L2231.StubTechCost_onshore_wind_EUR}. The corresponding file in the
original data system was \code{L2231.wind_update.R} (energy level2).
}
\description{
Updates region-specific onshore wind supply curves for europe with the new electricity segments
}
\details{
Updates region-specific onshore wind supply curves using improved global wind resource estimate from Eurek et al. (2016).
}
\author{
MB GI AJS March 2019
}
