% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L2232.electricity_trade.R
\name{module_gcameurope_L2232.electricity_trade}
\alias{module_gcameurope_L2232.electricity_trade}
\title{module_gcameurope_L2232.electricity_trade}
\usage{
module_gcameurope_L2232.electricity_trade(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs
}
\description{
Generate GCAM-Europe model inputs for electrcity trade sectors at the level of grid regions.
}
\details{
This chunk generates input files to create electricity trade and passthrough sectors for the grid regions,
and balances electricity supply and demand for each grid region.
}
\author{
RC Oct 2017
}
