% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L2233.electricity_water.R
\name{module_gcameurope_L2233.electricity_water}
\alias{module_gcameurope_L2233.electricity_water}
\title{module_gcameurope_L2233.electricity_water}
\usage{
module_gcameurope_L2233.electricity_water(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2233.StubTech_elecPassthru_EUR}, \code{L2233.StubTechProd_elecPassthru_EUR},
 \code{L2233.PassThroughSector_elec_cool_EUR},
\code{L2233.Supplysector_elec_cool_EUR}, \code{L2233.ElecReserve_elec_cool_EUR},
\code{L2233.SubsectorShrwtFllt_elec_cool_EUR}, \code{L2233.SubsectorLogit_elec_cool_EUR},
 \code{L2233.StubTech_elec_cool_EUR}, \code{L2233.StubTechEff_elec_cool_EUR},
\code{L2233.StubTechProd_elec_cool_EUR}, \code{L2233.StubTechCapFactor_elec_cool_EUR}, \code{L2233.StubTechFixOut_hydro_EUR},
  \code{L2233.StubTechShrwt_elec_cool_EUR}
}
\description{
Generates GCAM model inputs for electricity sector with cooling system types disaggregated.
}
\details{
Disaggregates electricity sector for all cooling system types.
}
\author{
RH Feb 2024
}
