% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L2234.elec_segments.R
\name{module_gcameurope_L2234.elec_segments}
\alias{module_gcameurope_L2234.elec_segments}
\title{module_gcameurope_L2234.elec_segments}
\usage{
module_gcameurope_L2234.elec_segments(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs
}
\description{
Generates GCAM-Europe model inputs for multiple load segments electricity sector by country
}
\details{
This chunk generates input files to create an electricity generation sector with multiple load segments
for each state and creates the demand for the state-level electricity sectors in the grid regions.
}
\author{
MTB Aug 2018
}
