% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L2235.electricity_water_segments.R
\name{module_gcameurope_L2235.elec_segments_water}
\alias{module_gcameurope_L2235.elec_segments_water}
\title{module_gcameurope_L2235.elec_segments_water}
\usage{
module_gcameurope_L2235.elec_segments_water(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs:
}
\description{
Generates GCAM-Europe model inputs for multiple load segments electricity sector and cooling systems by state.
}
\details{
This chunk generates input files to create an electricity generation sector with multiple load segments
for each country and creates the demand by load, fuel, power plant, and cooling system type.
}
\author{
NTG May 2020
}
