% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L2236.elecS_emissions.R
\name{module_gcameurope_L2236.elecS_emissions}
\alias{module_gcameurope_L2236.elecS_emissions}
\title{module_gcameurope_L2236.elecS_emissions}
\usage{
module_gcameurope_L2236.elecS_emissions(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs
}
\description{
European electricity non CO2 GHG emission and future emission coefficients by sector, fuel and cooling techs
}
\details{
Write electricity emission coefficients to multiple load segments and cooling technology
}
\author{
KD August 2018 / YO March 2022 / RH Sept 2024
}
