% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L226.en_distribution.R
\name{module_gcameurope_L226.en_distribution}
\alias{module_gcameurope_L226.en_distribution}
\title{module_gcameurope_L226.en_distribution}
\usage{
module_gcameurope_L226.en_distribution(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L226.SectorLogitTables[[ curr_table ]]$data}, \code{L226.Supplysector_en}, \code{L226.SubsectorLogitTables[[ curr_table ]]$data}, \code{L226.SubsectorLogit_en}, \code{L226.SubsectorShrwt_en}, \code{L226.SubsectorShrwtFllt_en}, \code{L226.SubsectorInterp_en}, \code{L226.SubsectorInterpTo_en}, \code{L226.StubTech_en}, \code{L226.StubTechCoef_elecownuse_EUR}, \code{L226.StubTechCoef_electd_EUR}, \code{L226.StubTechCoef_gaspipe_EUR}.
}
\description{
Generate the level 2 data tables for the energy distribution sector,
including capital costs, shareweights, logits, and interpolations as well as energy use coefficients for electricity and gas pipeline
}
\details{
Prepares Level 2 data on energy distribution sector for the generation of en_distribution.xml.
Creates global technology database info--cost, shareweight, logit, efficiencies, and interpolations--and regional values where applicable for electricity net ownuse, gas pipelines, and transmission and distribution.
}
\author{
RH Feb 2024
}
