% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L2322.Fert.R
\name{module_gcameurope_L2322.Fert}
\alias{module_gcameurope_L2322.Fert}
\title{module_gcameurope_L2322.Fert}
\usage{
module_gcameurope_L2322.Fert(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs:  \code{L2322.Supplysector_Fert_EUR}, \code{L2322.FinalEnergyKeyword_Fert_EUR}, \code{L2322.SubsectorLogit_Fert_EUR},
\code{L2322.SubsectorShrwtFllt_Fert_EUR}, \code{L2322.SubsectorInterp_Fert_EUR}, \code{L2322.StubTech_Fert_EUR},
\code{L2322.StubTechProd_FertProd_EUR}, \code{L2322.StubTechCoef_Fert_EUR}. The corresponding file in the
original data system was \code{L2322.Fert.R} (energy level2).
}
\description{
Provide supply sector information/keywords, subsector shareweights, global technology lifetime,
energy inputs and coefficients, global fertilizer manufacturing technologies, etc. for the fertilizer sector.
}
\details{
This chunk provides supply sector information/keywords, subsector shareweights, global technology lifetime,
energy inputs and coefficients, global fertilizer manufacturing technologies, etc. for the fertilizer sector.
}
\author{
LF September 2017
}
