% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L238.iron_steel_trade.R
\name{module_gcameurope_L238.iron_steel_trade}
\alias{module_gcameurope_L238.iron_steel_trade}
\title{module_gcameurope_L238.iron_steel_trade}
\usage{
module_gcameurope_L238.iron_steel_trade(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L238.Supplysector_tra},
  \code{L238.SectorUseTrialMarket_tra}, \code{L238.SubsectorAll_tra}, \code{L238.TechShrwt_tra},
  \code{L238.TechCost_tra}, \code{L238.TechCoef_tra}, \code{L238.Production_tra}, \code{L238.Supplysector_reg},
  \code{L238.SubsectorAll_reg}, \code{L238.TechShrwt_reg}, \code{L238.TechCoef_reg}, \code{L238.Production_reg_imp},
  \code{L238.Production_reg_dom}.
}
\description{
Model input for regional and (globally) traded iron and steel
}
\author{
Siddarth Durga July 2022
}
