% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L239.ff_trade.R
\name{module_gcameurope_L239.ff_trade}
\alias{module_gcameurope_L239.ff_trade}
\title{module_gcameurope_L239.ff_trade}
\usage{
module_gcameurope_L239.ff_trade(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L239.PrimaryConsKeyword_en_EUR}, \code{L239.Supplysector_tra_EUR},
  \code{L239.SectorUseTrialMarket_tra_EUR}, \code{L239.SubsectorAll_tra_EUR}, \code{L239.TechShrwt_tra_EUR},
  \code{L239.TechCost_tra_EUR}, \code{L239.TechCoef_tra_EUR}, \code{L239.Production_tra_EUR}, \code{L239.Supplysector_reg_EUR},
  \code{L239.SubsectorAll_reg_EUR}, \code{L239.TechShrwt_reg_EUR}, \code{L239.TechCoef_reg_EUR}, \code{L239.Production_reg_imp_EUR},
  \code{L239.Production_reg_dom_EUR}, \code{L239.Consumption_intraregional_EUR}, \code{L239.CarbonCoef_EUR}.
}
\description{
Model input for regionally and globally traded fossil fuel commodities
}
\author{
RH February 2024
}
