% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L2391.gas_trade_flows.R
\name{module_gcameurope_L2391.gas_trade_flows}
\alias{module_gcameurope_L2391.gas_trade_flows}
\title{module_gcameurope_L2391.gas_trade_flows}
\usage{
module_gcameurope_L2391.gas_trade_flows(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
 \code{command} is "MAKE") all the generated outputs: \code{L2391.NG_export_calOutput_LNG_EUR},
 \code{L2391.NG_export_calOutput_pipeline_EUR}, \code{L2391.NG_import_calOutput_LNG_EUR},
 \code{L2391.NG_import_calOutput_pipeline_EUR},  \code{L2391.NG_import_calOutput_statdiff_EUR},
 \code{L2391.NG_export_calOutput_statdiff_EUR}.
}
\description{
Model input for natural gas trade by LNG and regional pipeline networks.
}
\author{
RH Feb 2024
}
