% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L240.ag_trade.R
\name{module_gcameurope_L240.ag_trade}
\alias{module_gcameurope_L240.ag_trade}
\title{module_gcameurope_L240.ag_trade}
\usage{
module_gcameurope_L240.ag_trade(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs:
}
\description{
Adjust ag trade for European Single Market
}
\details{
Build datasets for ssp4 agricultural trade: food and nonfood trade coefficients, feed trade
coefficients, restricted agricultural trade, and trade regions.
}
\author{
RH November 2024
}
