% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L241.fgas.R
\name{module_gcameurope_L241.fgas}
\alias{module_gcameurope_L241.fgas}
\title{module_gcameurope_L241.fgas}
\usage{
module_gcameurope_L241.fgas(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L241.hfc_all_EUR}, \code{L241.pfc_all_EUR}, \code{L241.hfc_future_EUR}, \code{L241.fgas_all_units_EUR}. The corresponding file in the
original data system was \code{L241.fgas.R} (emissions level2).
}
\description{
Format fgases emission inputs for GCAM and estimates future emission factors for f gases for the SSP scenarios.
}
\details{
Formats hfc and pfc gas emissions for input. Calculates future emission factors for hfc gases based on 2010 region emissions and Germany emission factors and emission factors from Guus Velders (http://www.sciencedirect.com/science/article/pii/S135223101530488X) for the  SSP scenarios.
}
\author{
KD July 2017
}
