% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L243.bio_trade_input.R
\name{module_gcameurope_L243.bio_trade_input}
\alias{module_gcameurope_L243.bio_trade_input}
\title{module_gcameurope_L243.bio_trade_input}
\usage{
module_gcameurope_L243.bio_trade_input(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs:
}
\description{
Add European Single Market to bioenergy trade.
}
\details{
This chunk sets up a structure for regionally differentiated bioenergy trade. Each
region consumes a blend of domestic and international bioenergy. Regions can supply to either
the domestic or international market. Share weights in the default case depend on the amount of
cropland a region has, with the largest region having a share weight of 1. Share weights in SSP3
and SSP4 are adjusted to reflect the trade frictions in the SSP storylines.
}
\author{
RH October 2024
}
