% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L244.building_det.R
\name{module_gcameurope_L244.building_det}
\alias{module_gcameurope_L244.building_det}
\title{module_gcameurope_L244.building_det}
\usage{
module_gcameurope_L244.building_det(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L244.SubregionalShares_EUR},
\code{L244.PriceExp_IntGains_EUR}, \code{L244.Floorspace_EUR}, \code{L244.DemandFunction_serv_EUR},
\code{L244.DemandFunction_flsp_EUR}, \code{L244.Satiation_flsp_EUR}, \code{L244.SatiationAdder_EUR}, \code{L244.ThermalBaseService_EUR}, \code{L244.GenericBaseService_EUR},
\code{L244.ThermalServiceSatiation_EUR}, \code{L244.GenericServiceSatiation_EUR}, \code{L244.Intgains_scalar_EUR}, \code{L244.ShellConductance_bld_EUR},
\code{L244.Supplysector_bld_EUR}, \code{L244.FinalEnergyKeyword_bld_EUR}, \code{L244.SubsectorShrwt_bld_EUR}, \code{L244.SubsectorShrwtFllt_bld_EUR}, \code{L244.SubsectorInterp_bld_EUR},
\code{L244.SubsectorInterpTo_bld_EUR}, \code{L244.SubsectorLogit_bld_EUR}, \code{L244.FuelPrefElast_bld_EUR}, \code{L244.StubTech_bld_EUR}, \code{L244.StubTechEff_bld_EUR},
\code{L244.StubTechCalInput_bld_EUR}, \code{L244.StubTechIntGainOutputRatio_EUR}, \code{L244.GlobalTechShrwt_bld_EUR}, \code{L244.PrimaryRenewKeyword_bld_EUR}, \code{L244.GlobalTechCost_bld_EUR}, \code{L244.GlobalTechSCurve_bld_EUR},
\code{L244.DeleteGenericService_EUR}, \code{L244.DeleteThermalService_EUR} and \code{L244.GompFnParam_EUR}, \code{L244.Satiation_impedance_EUR},
\code{L244.GenericServiceImpedance_EUR}, \code{L244.GenericServiceAdder_EUR}, \code{L244.ThermalServiceImpedance_EUR}, \code{L244.ThermalServiceAdder_EUR}
\code{L244.GenericTradFuelParams_EUR}, \code{L244.ThermalTradFuelParams_EUR}, \code{L244.GenericServiceCoef_EUR},\code{L244.ThermalServiceCoef_EUR},
\code{L244.GenericShares_EUR}, \code{L244.ThermalShares_EUR},\code{L244.GenericServicePrice_EUR}, \code{L244.ThermalServicePrice_EUR},
\code{L244.GenericBaseDens_EUR}, \code{L244.ThermalBaseDens_EUR}, \code{L244.HDDCDD_constdd_no_GCM_EUR}
\code{L244.HDDCDD_A2_CCSM3x_EUR}, \code{L244.HDDCDD_A2_HadCM3_EUR}, \code{L244.HDDCDD_B1_CCSM3x_EUR}, \code{L244.HDDCDD_B1_HadCM3_EUR}
The corresponding file in the original data system was \code{L244.building_det.R} (energy level2).
}
\description{
Creates level2 data for the building sector.
}
\details{
Creates level2 data for the building sector.
}
\author{
RLH September 2017
}
