% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L251.en_ssp_nonco2.R
\name{module_gcameurope_L251.en_ssp_nonco2}
\alias{module_gcameurope_L251.en_ssp_nonco2}
\title{module_gcameurope_L251.en_ssp_nonco2}
\usage{
module_gcameurope_L251.en_ssp_nonco2(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L251.ctrl.delete_EUR}, \code{L251.ssp15_ef_EUR}, \code{L251.ssp2_ef_EUR}, \code{L251.ssp2_ef_residTradBio_EUR}, \code{L251.ssp34_ef_EUR}, \code{L251.ssp15_ef_vin_EUR}, \code{L251.ssp2_ef_vin_EUR}, \code{L251.ssp34_ef_vin_EUR}. The corresponding file in the
original data system was \code{L251.en_ssp_nonco2.R} (emissions level2).
}
\description{
Produce regional non-CO2 emissions coefficient data for SSPs 1/5, 2, and 3/4 as well as a GDP control.
}
\details{
This section takes in the non-CO2 emissions factors for SSP 1/5, 2, and 3/4 across sectors.
First, create data that spans the years 2010-2100 in five year increments by interpolation of input data.
Next, add emissions controls for future years of vintaged technologies for SSP emission factors.
Then, add columns that have regional SO2 emission species.
A GDP control of regional non-CO2 emissions in all regions is also created.
}
\author{
CRB June 2024
}
