% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L252.MACC.R
\name{module_gcameurope_L252.MACC}
\alias{module_gcameurope_L252.MACC}
\title{module_gcameurope_L252.MACC}
\usage{
module_gcameurope_L252.MACC(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L252.ResMAC_fos_EUR}, \code{L252.AgMAC_EUR}, \code{L252.MAC_an_EUR}, \code{L252.MAC_prc_EUR},
\code{L252.MAC_higwp_EUR}, \code{L252.ResMAC_fos_tc_average_EUR}, \code{L252.AgMAC_tc_average_EUR}, \code{L252.MAC_an_tc_average_EUR},
\code{L252.MAC_prc_tc_average_EUR}, \code{L252.MAC_higwp_tc_average_EUR}. The corresponding file in the
original data system was \code{L252.MACC.R} (emissions level2).
}
\description{
Creates marginal abatement cost curves "MACC", for fossil resources, agriculture, animals, and processing.
}
\details{
Creates marginal abatement cost curves "MACC", for fossil resources, agriculture, animals, and processing.
}
\author{
RH August 2017
}
