% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L270.EFW_input_coefs.R
\name{module_gcameurope_L270.EFW_input_coefs}
\alias{module_gcameurope_L270.EFW_input_coefs}
\title{module_gcameurope_L270.EFW_input_coefs}
\usage{
module_gcameurope_L270.EFW_input_coefs(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L270.TechCoef_EFW_EUR}. This did not exist in the original
  data system.
}
\description{
Energy-for-water input names and coefficients to water T&D technologies
}
\details{
Describe in detail what this chunk does.
}
\author{
GPK January 2019
}
