% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L271.desalination.R
\name{module_gcameurope_L271.desalination}
\alias{module_gcameurope_L271.desalination}
\title{module_gcameurope_L271.desalination}
\usage{
module_gcameurope_L271.desalination(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L271.Supplysector_desal},
  \code{L271.FinalEnergyKeyword_desal}, \code{L271.SubsectorLogit_desal}, \code{L271.SubsectorShrwtFllt_desal},
  \code{L271.SubsectorInterpTo_desal}, \code{L271.StubTech_desal}, \code{L271.StubTechProd_desal_EUR}.
}
\description{
Generate sector and technologies for desalination
}
\details{
This chunk translates input assumptions from CSV to XML, and calibrates the production by technology in each
  region and model base year
}
\author{
RH Feb 2024
}
