% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L272.EFW_irrigation.R
\name{module_gcameurope_L272.EFW_irrigation}
\alias{module_gcameurope_L272.EFW_irrigation}
\title{module_gcameurope_L272.EFW_irrigation}
\usage{
module_gcameurope_L272.EFW_irrigation(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L272.Supplysector_irr_EUR},
  \code{L272.FinalEnergyKeyword_irr_EUR}, \code{L272.SubsectorLogit_irr_EUR}, \code{L272.SubsectorShrwtFllt_irr_EUR},
  \code{L272.SubsectorInterpTo_irr_EUR}, \code{L272.StubTech_irr_EUR}, \code{L272.StubTechCoef_irr_EUR}.
}
\description{
Generate irrigation energy-for-water input file
}
\details{
This chunk translates input assumptions from CSV to XML, and calibrates the input-output coefficients by
  technology in each region and model base year
}
\author{
RH Feb 2024
}
