% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L273.EFW_manufacturing.R
\name{module_gcameurope_L273.EFW_manufacturing}
\alias{module_gcameurope_L273.EFW_manufacturing}
\title{module_gcameurope_L273.EFW_manufacturing}
\usage{
module_gcameurope_L273.EFW_manufacturing(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L273.Supplysector_ind},
  \code{L273.FinalEnergyKeyword_ind}, \code{L273.SubsectorLogit_ind}, \code{L273.SubsectorShrwtFllt_ind},
  \code{L273.SubsectorInterpTo_ind}, \code{L273.StubTech_ind}, \code{L273.GlobalTechCoef_ind},
  \code{L273.GlobalTechShrwt_ind}, \code{L273.StubTechCoef_ind_EUR}. This did not
  exist in the original data system.
}
\description{
Generate manufacturing energy-for-water input file
}
\details{
This chunk translates input assumptions from CSV to XML, and calibrates the input-output coefficients by
  technology in each region and model base year
}
\author{
GPK January 2019
}
