% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L274.EFW_municipal.R
\name{module_gcameurope_L274.EFW_municipal}
\alias{module_gcameurope_L274.EFW_municipal}
\title{module_gcameurope_L274.EFW_municipal}
\usage{
module_gcameurope_L274.EFW_municipal(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L274.Supplysector_muni},
  \code{L274.FinalEnergyKeyword_muni}, \code{L274.SubsectorLogit_muni}, \code{L274.SubsectorShrwtFllt_muni},
  \code{L274.SubsectorInterpTo_muni}, \code{L274.StubTech_muni}, \code{L274.StubTechCoef_muni_EUR}.
}
\description{
Generate municipal water sector energy-for-water input file
}
\details{
This chunk translates input assumptions from CSV to XML, and calibrates the input-output coefficients by
  technology in each region and model base year
}
\author{
RH Feb 2024
}
