% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_xml_water_demand_industry.R
\name{module_gcameurope_water_demand_industry_xml}
\alias{module_gcameurope_water_demand_industry_xml}
\title{module_gcameurope_water_demand_industry_xml}
\usage{
module_gcameurope_water_demand_industry_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{water_demand_industry_EUR.xml}.(water XML).
}
\description{
Construct XML data structure for \code{water_demand_industry_EUR.xml}.
}
