% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/zchunk_batch_agg_elect_global_solar_resource_india_xml.R
\name{module_gcamindia_batch_agg_elect_global_solar_resource_xml}
\alias{module_gcamindia_batch_agg_elect_global_solar_resource_xml}
\title{module_gcamindia_batch_agg_elect_global_solar_resource_xml}
\usage{
module_gcamindia_batch_agg_elect_global_solar_resource_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{agg_elect_global_solar_resource_india.xml}. The corresponding file in the
(gcamindia XML).
}
\description{
Construct XML data structure for \code{comm_india.xml}.
}
\author{
PNK Nov20
}
