% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_electricity_india_xml.R
\name{module_gcamindia_batch_electricity_xml}
\alias{module_gcamindia_batch_electricity_xml}
\title{module_gcamindia_batch_electricity_xml}
\usage{
module_gcamindia_batch_electricity_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{electricity_india.xml}. The corresponding file in the
original data system was \code{batch_electricity_india_xml.R} (gcamindia XML).
}
\description{
Construct XML data structure for \code{electricity_india.xml}.
}
\author{
VC Nov20
}
