% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_HDDCDD_constdds_india_xml.R
\name{module_gcamindia_batch_HDDCDD_constdds_xml}
\alias{module_gcamindia_batch_HDDCDD_constdds_xml}
\title{module_gcamindia_batch_HDDCDD_constdds_xml}
\usage{
module_gcamindia_batch_HDDCDD_constdds_xml(command, ...)
}
\description{
Construct XML data structure for \code{HDDCDD_constdds_india.xml}.
}
\author{
PNK Nov20
}
