% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_socioeconomics_india_xml.R
\name{module_gcamindia_batch_socioeconomics_xml}
\alias{module_gcamindia_batch_socioeconomics_xml}
\title{module_gcamindia_batch_socioeconomics_xml}
\usage{
module_gcamindia_batch_socioeconomics_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{socioeconomics_india.xml}.
}
\description{
Construct XML data structure for \code{socioeconomics_india.xml}.
}
\author{
AM Nov20
}
