% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_transportation_INDIA_xml.R
\name{module_gcamindia_batch_transportation_India_xml}
\alias{module_gcamindia_batch_transportation_India_xml}
\title{module_gcamindia_batch_transportation_India_xml}
\usage{
module_gcamindia_batch_transportation_India_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{transportation_India.xml}. The corresponding file in the
original data system was \code{batch_transportation_India_xml.R} (gcamIndia XML).
}
\description{
Construct XML data structure for \code{transportation_India.xml}.
}
