% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/zchunk_L210.agg_elec_global_solar_resource_india.R
\name{module_gcamindia_L210.agg_global_solar_resource_india}
\alias{module_gcamindia_L210.agg_global_solar_resource_india}
\title{module_gcamindia_L210.agg_global_solar_resource_india}
\usage{
module_gcamindia_L210.agg_global_solar_resource_india(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L210.india_state_UnlimitRsrc_agg_india}, \code{L210.india_state_UnlimitRsrcPrice_agg_india}
original data system was \code{L223.building_agg.R} (energy level2).
}
\description{
Calculate supply sector, subsector, and technology information for the agg_elect_td_ind sector
}
\details{
Calculate shareweights, cost, price elasticity, calibrated, and other data for the building sector
}
\author{
PNK Nov20
}
