% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L210.resources_india.R
\name{module_gcamindia_L210.resources}
\alias{module_gcamindia_L210.resources}
\title{module_gcamindia_L210.resources}
\usage{
module_gcamindia_L210.resources(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L210.DeleteRenewRsrc_indiarsrc}, \code{L210.DeleteUnlimitRsrc_indiarsrc}, \code{L210.india_state_RenewRsrc},
\code{L210.india_state_UnlimitRsrc}, \code{L210.india_state_UnlimitRsrcPrice}, \code{L210.india_state_SmthRenewRsrcTechChange},
\code{L210.india_state_SmthRenewRsrcCurves_wind}, \code{L210.india_state_GrdRenewRsrcCurves_geo}, \code{L210.india_state_GrdRenewRsrcMax_geo},
\code{L210.india_state_GrdRenewRsrcCurves_tradbio},\code{L210.india_state_GrdRenewRsrcMax_tradbio},\code{L210.india_state_SmthRenewRsrcCurvesGdpElast_roofPV}.
The corresponding file in the
original data system was \code{L210.resources_india.R} (gcam-india level2).
}
\description{
GCAM-india resource market information, prices, TechChange parameters, and supply curves.
}
\details{
GCAM-india resource market information, prices, TechChange parameters, and supply curves.
}
\author{
SJ NOV21
}
