% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L223.agg_elec_td_trn_india.R
\name{module_gcamindia_L223.agg_elect_td_trn}
\alias{module_gcamindia_L223.agg_elect_td_trn}
\title{module_gcamindia_L223.agg_elect_td_trn}
\usage{
module_gcamindia_L223.agg_elect_td_trn(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L223.india_state_Supplysector_agg_elect_td_trn}, \code{L223.india_state_SubsectorLogit_agg_elect_td_trn}, \code{L223.india_state_SubsectorShrwtFllt_agg_elect_td_trn},
\code{L223.india_state_TechShrwt_agg_elect_td_trn}, \code{L223.india_state_TechCoeff_agg_elect_td_trn},
original data system was \code{L223.building_agg.R} (energy level2).
}
\description{
Calculate supply sector, subsector, and technology information for the agg_elect_td_trn sector
}
\details{
Calculate shareweights, cost, price elasticity, calibrated, and other data for the building sector
}
\author{
AM Nov20
}
