% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L223.electricity_india.R
\name{module_gcamindia_L223.electricity}
\alias{module_gcamindia_L223.electricity}
\title{module_gcamindia_L223.electricity}
\usage{
module_gcamindia_L223.electricity(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\description{
Generates GCAM-india model inputs for electrcity sector by grid regions and states.
}
\details{
This chunk sets up the india energy electricity technology databases as well as writing out assumptions to all states/sectors/markets for shareweights and logits.
Calibrated outputs and I:O coefficients are updated from global values produced by \code{\link{module_gcamindia_L223.electricity}}.
}
\author{
VC Nov20
}
