% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L226.en_distribution_india.R
\name{module_gcamindia_L226.en_distribution}
\alias{module_gcamindia_L226.en_distribution}
\title{module_gcamindia_L226.en_distribution}
\usage{
module_gcamindia_L226.en_distribution(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs:
\itemize{
\item{\code{L226.india_state_DeleteSupplysector_elec}: Removing the electricity T&D sectors of the India region.}
\item{\code{L226.india_state_StubTechCoef_electd}: Stub technology coefficients elec T&D when using national elec markets. State elect_td sectors are treated as stub technologies.}
\item{\code{L226.india_state_TechShrwt_electd}: tech share weights for elec T&D when using regional electricity markets. The elect_td sectors can not use the global tech database as their input is different.}
\item{\code{L226.india_state_TechCost_electd}: Tech costs for elec T&D when using regional electricity markets.}
\item{\code{L226.india_state_TechCoef_electd}: Tech coeff for elec T&D when using regional electricity markets.}
\item{\code{L226.india_state_Supplysector_en}: Supply sector information for energy handling and delivery sectors for India grid regions. Currently using FERC regions as a proxy for regional energy markets.}
\item{\code{L226.india_state_SubsectorShrwtFllt_en}: Subsector shareweights of energy handling and delivery.}
\item{\code{L226.india_state_SubsectorLogit_en}: Logit info for energy subsectors. There is only one tech per subsector so the logit choice does not matter.}
\item{\code{L226.india_state_TechShrwt_en}: Technology shareweights of energy handling and delivery. Can't use stub technologies because these would inherit the wrong energy-inputs.}
\item{\code{L226.india_state_TechCoef_en}: Technology coefficients and market names of energy handling and delivery.}
\item{\code{L226.india_state_TechCost_en}: Regional price adjustments/cost adders for India energy.}
\item{\code{L226.india_state_Ccoef}: Carbon coef for India cost adder sectors.}
\item{\code{L226.india_state_Supplysector_electd}: India supply sector input, output, and logit info for elec T&D by state.}
\item{\code{L226.india_state_SubsectorLogit_electd}: India subsector logit info for elec T&D by grid_region.}
\item{\code{L226.india_state_SubsectorShrwtFllt_electd}: India subsector shareweight fillout for elec T&D by state.}
\item{\code{L226.india_state_SubsectorInterp_electd}: India interpolation info for elec T&D by state.}
}
The corresponding file in the original data system was \code{L226.india_state_en_distribution.R} (gcam-india level2).
}
\description{
Create a variety of energy and electricity outputs for India at the state and/or grid_region level.
}
\author{
VC Nov20
}
