% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L232.industry_india.R
\name{module_gcamindia_L232.industry}
\alias{module_gcamindia_L232.industry}
\title{module_gcamindia_L232.industry}
\usage{
module_gcamindia_L232.industry(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L232.india_state_DeleteSupplysector_ind}, \code{L2321.india_state_DeleteSupplysector_cement}, \code{L232.india_state_DeleteFinalDemand_ind},
\code{L2321.india_state_DeleteFinalDemand_cement}, \code{L232.india_state_StubTechCalInput_indenergy}, \code{L232.india_state_StubTechCalInput_indfeed}, \code{L232.india_state_StubTechProd_industry},
\code{L232.india_state_StubTechCoef_industry}, \code{L232.india_state_StubTechMarket_ind}, \code{L232.india_state_StubTechSecMarket_ind},
\code{L232.india_state_BaseService_ind}, \code{L232.india_state_Supplysector_ind}, \code{L232.india_state_FinalEnergyKeyword_ind},
\code{L232.india_state_SubsectorLogit_ind}, \code{L232.india_state_SubsectorShrwtFllt_ind}, \code{L232.india_state_SubsectorInterp_ind},
\code{L232.india_state_StubTech_ind}, \code{L232.india_state_StubTechInterp_ind}, \code{L232.india_state_PerCapitaBased_ind},
\code{L232.india_state_PriceElasticity_ind}, \code{L232.india_state_IncomeElasticity_ind_gcam3}.
The corresponding file in the original data system was \code{L232.industry_india.R} (gcam-india level2).
}
\description{
Prepare level 2 industry sector files for india.
}
\details{
Prepare level 2 industry sector files for india.
}
\author{
AM Nov20
}
