% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2321.cement_india.R
\name{module_gcamindia_L2321.cement_India}
\alias{module_gcamindia_L2321.cement_India}
\title{module_gcamindia_L2321.cement_India}
\usage{
module_gcamindia_L2321.cement_India(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2321.DeleteSupplysector_Indiacement}, \code{L2321.DeleteFinalDemand_Indiacement}, \code{L2321.StubTechProd_cement_India},
\code{L2321.StubTechCoef_cement_India}, \code{L2321.StubTechCalInput_cement_heat_India}, \code{L2321.StubTechMarket_cement_India}, \code{L2321.BaseService_cement_India},
\code{L2321.Supplysector_cement_India}, \code{L2321.FinalEnergyKeyword_cement_India}, \code{L2321.SubsectorLogit_cement_India},
\code{L2321.SubsectorShrwtFllt_cement_India}, \code{L2321.StubTech_cement_India}, \code{L2321.PerCapitaBased_cement_India},
 \code{L2321.PriceElasticity_cement_India}, \code{L2321.IncomeElasticity_cement_gcam3_India} .
 The corresponding file in the original data system was \code{L2321.cement_India.R} (gcam-india level2).
}
\description{
Make the logit and input tables for the cement sector in gcam-india
}
\details{
Make the logit and input tables for the cement sector in gcam-usa
}
\author{
AM Dec20
}
