% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L242.agg_tradbio_india.R
\name{module_gcamindia_L242.agg_tradbio}
\alias{module_gcamindia_L242.agg_tradbio}
\title{module_gcamindia_L242.agg_tradbio}
\usage{
module_gcamindia_L242.agg_tradbio(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L242.india_state_Supplysector_agg_tradbio}, \code{L242.india_state_SubsectorLogit_agg_tradbio}, \code{L242.india_state_SubsectorShrwtFllt_agg_tradbio},
\code{L242.india_state_TechShrwt_agg_tradbio}, \code{L242.india_state_TechEff_agg_tradbio},
original data system was \code{L223.building_agg.R} (energy level2).
}
\description{
Calculate supply sector, subsector, and technology information for the agg_elect_td_ind sector
}
\details{
Calculate shareweights, cost, price elasticity, calibrated, and other data for the building sector
}
\author{
PNK Nov20
}
