% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L242.building_agg_comm_india.R
\name{module_gcamindia_L242.building_agg_comm}
\alias{module_gcamindia_L242.building_agg_comm}
\title{module_gcamindia_L242.building_agg_comm}
\usage{
module_gcamindia_L242.building_agg_comm(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L242.india_state_Supplysector_comm}, \code{L242.india_state_FinalEnergyKeyword_comm}, \code{L242.india_state_SubsectorLogit_comm},
\code{L242.india_state_SubsectorShrwtFllt_comm}, \code{L242.india_state_SubsectorInterp_comm}, \code{L242.india_state_StubTech_comm},
\code{L242.india_state_GlobalTechInterp_comm}, \code{L242.india_state_GlobalTechShrwt_comm}, \code{L242.india_state_GlobalTechEff_comm},
\code{L242.india_state_GlobalTechCost_comm}, \code{L242.india_state_StubTechCalInput_comm}, \code{L242.india_state_StubTechMarket_comm}, \code{L242.india_state_FuelPrefElast_comm},
\code{L242.india_state_PerCapitaBased_comm}, \code{L242.india_state_PriceElasticity_comm}, \code{L242.india_state_IncomeElasticity_comm}, \code{L242.india_state_BaseService_comm}. The corresponding file in the
original data system was \code{L242.building_agg.R} (energy level2).
}
\description{
Calculate supply sector, subsector, and technology information for the comm sector
}
\details{
Calculate shareweights, cost, price elasticity, calibrated, and other data for the building sector
}
\author{
AJS August 2017
}
