% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L242.building_agg_resid_india.R
\name{module_gcamindia_L242.building_agg_resid}
\alias{module_gcamindia_L242.building_agg_resid}
\title{module_gcamindia_L242.building_agg_resid}
\usage{
module_gcamindia_L242.building_agg_resid(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L242.india_state_DeleteSupplysector_bld}, \code{L242.india_state_DeleteFinalDemand_bld}, \code{L242.india_state_Supplysector_resid}, \code{L242.india_state_FinalEnergyKeyword_resid}, \code{L242.india_state_SubsectorLogit_resid},
\code{L242.india_state_SubsectorShrwtFllt_resid}, \code{L242.india_state_SubsectorInterp_resid}, \code{L242.india_state_StubTech_resid},
\code{L242.india_state_GlobalTechInterp_resid}, \code{L242.india_state_GlobalTechShrwt_resid}, \code{L242.india_state_GlobalTechEff_resid},
\code{L242.india_state_GlobalTechCost_resid}, \code{L242.india_state_StubTechCalInput_resid}, \code{L242.india_state_StubTechMarket_resid}, \code{L242.india_state_FuelPrefElast_resid},
\code{L242.india_state_PerCapitaBased_resid}, \code{L242.india_state_PriceElasticity_resid}, \code{L242.india_state_IncomeElasticity_resid}, \code{L242.india_state_BaseService_resid}. The corresponding file in the
original data system was \code{L242.building_agg.R} (energy level2).
}
\description{
Calculate supply sector, subsector, and technology information for the resid sector
}
\details{
Calculate shareweights, cost, price elasticity, calibrated, and other data for the building sector
}
\author{
AJS August 2017
}
