% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L244.building_INDIA.R
\name{module_gcamindia_L244.building_INDIA}
\alias{module_gcamindia_L244.building_INDIA}
\title{module_gcamindia_L244.building_INDIA}
\usage{
module_gcamindia_L244.building_INDIA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\description{
Creates GCAM-INDIA building output files for writing to xml.
}
\details{
Calculate shareweights, cost, price elasticity, calibrated, and other data for the building sector
}
\author{
AM Jul21
}
