% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA100.socioeconomics_india.R
\name{module_gcamindia_LA100.socioeconomics}
\alias{module_gcamindia_LA100.socioeconomics}
\title{module_gcamindia_LA100.socioeconomics}
\usage{
module_gcamindia_LA100.socioeconomics(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.pcGDP_thous90usd_state_india}, \code{L100.GDP_mil90usd_state_india}, \code{L100.Pop_thous_state_india}. The corresponding file in the
}
\description{
This chunk estimates state-wise per capita GDP, Popualtion, and  total GDP (at 90USD prices) in a time series.
}
\details{
This chunk includes a series of estimations. First is to fill the data gaps with linear interpolation,
followed by estimation of state-wise GDP at 90USD prices using shares and population.
}
\author{
AM Nov20
}
