% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA101.energy_balance_india.R
\name{module_gcamindia_LA101.energy_balance}
\alias{module_gcamindia_LA101.energy_balance}
\title{module_gcamindia_LA101.energy_balance}
\usage{
module_gcamindia_LA101.energy_balance(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L101.india_state_A10.EB_SecFuel_mtoe}, \code{L101.india_state_EB_EJ_state_S_F}.
}
\description{
Produce two ouput tables from GHG Platform state energy database:
\itemize{
 \item{L101.india_state_EB_EJ_state_S_F: Energy data by GCAM sector and fuel, state, and year; energy units in EJ, years from 2005 - 2010, includes only rows that have a defined sector and fuel}
 \item{L101.india_state_A10.EB_SecFuel_mtoe: Energy data by sector and fuel, GCAM sector and fuel, state, and year; energy units in mtoe, years from 1960-2011, includes all original data}
}
}
\details{
See above
}
\author{
AM Nov20
}
