% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA114.wind_india.R
\name{module_gcamindia_LA114.wind}
\alias{module_gcamindia_LA114.wind}
\title{module_gcamindia_LA114.wind}
\usage{
module_gcamindia_LA114.wind(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L114.india_state_CapacityFactor_wind}.
}
\description{
Compute capacity factors for wind by India state.
}
\details{
Computes capacity factors for wind by India state.
}
\author{
AM Nov20
}
