% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA115.rooftopPV_india.R
\name{module_gcamindia_LA115.rooftopPV}
\alias{module_gcamindia_LA115.rooftopPV}
\title{module_gcamindia_LA115.rooftopPV}
\usage{
module_gcamindia_LA115.rooftopPV(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L115.india_state_rsrc_rooftopPV}. The corresponding file in the
original data system was \code{LA115.RooftopPV.R} (gcam-india level1).
}
\description{
Prepare resource curves for rooftop PV (commercial and residential combined).
}
\details{
Prepare resource curves for rooftop PV (commercial and residential combined).
}
\author{
PNK Nov20
}
