% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA119.solar_india.R
\name{module_gcamindia_LA119.solar}
\alias{module_gcamindia_LA119.solar}
\title{module_gcamindia_LA119.solar}
\usage{
module_gcamindia_LA119.solar(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L119.india_state_CapFacScaler_PV}, \code{L119.india_state_CapFacScaler_CSP}.
}
\description{
Compute scalars by state to vary capacity factors by state for India.
}
\details{
This chunk computes scalars by state to vary capacity factors for central station PV and CSP technologies by state.
}
\author{
AM Nov20
}
