% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA132.industry_india.R
\name{module_gcamindia_LA132.industry}
\alias{module_gcamindia_LA132.industry}
\title{module_gcamindia_LA132.industry}
\usage{
module_gcamindia_LA132.industry(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L132.india_state_in_EJ_indnochp_F}, \code{L132.india_state_in_EJ_indchp_F}, \code{L132.india_state_out_EJ_indchp_F}, \code{L132.india_state_in_EJ_indfeed_F}. The corresponding file in the
original data system was \code{LA132.Industry.R} (gcam-india level1).
}
\description{
Provides industrial energy consumption and industrial feedstock consumption by region/fuel/historical year.
}
\details{
Provides for each US state industrial energy consumption and industrial feedstock consumption by region/fuel/historical year.
}
\author{
AM Nov20
}
