% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA1321.cement_india.R
\name{module_gcamindia_LA1321.cement}
\alias{module_gcamindia_LA1321.cement}
\title{module_gcamindia_LA1321.cement}
\usage{
module_gcamindia_LA1321.cement(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1321.out_Mt_india_state_cement_Yh}, \code{L1321.IO_GJkg_india_state_cement_F_Yh}, \code{L1321.in_india_EJ_state_cement_F_Y}.
}
\description{
Allocate across the states national cement production, input-output cofficients, and energy inputs to cement production
}
\details{
The tables for cement production, i.e., out, and energy inputs, i.e., in, were calculated by applying state shares to national data.

The state shares were determined by the states' relative values of cement shipments.

The input-out coefficients were downscaled to the states in proportation to the national data.
}
\author{
AM Dec20
}
