% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA142.buildings_india.R
\name{module_gcamindia_LA142.buildings}
\alias{module_gcamindia_LA142.buildings}
\title{module_gcamindia_LA142.buildings}
\usage{
module_gcamindia_LA142.buildings(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L142.india_state_in_EJ_comm_F}, \code{L142.india_state_in_EJ_resid_F}. The corresponding file in the
original data system was \code{LA142.Buildings.R} (gcam-india level1).
}
\description{
Provides residential and commercial building energy consumption by region/fuel/historical year.
}
\details{
Provides for each US state industrial energy consumption and industrial feedstock consumption by region/fuel/historical year.
}
\author{
AM Jul21
}
