% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA144.Commercial_gcamindia.R
\name{module_gcamindia_LA144.Commercial}
\alias{module_gcamindia_LA144.Commercial}
\title{module_gcamindia_LA144.Commercial}
\usage{
module_gcamindia_LA144.Commercial(command, ...)
}
\description{
Calculates commercial floorspace by state and energy consumption by state/fuel/end use
}
\author{
PNK Nov20
}
