% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA144.Residential_gcamindia.R
\name{module_gcamindia_LA144.Residential}
\alias{module_gcamindia_LA144.Residential}
\title{module_gcamindia_LA144.Residential
Calculate residential floorspace by state and residential energy consumption by state/fuel/end use.}
\usage{
module_gcamindia_LA144.Residential(command, ...)
}
\description{
module_gcamindia_LA144.Residential
Calculate residential floorspace by state and residential energy consumption by state/fuel/end use.
}
\details{
Calculate residential floorspace by state and residential energy consumption by state/fuel/end use.
}
\author{
PNK Nov20
}
