% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB123.electricity_india.R
\name{module_gcamindia_LB123.electricity}
\alias{module_gcamindia_LB123.electricity}
\title{module_gcamindia_LB123.electricity}
\usage{
module_gcamindia_LB123.electricity(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L123.india_state_in_EJ_elec_F}, \code{L123.india_state_out_EJ_elec_F}, \code{L123.india_state_in_EJ_ownuse_elec}, \code{L123.india_state_out_EJ_ownuse_elec}.
The corresponding file in the original data system was \code{LB123.Electricity.R} (gcam-india level1).
}
\description{
Calculate electricity fuel consumption, electricity generation, and inputs and outputs of net ownuse
(the electricity used by production/transformation facilities) by state.
}
\details{
By state, calculates electricity fuel consumption, electricity generation, and inputs and outputs of net ownuse.
}
\author{
VC Nov20
}
