% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1232.elec_subregions_india.R
\name{module_gcamindia_LB1232.elec_subregions}
\alias{module_gcamindia_LB1232.elec_subregions}
\title{module_gcamindia_LB1232.elec_subregions}
\usage{
module_gcamindia_LB1232.elec_subregions(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1232.india_state_out_EJ_sR_elec}. The corresponding file in the
original data system was \code{LB1232.Elec_subregions.R} (gcam-usa level1).
}
\description{
Aggregates Indian state electricity generation to electricity subregions/grids.
}
\details{
Aggregates USA state electricity generation to electricity subregions.
}
\author{
PNK Nov20
}
