% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_building_segments_USA_xml.R
\name{module_gcamusa_batch_building_segments_USA_xml}
\alias{module_gcamusa_batch_building_segments_USA_xml}
\title{module_gcamusa_batch_building_segments_USA_xml}
\usage{
module_gcamusa_batch_building_segments_USA_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{building_segments_USA.xml}, \code{building_segments_USA_rcp4p5.xml},
\code{building_segments_USA_rcp8p5.xml}. The corresponding file in the
original data system was \code{batch_segments_building_USA.xml} (dispatch branch gcamusa XML).
}
\description{
Construct XML data structure for \code{building_segments_USA.xml}.
}
