% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_elec_RPS_USA.R
\name{module_gcamusa_batch_RPS_USA_xml}
\alias{module_gcamusa_batch_RPS_USA_xml}
\title{module_gcamusa_batch_RPS_USA_xml}
\usage{
module_gcamusa_batch_RPS_USA_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{RPS_USA_Gas.xml}, \code{RPS_USA_noGas.xml}. The corresponding file in the
original data system was \code{batch_RPS_USA.xml.R} (gcamusa XML batch).
}
\description{
Construct XML data structure for \code{RPS_USA.xml}.
}
