% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_xml_coal_retire.R
\name{module_gcamusa_coal_retire_xml}
\alias{module_gcamusa_coal_retire_xml}
\title{module_gcamusa_coal_retire_xml}
\usage{
module_gcamusa_coal_retire_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{coal_retire_vintage_USA.xml}.
}
\description{
Construct XML data structure for \code{coal_retire_vintage_USA.xml}.
}
