% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_xml_elecS_costs_ptc.R
\name{module_gcamusa_elecS_costs_ptc_xml}
\alias{module_gcamusa_elecS_costs_ptc_xml}
\title{module_gcamusa_elecS_costs_ptc_xml}
\usage{
module_gcamusa_elecS_costs_ptc_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{elecS_costs_USA_ptc.xml}. The corresponding file in the
original data system was \code{batch_elecS_costs_USA_ptc.xml} (gcamusa xml-batch).
}
\description{
Construct XML data structure for \code{elecS_costs_USA_ptc.xml}.
}
