% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_xml_HDDCDD_AEO_2015.R
\name{module_gcamusa_HDDCDD_AEO_2015_xml}
\alias{module_gcamusa_HDDCDD_AEO_2015_xml}
\title{module_gcamusa_HDDCDD_AEO_2015_xml}
\usage{
module_gcamusa_HDDCDD_AEO_2015_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
The generated outputs: \code{HDDCDD_AEO_2015_USA.xml}.
The corresponding file in the original data system was \code{batch_HDDCDD_QER_QER_USA.xml} (gcamusa XML batch).
}
\description{
Construct XML data structure for \code{HDDCDD_AEO_2015_USA.xml}.
}
