% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L102.water.supply.unlimited_USA.R
\name{module_gcamusa_L102.water.supply.unlimited}
\alias{module_gcamusa_L102.water.supply.unlimited}
\title{module_gcamusa_L102.water.supply.unlimited}
\usage{
module_gcamusa_L102.water.supply.unlimited(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L102.unlimited_water_price_state_R_W_Y_75USDm3}. The corresponding file in the
original data system was \code{L102.water.supply.unlimited.R} (water level1).
}
\description{
Summarize the water price values by water type and US state.
}
\details{
We generate all the possible combinations between the 50 GCAM-USA states, 4 water types, and all the defined years.
}
\author{
NTG August 2019
}
