% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L104.bcoc_en_S_T_Y.R
\name{module_gcamusa_L104.bcoc_en_S_T_Y}
\alias{module_gcamusa_L104.bcoc_en_S_T_Y}
\title{module_gcamusa_L104.bcoc_en_S_T_Y}
\usage{
module_gcamusa_L104.bcoc_en_S_T_Y(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L104.bcoc_tgej_USA_en_T_1990}. The corresponding file in the
original data system was \code{L104.bcoc_en_USA_S_T_Y.R} (emissions level1).
}
\description{
Creates USA BC and OC emission factors for combustion sectors
}
\details{
1990 USA BC and OC emissions are divided by year 2000 USA energy use to generate emission factors for GCAM aggregate sectors
}
\author{
SJS May 2017
}
