% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L120.offshore_wind_reeds.R
\name{module_gcamusa_L120.offshore_wind_reeds}
\alias{module_gcamusa_L120.offshore_wind_reeds}
\title{module_gcamusa_L120.offshore_wind_reeds}
\usage{
module_gcamusa_L120.offshore_wind_reeds(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L120.RsrcCurves_EJ_R_offshore_wind_USA}, \code{L120.GridCost_offshore_wind_USA}, \code{L120.RegCapFactor_offshore_wind_USA}.
The corresponding file in the
original data system was \code{LA120.offshore_wind_reeds_USA.R} (gcam_usa level1).}
}
\value{
Depends on \code{command}: either a vector of required inputs,
}
\description{
Takes in data on country-level offshore wind energy potential and global offshore wind capital cost assumptions and generates tables containing region-level offshore wind data.
}
\details{
Takes in data on US-level offshore wind energy potential and global offshore wind capital cost assumptions and generates tables containing state-level offshore wind data.
}
\author{
MB GI AJS March 2019
}
