% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L1234.elec_gridregions.R
\name{module_gcamusa_L1234.elec_gridregions}
\alias{module_gcamusa_L1234.elec_gridregions}
\title{module_gcamusa_L1234.elec_gridregions}
\usage{
module_gcamusa_L1234.elec_gridregions(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1234.in_EJ_grid_elec_F}, \code{L1234.out_EJ_grid_elec_F}.
The corresponding file in the original data system was \code{LB1234.elec_gridregions.R} (gcam-usa level1).
}
\description{
Calculate electricity fuel consumption and electricity generation by grid region.
}
\details{
By grid region, calculates electricity fuel consumption and electricity generation.
}
\author{
MTB August 2018
}
