% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L1239.elec_state_fractions.R
\name{module_gcamusa_L1239.elec_state_fractions}
\alias{module_gcamusa_L1239.elec_state_fractions}
\title{module_gcamusa_L1239.elec_state_fractions}
\usage{
module_gcamusa_L1239.elec_state_fractions(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1239.state_elec_supply_USA}.

The corresponding file in the original data system was \code{LB1239.elec_state_fractions.R} (gcam-usa level1).
}
\description{
Map electricity generation by fuel | grid region | horizontal segment to generation by fuel | state | segment.
The fraction of generation by fuel by horizontal segment is assumed to be equal for all states within a grid region.
}
\details{
Calculates the fraction of electricity generation by fuel, by horizontal load segment, by state.
}
\author{
MTB August 2018
}
