% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L132.water_demand_industry.R
\name{module_gcamusa_L132.water_demand_industry}
\alias{module_gcamusa_L132.water_demand_industry}
\title{module_gcamusa_L132.water_demand_industry}
\usage{
module_gcamusa_L132.water_demand_industry(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L132.water_km3_state_ind_Yh}.
}
\description{
Computes industrial water withdrawals and consumption by US state and historical year.
}
\details{
This chunks extends the USGS industrial water withdrawals and consumption data to all historical years. Missing years are
filled with interpolated water coefficients and state industrial energy consumption.
}
\author{
RC March 2019
}
