% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L143.HDDCDD.R
\name{module_gcamusa_L143.HDDCDD}
\alias{module_gcamusa_L143.HDDCDD}
\title{module_gcamusa_L143.HDDCDD}
\usage{
module_gcamusa_L143.HDDCDD(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L143.share_state_Pop_CDD_sR9}, \code{L143.share_state_Pop_CDD_sR13}, \code{L143.share_state_Pop_HDD_sR9}, \code{L143.share_state_Pop_HDD_sR13}, \code{L143.HDDCDD_scen_state}. The corresponding file in the
original data system was \code{LA143.HDDCDD.R} (gcam-usa level1).
}
\description{
Estimate heating and cooling degree days for gcam-usa.
}
\details{
Estimate heating and cooling degree days for gcam-usa.
}
\author{
KRD Nov 2017
}
