% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L145.water_demand_municipal.R
\name{module_gcamusa_L145.water_demand_municipal}
\alias{module_gcamusa_L145.water_demand_municipal}
\title{module_gcamusa_L145.water_demand_municipal}
\usage{
module_gcamusa_L145.water_demand_municipal(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L145.municipal_water_state_W_Yh_km3}, \code{L145.municipal_water_cost_state_75USD_m3},
\code{L145.municipal_water_eff_state_Yh}.
}
\description{
Generate US state municipal water withdrawals, municipal water base delivery cost, and municipal water use efficiency.
}
\details{
Generate municipal water withdrawals, municipal water base delivery cost, and municipal water use efficiency
by US state and historical year.
}
\author{
RC March 2019
}
