% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L170.nonghg_ceds_scaling_USA.R
\name{module_gcamusa_L170.nonghg_ceds_scaling_USA}
\alias{module_gcamusa_L170.nonghg_ceds_scaling_USA}
\title{module_gcamusa_L170.nonghg_ceds_scaling_USA}
\usage{
module_gcamusa_L170.nonghg_ceds_scaling_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L170.NEI_1990_2017_GCAM_sectors}, \code{L170.NEI_CEDS_scaling_diagnostic}
}
\description{
Scales pre-processed NEI non-CO2 emissions to CEDS at the national level
The NEI data is processed exogenously. It is scaled to EPA Tier 1, interpolated
between NEI years, extrapolated back to 1990, and mapped to GCAM sectors,
CEDS sectors, and CEDS fuels
}
\details{
Non-CO2 emissions from NEI scaled to CEDS
}
\author{
MAW February 2021
}
