% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L171.nonghg_trn_USA.R
\name{module_gcamusa_L171.nonghg_trn_USA}
\alias{module_gcamusa_L171.nonghg_trn_USA}
\title{module_gcamusa_L171.nonghg_trn_USA}
\usage{
module_gcamusa_L171.nonghg_trn_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L171.nonco2_tgpkm_censusR_trn_SMarkal_F_V_Y}, \code{L171.nonco2_tgpkm_state_trn_SMarkal_F_Y}. The corresponding file in the
original data system was \code{LA171.nonCO2_trn_USA_S_T_Y.R} (gcam-usa level1).
}
\description{
Emissions factors for transportation by MARKAL technology in the USA
}
\details{
Emissions factors in Tg/million pass-km for transportation by MARKAL technology in the USA.
In L171.nonco2_tgpkm_censusR_trn_SMarkal_F_V_Y, the evolution of EFs are computed with two different methods
1) For the future evolution of fleet EFs for vehicles existing in the base year, we account for each
individual vintage degrading as they age, and older vintages retiring.
2) For the degradation of future year EFs, EFs are taken directly from MARKAL aside from a unit conversion.
In L171.nonco2_tgpkm_state_trn_SMarkal_F_Y, EFs are computed with two different methods:
1) Base year EFs are the weighted means of the vintaged data using the age fractions as weights.
2) Future year EFs are taken directly from MARKAL for the year = vintage.
}
\author{
BY Aug 2019, MAW March 2022
}
