% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L201.socioeconomics_USA.R
\name{module_gcamusa_L201.socioeconomics_USA}
\alias{module_gcamusa_L201.socioeconomics_USA}
\title{module_gcamusa_L201.socioeconomics_USA}
\usage{
module_gcamusa_L201.socioeconomics_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L201.InterestRate_GCAMUSA}, \code{L201.Pop_GCAMUSA},\code{L201.Pop_GCAMUSA_SSP}, \code{L201.BaseGDP_GCAMUSA},
\code{L201.LaborForceFillout_GCAMUSA}, \code{L201.LaborProductivity_GCAMUSA}, \code{L201.Pop_national_updated_USA},
\code{L201.Pop_national_updated_USA_SSP}, \code{L201.BaseGDP_national_updated_USA}, \code{L201.LaborProductivity_national_updated_USA}.
The corresponding file in the original data system was \code{L201.socioeconomics_USA.R} (gcam-usa level2).
}
\description{
Interest rate, population, and GDP for GCAM-USA.
}
\details{
Interest rate, population, and GDP for GCAM-USA.
}
\author{
RLH October 2017
}
