% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L203.water_mapping_USA.R
\name{module_gcamusa_L203.water_mapping_USA}
\alias{module_gcamusa_L203.water_mapping_USA}
\title{module_gcamusa_L203.water_mapping_USA}
\usage{
module_gcamusa_L203.water_mapping_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L203.Supplysector}, \code{L203.SubsectorLogit}, \code{L203.SubsectorShrwtFllt}, \code{L203.TechShrwt}, \code{L203.TechCoef}. The corresponding file in the
original data system was \code{L203.water.mapping.R} (water level2).
}
\description{
Mapping of water consumption/withdrawal to sectoral demands at the state level.
}
\details{
Generates water mapping sector input files to group demands by sectors.
}
\author{
NTG Oct 2019
}
