% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L203.water_td_USA.R
\name{module_gcamusa_L203.water_td_USA}
\alias{module_gcamusa_L203.water_td_USA}
\title{module_gcamusa_L203.water_td_USA}
\usage{
module_gcamusa_L203.water_td_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L203.DeleteSupplysector_USA}, \code{L203.Supplysector_USA}, \code{L203.SubsectorLogit_USA},
\code{L203.SubsectorShrwt_USA}, \code{L203.TechShrwt_USA}, \code{L203.TechCoef_USA}, \code{L203.TechPmult_USA},
\code{L203.TechDesalCoef_USA}, \code{L203.TechDesalShrwt_USA}, \code{L203.TechDesalCost_USA}.
The corresponding file in the original data system was \code{L203.water.mapping.R} (water level2).
}
\description{
Mapping of water consumption/withdrawal to sectoral demands at the state level.
}
\details{
Generates water mapping sector input files to group demands by sectors.
}
\author{
NTG May 2020
}
