% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L2235.elec_segments_FERC.R
\name{module_gcamusa_L2235.elec_segments_FERC}
\alias{module_gcamusa_L2235.elec_segments_FERC}
\title{module_gcamusa_L2235.elec_segments_FERC}
\usage{
module_gcamusa_L2235.elec_segments_FERC(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2235.DeleteSupplysector_elec_USA}, \code{L2235.InterestRate_FERC_USA},
\code{L2235.Pop_FERC_USA}, \code{L2235.BaseGDP_FERC_USA}, \code{L2235.LaborForceFillout_FERC_USA},
\code{L2235.Supplysector_elec_USA}, \code{L2235.ElecReserve_elecS_grid_vertical_USA}, \code{L2235.SubsectorLogit_elec_USA},
\code{L2235.SubsectorShrwtFllt_elec_USA}, \code{L2235.SubsectorInterp_elec_USA}, \code{L2235.SubsectorShrwt_elec_USA},
\code{L2235.SubsectorInterpTo_elec_USA}, \code{L2235.SubsectorShrwtFllt_elecS_grid_vertical_USA},
\code{L2235.SubsectorShrwtInterp_elecS_grid_vertical_USA}, \code{L2235.TechShrwt_elec_USA}, \code{L2235.TechCoef_elec_USA},
\code{L2235.Production_exports_elec_USA}, \code{L2235.TechShrwt_elecS_grid_vertical_USA}, \code{L2235.TechCoef_elecS_grid_vertical_USA},
\code{L2235.Supplysector_elec_FERC_USA}, \code{L2235.SubsectorLogit_elec_FERC_USA}, \code{L2235.SubsectorShrwtFllt_elec_FERC_USA},
\code{L2235.SubsectorInterp_elec_FERC_USA}, \code{L2235.SubsectorShrwt_elec_FERC}, \code{L2235.SubsectorInterpTo_elec_FERC},
\code{L2235.TechShrwt_elec_FERC_USA}, \code{L2235.TechCoef_elec_FERC_USA}, \code{L2235.TechCoef_elecownuse_FERC_USA},
\code{L2235.Production_imports_FERC_USA},\code{L2235.Production_elec_gen_FERC_USA}.
The corresponding file in the original data system was \code{L2235.elec_segments_FERC.R} (gcam-usa level2).
}
\description{
Creates the vertical segment supplysectors in the grid regions and also the domestic supply and
electricity trade sectors.
}
\details{
This chunk generates input files to create the vertical segment supplysectors in the grid regions
and also the domestic supply and electricity trade sectors for the grid regions.
}
\author{
MTB Aug 2018
}
