% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2236.elecS_ghg_emissions_USA.R
\name{module_gcamusa_L2236.elecS_ghg_emissions_USA}
\alias{module_gcamusa_L2236.elecS_ghg_emissions_USA}
\title{module_gcamusa_L2236.elecS_ghg_emissions_USA}
\usage{
module_gcamusa_L2236.elecS_ghg_emissions_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs:  \code{L2236.elecS_ghg_emissions_USA}, \code{L2236.elecS_cool_ghg_tech_coeff_USA}.
The corresponding file in the original data system was \code{L2236.elecS_ghg_emissions_USA} (gcam-usa level2).
}
\description{
U.S. electricity non CO2 GHG emission and future emission coefficients by sector, fuel and cooling techs
}
\details{
Write electricity emission coefficients to multiple load segments and cooling technology then compute state
shares for each category in the fuel input table.
}
\author{
KD August 2018 / YO March 2022
}
