% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2236.rooftop_PV_USA.R
\name{module_gcamusa_L2236.rooftop_PV_USA}
\alias{module_gcamusa_L2236.rooftop_PV_USA}
\title{module_gcamusa_L2236.rooftop_PV_USA}
\usage{
module_gcamusa_L2236.rooftop_PV_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2236.Supplysector_RPV_USA}, \code{L2236.ElecReserve_RPV_USA},
\code{L2236.SubsectorLogit_RPV_USA}, \code{L2236.SubsectorShrwtFllt_RPV_USA},
\code{L2236.SubsectorInterpTo_RPV_USA}, \code{L2236.StubTech_RPV_USA}, \code{L2236.StubTechCapFactor_RPV_USA},
\code{L2236.StubTechMarket_RPV_USA}, \code{L2236.StubTechElecMarket_RPV_USA}.
}
\description{
Create rooftop PV technologies for GCAM-USA.
}
\details{
Create rooftop PV technologies for GCAM-USA
}
\author{
MTB September 2018
}
