% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2241.coal_retire_USA.R
\name{module_gcamusa_L2241.coal_retire_USA}
\alias{module_gcamusa_L2241.coal_retire_USA}
\title{module_gcamusa_L2241.coal_retire_USA}
\usage{
module_gcamusa_L2241.coal_retire_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2241.Production_coalret_vintage_dispatch_gcamusa}, \code{L2241.CapacityTech_coalret_vintage_dispatch_gcamusa},
\code{L2241.TechEff_coalret_vintage_dispatch_gcamusa}, \code{L2241.TechSCurve_coalret_vintage_dispatch_gcamusa},
\code{L2241.TechShrwt_coalret_vintage_dispatch_gcamusa}, \code{L2241.TechOMvar_coalret_vintage_dispatch_gcamusa},
\code{L2241.TechCapFac_coalret_vintage_dispatch_gcamusa}, \code{L2241.CapacityTechAvail_coalret_vintage_dispatch_gcamusa},
\code{L2241.CapacityTechMinCapFac_coalret_vintage_dispatch_gcamusa}, \code{L2241.TechProfitShutdown_coalret_vintage_dispatch_gcamusa}.
The corresponding file in the original data system was \code{L2241.coal_slow_fast_retire_USA.R} (gcam-usa level2).
}
\description{
Generates GCAM-USA model input for removing coal capacity retired between 2011 and 2015, 2016 and 2020, and vintaging capacity which continues to operate beyond 2020.
}
\details{
This chunk creates add-on files to take the fraction of reduction in coal electricity generation between 2010 and 2015 for each state and
forces that generation to retire in 2015. It also tempers retirement assumptions for the remaining coal fleet to allow
most 2015 generation to continue through mid-century.
}
\author{
RC Aug 2018 / YO Jul 2020
}
