% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2244.nuclear_USA.R
\name{module_gcamusa_L2244.nuclear_USA}
\alias{module_gcamusa_L2244.nuclear_USA}
\title{module_gcamusa_L2244.nuclear_USA}
\usage{
module_gcamusa_L2244.nuclear_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2244.CapacityTechYr_nuc_gen2_USA}, \code{L2244.TechSCurve_nuc_gen2_USA}.
The corresponding file in the original data system was \code{L2244.nuclear_USA.R} (gcam-usa level2).
}
\description{
Generates an add-on file to update nuclear assumptions in GCAM-USA.
}
\details{
This chunk creates an add-on file to update nuclear assumptions in GCAM-USA. Specifically, it reads in state-specific
s-curve retirement functions and lifetimes for existing nuclear vintage, based on planned retirements by state and nuclear power plant.
The current and future power generation data from existing nuclear plants in the USA is based on US Nuclear Regulatory Commission (NRC) of
nuclear reactor characteristic and operational history from January 2011, with additional updates to operating licenses and plant closures.
}
\author{
RC Sep 2018; edited MB Sep 2018
}
