% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L2245.geothermal_fixed.R
\name{module_gcamusa_L2245.geothermal_fixed}
\alias{module_gcamusa_L2245.geothermal_fixed}
\title{module_gcamusa_L2245.geothermal_fixed}
\usage{
module_gcamusa_L2245.geothermal_fixed(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2245.DeleteRenewRsrc_USAgeo}, \code{L2245.DeleteSupplysector_USAgeo},
\code{L2245.GlobalTechShrwt_USAgeo},\code{L2245.StubTechFixOut_geothermal_USA},
\code{L2245.SubsectorLogit_USAgeo},\code{L2245.SubsectorLogit_cool_USAgeo}.
}
\description{
Provide the option for users to fix future geothermal generation capabilities to base year values.
}
\details{
Update state-level geothermal outputs to fixed while deleting renewresource
}
\author{
NTG February 2022
}
