% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2245a.elec_RPS_USA_Gas.R
\name{module_gcamusa_L2245a.elec_RPS_USA_Gas}
\alias{module_gcamusa_L2245a.elec_RPS_USA_Gas}
\title{module_gcamusa_L2245a.elec_RPS_USA_Gas}
\usage{
module_gcamusa_L2245a.elec_RPS_USA_Gas(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2245.StubTechRESSecondaryOutput_RPS_USA_Gas},
\code{L2245.StubTechAdjCoef_RPS_USA_Gas}, \code{L2245.RESPolicy_RPS_USA_Gas}.
The corresponding file in the original data system was \code{L2245.RPS_USA.R} (gcam-usa level2).
}
\description{
Generates the vintage coefficient-based RPS structure for GCAM-USA created for UMD/CGS Projects.
}
\details{
This chunk generates input files to create an electricity generation sector with multiple load segments
for each state and creates the demand for the state-level electricity sectors in the grid regions.
}
\author{
AJS Oct 2019 / YO Jan 2020
}
