% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L2247.elecS_tech_costs.R
\name{module_gcamusa_L2247.elecS_tech_costs}
\alias{module_gcamusa_L2247.elecS_tech_costs}
\title{module_gcamusa_L2247.elecS_tech_costs}
\usage{
module_gcamusa_L2247.elecS_tech_costs(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2247.GlobalTechCapitalOnly_elecS_USA}, \code{L2247.GlobalIntTechCapitalOnly_elecS_USA},
\code{L2247.GlobalTechCapitalOnly_elecS_adv_USA}, \code{L2247.GlobalIntTechCapitalOnly_elecS_adv_USA},
\code{L2247.GlobalTechOMfixedOnly_elecS_adv_USA}, \code{L2247.GlobalIntTechOMfixedOnly_elecS_adv_USA},
\code{L2247.GlobalTechOMvar_elecS_adv_USA}, \code{L2247.GlobalIntTechOMvar_elecS_adv_USA}, \code{L2247.GlobalTechFCROnly_elecS_itc_USA},
\code{L2247.GlobalIntTechFCROnly_elecS_itc_USA}, \code{L2247.GlobalTechCost_ptc_USA}, \code{L2247.GlobalIntTechCost_ptc_USA}.
The corresponding file in the original data system was \code{L2247.elecS_tech_costs.R} (gcam-usa level2).
}
\description{
Generates add-ons to harmonize reference case generation technology costs for multiple load segments with AEO-2016.
}
\details{
This chunk creates add-ons to harmonize reference case generation technology costs for multiple load segments with AEO-2016.
These assumptions are also consistent with the Base case scenario. For technologies in GCAM that are not listed in the AEO,
we simply use the cost assumptions from the global version of GCAM. This chunk also creates add-ons for Advanced Tech scenario
for the multiple load segments consistent with the Advanced technology scenario.
This chunk also creates add-ons for ITC and PTC policies for multiple load segments technologies.
}
\author{
RC Sep 2018
}
