% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L231.proc_sector_USA.R
\name{module_gcamusa_L231.proc_sector_USA}
\alias{module_gcamusa_L231.proc_sector_USA}
\title{module_gcamusa_L231.proc_sector_USA}
\usage{
module_gcamusa_L231.proc_sector_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs:\code{L231.DeleteSupplysector_industry_USA}, \code{L231.DeleteSupplysector_urban_processes_USA},
\code{L231.DeleteFinalDemand_urban_processes_USA}, \code{L231.UnlimitRsrc_USA}, \code{L231.UnlimitRsrcPrice_USA}, \code{L231.FinalDemand_urb_USA},
\code{L231.Supplysector_urb_ind_USA}, \code{L231.SubsectorLogit_urb_ind_USA},
\code{L231.SubsectorShrwt_urb_ind_USA}, \code{L231.SubsectorShrwtFllt_urb_ind_USA},
\code{L231.SubsectorInterp_urb_ind_USA}, \code{L231.SubsectorInterpTo_urb_ind_USA},
\code{L231.StubTech_urb_ind_USA}, \code{L231.RegionalTechCalValue_urb_ind_USA}, \code{L231.IndCoef_USA}.
}
\description{
Writes urban & industrial processing sector outputs for USA.
}
\details{
For urban processes and industrial processes, subsectors
and supplysectors, produces logit, shareweights, and interpolation files by writing assumption file data to all regions.
Writes out coefficients and prices for misc emissions sources from assumption file data.
Outputs urban processing demand using constants. StubTech mapping file created from assumption mappings.
Regional technology calibration values created with constant. Industry input-output coefficient file created.
}
\author{
BY September 2019
}
