% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L232.water_demand_industry.R
\name{module_gcamusa_L232.water_demand_industry}
\alias{module_gcamusa_L232.water_demand_industry}
\title{module_gcamusa_L232.water_demand_industry}
\usage{
module_gcamusa_L232.water_demand_industry(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L232.TechCoef_USA}.
}
\description{
Computes industrial water withdrawal/consumption coefficients (m3/GJ output) by US state and year.
}
\details{
This chunk generates industrial water withdrawal and consumption coefficients by US state and year.
Water coefficients are as total withdrawals/consumption divided by industrial sector output (base-service).
}
\author{
RC March 2019
}
