% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L2322.industry_vintage.R
\name{module_gcamusa_L2322.industry_vintage}
\alias{module_gcamusa_L2322.industry_vintage}
\title{module_gcamusa_L2322.industry_vintage}
\usage{
module_gcamusa_L2322.industry_vintage(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2322.StubTechSCurve_industry_USA}, \code{L2322.StubTechProfitShutdown_industry_USA}.
 The corresponding file in the original data system was \code{L2322.industry_vintage_USA.R} (gcam-usa level2).
}
\description{
Creates a vintage structure and retirement parameters for U.S. industrial energy use
}
\details{
This chunk creates a vintage structure and retirement parameters for U.S. industrial energy use.
}
\author{
KD August 2018; edited MTB October 2018
}
