% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L244.building_det_USA.R
\name{module_gcamusa_L244.building_det_USA}
\alias{module_gcamusa_L244.building_det_USA}
\title{module_gcamusa_L244.building_det_USA}
\usage{
module_gcamusa_L244.building_det_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: .
}
\description{
Construct XML data structure for \code{building_USA.xml}.
}
