% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2441.building_segments_USA.R
\name{module_gcamusa_L2441.building_segments_USA}
\alias{module_gcamusa_L2441.building_segments_USA}
\title{module_gcamusa_L2441.building_segments_USA}
\usage{
module_gcamusa_L2441.building_segments_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2441.DeleteThermalService_gcamusa}, \code{L2441.ThermalBaseService_gcamusa}, \code{L2441.ThermalServiceSatiation_gcamusa},
\code{L2441.ThermalServiceCalSatiationValue_gcamusa}, \code{L2441.ThermalDefaultCoef_gcamusa}, \code{L2441.Intgains_scalar_gcamusa},
\code{L2441.Supplysector_bld_gcamusa}, \code{L2441.FinalEnergyKeyword_bld_gcamusa}, \code{L2441.SubsectorLogit_bld_gcamusa},
\code{L2441.SubsectorShrwtFllt_bld_gcamusa}, \code{L2441.SubsectorInterp_bld_gcamusa}, \code{L2441.SubsectorInterpTo_bld_gcamusa},
\code{L2441.StubTechFromSector_bld_gcamusa}, \code{L2441.StubTechDeleteInput_gcamusa}, \code{L2441.StubTechEff_segmentinputs_gcamusa},
\code{L2441.StubTechCalInput_bld_gcamusa}, \code{L2441.StubTechMarket_bld_gcamusa}, \code{L2441.TechCoef_nonthermal_load_curve_gcamusa},
\code{L2441.HDDCDD_Fixed_gcamusa}, \code{L2441.HDDCDD_Fixed_rcp4p5_gcamusa}, \code{L2441.HDDCDD_Fixed_rcp8p5_gcamusa}.
The corresponding file in the original data system was \code{L2441.building_segments_USA.R} (dispatch branch gcam-usa level2).
}
\description{
Creates GCAM-USA building dispatch output files for writing to xml.
}
\details{
Creates GCAM-USA building output files for writing to xml.
}
\author{
YO May 2020
}
