% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L245.water_demand_municipal.R
\name{module_gcamusa_L245.water_demand_municipal}
\alias{module_gcamusa_L245.water_demand_municipal}
\title{module_gcamusa_L245.water_demand_municipal}
\usage{
module_gcamusa_L245.water_demand_municipal(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L245.Supplysector_USA}, \code{L245.SubsectorLogit_USA}, \code{L245.SubsectorShrwtFllt_USA},
\code{L245.TechShrwt_USA}, \code{L245.TechCoef_USA}, \code{L245.TechCost_USA}, \code{L245.PerCapitaBased_USA},
\code{L245.BaseService_USA}, \code{L245.IncomeElasticity_USA}, \code{L245.PriceElasticity_USA}, \code{L245.aeei_USA}.
}
\description{
Genereate GCAM-USA municipal water sector input files.
}
\details{
This chunk generates the input files of municipal water information (cost, efficiency, coefficients)
across US states and model years.
}
\author{
RC March 2019
}
