% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L270.nonghg_NEI_to_GCAM_USA.R
\name{module_gcamusa_L270.nonghg_nei_to_gcam}
\alias{module_gcamusa_L270.nonghg_nei_to_gcam}
\title{module_gcamusa_L270.nonghg_nei_to_gcam}
\usage{
module_gcamusa_L270.nonghg_nei_to_gcam(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L270.nonghg_tg_state_elec_F_Yb}, \code{L270.nonghg_tg_state_refinery_F_Yb}, \code{L270.nonghg_tg_state_bld_F_Yb},
\code{L270.nonghg_tg_state_indenergy_F_Yb}, \code{L270.nonghg_tg_state_othertrn_F_Yb}, \code{L270.nonghg_tg_state_prc_F_Yb}.
}
\description{
Produce tables containing non-GHG emissions for all model base years by sector.
}
\details{
This chunk isolates sectoral non-ghg input emissions by U.S. state / sector / fuel / pollutant / year from NEI.
}
\author{
MAW February 2022
}
