% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L271.ghg_trn_USA.R
\name{module_gcamusa_L271.ghg_trn_USA}
\alias{module_gcamusa_L271.ghg_trn_USA}
\title{module_gcamusa_L271.ghg_trn_USA}
\usage{
module_gcamusa_L271.ghg_trn_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L271.ghg_trn_tech_coeff_USA}
The corresponding file in the original data system was \code{L271.trn_nonCO2_USA.R} (gcam-usa level2).
}
\description{
Non-CO2 GHG emissions parameters for transportation technologies in the USA
}
\details{
Prepare level 2 transportation sector non-CO2 GHG emissions files for USA.
}
\author{
BY September 2019 / YO 2022
}
