% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L272.nonghg_elc_USA.R
\name{module_gcamusa_L272.elc_nonghg_USA}
\alias{module_gcamusa_L272.elc_nonghg_USA}
\title{module_gcamusa_L272.elc_nonghg_USA}
\usage{
module_gcamusa_L272.elc_nonghg_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2721.nonghg_elec_tech_coeff_USA},\code{L2721.elec_MAC_CSAPR}. The corresponding file in the
original data system was \code{L272.elc_nonghg_USA.R} (gcam-usa level2).
}
\description{
Calculate Non-GHG emissions parameters for electricity technologies in the USA
}
\details{
This chunk calculates Non-GHG emissions parameters for electricity technologies in the USA and the electricity MAC curve
}
\author{
AJS and BY August 2019, MAW March 2022
}
