% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2722.nonghg_elc_linear_control_USA.R
\name{module_gcamusa_L2722.nonghg_elc_linear_control_USA}
\alias{module_gcamusa_L2722.nonghg_elc_linear_control_USA}
\title{module_gcamusa_L2722.nonghg_elc_linear_control_USA}
\usage{
module_gcamusa_L2722.nonghg_elc_linear_control_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L272.nonghg_elec_tech_coeff_USA_linear_control}, \code{L2722.nonghg_elec_tech_coeff_USA_linear_control_off}.
}
\description{
Calculate NOx and SO2 emission factors and linear control parameters for coal - electricity technologies in the USA
}
\details{
This chunk calculates NOx and SO2 emission factors and linear control parameters for coal - electricity technologies in the USA
}
\author{
MAW October 2020
}
