% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L273.en_ghg_emissions_USA.R
\name{module_gcamusa_L273.en_ghg_emissions_USA}
\alias{module_gcamusa_L273.en_ghg_emissions_USA}
\title{module_gcamusa_L273.en_ghg_emissions_USA}
\usage{
module_gcamusa_L273.en_ghg_emissions_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L273.en_ghg_tech_coeff_USA}, \code{L273.en_ghg_emissions_USA}, \code{L273.out_ghg_emissions_USA},
and \code{L273.MAC_higwp_USA}. The corresponding file in the
original data system was \code{L273.en_ghg_emissions_USA.R} (gcam-usa level2).
}
\description{
Define non-CO2 GHG emissions for GCAM-USA states, including 1. CH4 and N2O in refinery,
buildings, N fertilizer, industrial energy use; 2. F-gases in cooling and power
transmission; 3. MACs for F-gases
}
\details{
KALYN MTB YO
}
\author{
KRD 2018; MTB 2018; YO 2021
}
