% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L273.nonghg_refinery.R
\name{module_gcamusa_L273.nonghg_refinery}
\alias{module_gcamusa_L273.nonghg_refinery}
\title{module_gcamusa_L273.nonghg_refinery}
\usage{
module_gcamusa_L273.nonghg_refinery(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L273.nonghg_state_refinery_USA}
}
\description{
Non-GHG input emissions parameters for refining sector in the USA
}
\details{
This chunk calculates Non-GHG emissions parameters for refining technologies in the USA.
}
\author{
MAW March 2022
}
