% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L274.nonghg_bld.R
\name{module_gcamusa_L274.nonghg_bld}
\alias{module_gcamusa_L274.nonghg_bld}
\title{module_gcamusa_L274.nonghg_bld}
\usage{
module_gcamusa_L274.nonghg_bld(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L274.nonghg_bld_tech_coeff_USA}. The corresponding file in the
original data system was \code{L274.bld_nonghg_USA.R} (gcam-usa level2).
}
\description{
Non-GHG input emissions parameters for buildings technologies in the USA
}
\details{
Non-GHG input emissions parameters for buildings technologies in the USA
}
\author{
BY Aug 2019, MAW July 2021
}
