% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L275.nonghg_indenergy_USA.R
\name{module_gcamusa_L275.indenergy_nonghg_USA}
\alias{module_gcamusa_L275.indenergy_nonghg_USA}
\title{module_gcamusa_L275.indenergy_nonghg_USA}
\usage{
module_gcamusa_L275.indenergy_nonghg_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L275.nonghg_indenergy_tech_coeff_USA}
The corresponding file in the original data system was \code{L275.indenergy_nonghg_USA.R} (gcam-usa level2).
}
\description{
Non-GHG input emissions parameters for industrial energy use sector in the USA
}
\details{
Non-GHG input emissions parameters for industrial energy use technologies in the USA
}
\author{
BY Aug 2019, MAW March 2022
}
