% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L276.nonghg_othertrn_USA.R
\name{module_gcamusa_L276.nonghg_othertrn_USA}
\alias{module_gcamusa_L276.nonghg_othertrn_USA}
\title{module_gcamusa_L276.nonghg_othertrn_USA}
\usage{
module_gcamusa_L276.nonghg_othertrn_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L276.nonghg_othertrn_tech_coeff_USA}. The corresponding file in the
original data system was \code{L276.othertrn_nonghg_USA.R} (gcam-usa level2).
}
\description{
Generates GCAM-USA input files of Non-GHG input emissions parameters for domestic aviation, ships, and rail.
}
\details{
This chunk generates Non-GHG input emissions parameters for domestic aviation, ships, and rail in the USA.
}
\author{
MAW March 2022
}
