% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_L277.ghg_prc.R
\name{module_gcamusa_L277.ghg_prc}
\alias{module_gcamusa_L277.ghg_prc}
\title{module_gcamusa_L277.ghg_prc}
\usage{
module_gcamusa_L277.ghg_prc(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L277.ghg_prc_USA}, \code{L277.MAC_prc_USA}, \code{L277.MAC_prc_tc_average_USA}, \code{L277.MAC_prc_phaseInTime_USA}.
}
\description{
Generates input ghg emissions and marginal abatement curves for industrial processes and urban processes by
energy technology for U.S. states
Writes out input emissions and MAC components to all process-related energy technologies and states
}
\details{
Generates input ghg emissions and marginal abatement curves for industrial processes and urban processes by
energy technology for U.S. states and writes out input emissions and MAC components to all process-related energy technologies and states
}
\author{
Yang Ou Feb 2023
}
