% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L277.nonghg_prc_USA.R
\name{module_gcamusa_L277.nonghg_prc_USA}
\alias{module_gcamusa_L277.nonghg_prc_USA}
\title{module_gcamusa_L277.nonghg_prc_USA}
\usage{
module_gcamusa_L277.nonghg_prc_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L277.nonghg_prc_USA}, \code{L277.nonghg_max_reduction_USA}, \code{L277.nonghg_steepness}.
}
\description{
Generates input emissions for process and cement sectors by energy technology for Non.CO2 air pollutants for U.S. states
Writes out max emissions reductions and steepness to all process-related energy technologies and states
}
\details{
Generates input emissions by energy technology for Non.CO2 air pollutants for U.S. states, Writes out max emissions reductions and steepness to all process-related energy technologies and states
}
\author{
BY September 2019, MAW May 2021
}
