% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA114.wind_USA.R
\name{module_gcamusa_LA114.wind}
\alias{module_gcamusa_LA114.wind}
\title{module_gcamusa_LA114.Wind}
\usage{
module_gcamusa_LA114.wind(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L114.CapacityFactor_wind_state_gcamusa}, \code{L114.CapacityFactor_wind_state_segment_gcamusa}.
The corresponding file in the original data system was \code{LA114.Wind.R} (gcam-usa level1).
}
\description{
Compute capacity factors for wind by US state.
}
\details{
Computes capacity factors for wind by US state.
}
\author{
ST September 2017 / YO June 2020
}
