% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA115.hydro_USA.R
\name{module_gcamusa_LA115.hydro}
\alias{module_gcamusa_LA115.hydro}
\title{module_gcamusa_LA115.hydro}
\usage{
module_gcamusa_LA115.hydro(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs:
\code{L115.CapacityFactor_hydro_state_gcamusa}, \code{L115.CapacityFactor_hydro_state_segment_gcamusa}.
}
\description{
Compute hydropower capacity factors by state and month.
}
\details{
This chunk computes capacity factors for hydropower by state and month.
}
\author{
MTB 2021/05
}
