% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA142.Building_USA.R
\name{module_gcamusa_LA142.Building_USA}
\alias{module_gcamusa_LA142.Building_USA}
\title{module_gcamusa_LA142.Building_USA}
\usage{
module_gcamusa_LA142.Building_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L142.in_EJ_state_bld_F}. The corresponding file in the
original data system was \code{LA142.Building.R} (gcam-usa level1).
}
\description{
Downscaling each state and sector's shares of USA building energy use by fuel
}
\details{
Scaled national-level building energy consumption by portion of total US building energy use by fuel for each state and sector from the SEDS table.
}
\author{
KD September 2017
}
