% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA161.Cstorage_USA.R
\name{module_gcamusa_LA161.Cstorage}
\alias{module_gcamusa_LA161.Cstorage}
\title{module_gcamusa_LA161.Cstorage}
\usage{
module_gcamusa_LA161.Cstorage(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L161.Cstorage_FERC}. The corresponding file in the
original data system was \code{LA161.Cstorage.R} (gcam-usa level1).
}
\description{
Calculates onshore CO2 storage by grid region.
}
\details{
Calculates onshore CO2 storage by grid region.
}
\author{
RLH October 2017
}
