% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA2233.elec_segments_water_USA.R
\name{module_gcamusa_LA2233.elec_segments_water_USA}
\alias{module_gcamusa_LA2233.elec_segments_water_USA}
\title{module_gcamusa_LA2233.elec_segments_water_USA}
\usage{
module_gcamusa_LA2233.elec_segments_water_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\description{
Generates GCAM-USA model inputs for multiple load segments electricity sector and cooling systems by state.
}
\details{
This chunk generates input files to create an electricity generation sector with multiple load segments
for each state and creates the demand by load, fuel, power plant, and cooling system type.
}
\author{
NTG May 2020
}
