% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB103.load_curve_enduse_USA.R
\name{module_gcamusa_LB103.load_curve_enduse_USA}
\alias{module_gcamusa_LB103.load_curve_enduse_USA}
\title{module_gcamusa_LB103.load_curve_enduse_USA}
\usage{
module_gcamusa_LB103.load_curve_enduse_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated output: \code{L103.load_segments_sector_gcamusa}. The corresponding file in the
original data system was \code{LA103.load_curves_enduse.R} (gcamusa dispatch level1).
}
\description{
Process EIA monthly retail electricity sales data by grid regions to get load curves for enduse sectors.
}
\details{
Compute load curve related parameters for enduse sectors
}
\author{
YO May 2020
}
