% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB104.load_curve_enduse_HDDCDD_USA.R
\name{module_gcamusa_LB104.load_curve_enduse_HDDCDD_USA}
\alias{module_gcamusa_LB104.load_curve_enduse_HDDCDD_USA}
\title{module_gcamusa_LB104.load_curve_enduse_HDDCDD_USA}
\usage{
module_gcamusa_LB104.load_curve_enduse_HDDCDD_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L104.HistoricalDD_S_Segment_gcamusa}, \code{L104.DD_S_Segment_all_gcamusa},
The corresponding file in the original data system was \code{LA104.load_curves_enduse_HDDCDD.R} (gcamusa dispatch level1).
}
\description{
Use HDDCDD to calculate buildings heating and cooling demand profiles.
}
\details{
Use HDDCDD to calculate buildings heating and cooling demand profiles.
}
\author{
YO May 2020
}
