% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB123.Electricity.R
\name{module_gcamusa_LB123.Electricity}
\alias{module_gcamusa_LB123.Electricity}
\title{module_gcamusa_LB123.Electricity}
\usage{
module_gcamusa_LB123.Electricity(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L123.in_EJ_state_elec_F_tech}, \code{L123.out_EJ_state_elec_F}, \code{L123.in_EJ_state_ownuse_elec},
\code{L123.out_EJ_state_ownuse_elec}, \code{L123.out_EJ_state_elec_F_tech}, \code{L123.capacity_EJ_state_elec_F_tech},
\code{L123.capacity_factor_EJ_state_elec_F_tech}
The corresponding file in the original data system was \code{LB123.Electricity.R} (gcam-usa level1).
}
\description{
Calculate electricity fuel consumption, electricity generation, and inputs and outputs of net ownuse
(the electricity used by production/transformation facilities) by state.
}
\details{
By state, calculates electricity fuel consumption, electricity generation, and inputs and outputs of net ownuse.
}
\author{
RLH August 2017 YO Jun 2020
}
