% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1233.Elec_water_Update.R
\name{module_gcamusa_LB1233.Elec_water_Update}
\alias{module_gcamusa_LB1233.Elec_water_Update}
\title{module_gcamusa_LB1233.Elec_water_Update}
\usage{
module_gcamusa_LB1233.Elec_water_Update(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1233.wdraw_coef_R_elec_F_tech_Yh_ref},\code{L1233.wcons_coef_R_elec_F_tech_Yh_ref}
The corresponding file in the
original data system was \code{LB1233.Elec_water.R} (gcam-usa level1).
}
\description{
Compute water withdrawals/consumption coefficients by state, fuel, technology, and cooling system type.
}
\details{
Compute water withdrawals/consumption coefficients by state, fuel, technology, and cooling system type.
}
\author{
Zarrar Khan September 2018
}
