% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1233.elec_water_USA.R
\name{module_gcamusa_LB1233.elec_water_USA}
\alias{module_gcamusa_LB1233.elec_water_USA}
\title{module_gcamusa_LB1233.elec_water_USA}
\usage{
module_gcamusa_LB1233.elec_water_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1233.out_EJ_state_elec_F_tech_cool}, \code{L1233.share_sR_elec_F_tech_cool}, \code{L1233.wdraw_km3_state_elec}, \code{L1233.wcons_km3_state_elec}. The corresponding file in the
original data system was \code{LB1233.Elec_water.R} (gcam-usa level1).
}
\description{
Compute water withdrawals/consumption by state, fuel, technology, and cooling system type.
}
\details{
Compute water withdrawals/consumption by state, fuel, technology, and cooling system type.
}
\author{
ST October 2017, NTG Oct 2019
}
