% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB172.nonghg_elc_USA.R
\name{module_gcamusa_LB172.nonghg_elc}
\alias{module_gcamusa_LB172.nonghg_elc}
\title{module_gcamusa_LB172.nonghg_elc}
\usage{
module_gcamusa_LB172.nonghg_elc(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L172.nonghg_tg_state_elec_F_Yb}, \code{L172.nonghg_tgej_state_elec_F_Yf}. The corresponding file in the
original data system was \code{LB172.nonghg_elc_USA.R} (gcam-usa level1).
}
\description{
Base-year and future year electricity non-ghg  emissions by U.S. state / fuel / pollutant / year
}
\details{
elec sector non-ghg emissions by U.S. state / sector / fuel / pollutant / base and future year.
}
\author{
YO July 2018
}
