% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB174.nonghg_bld_USA.R
\name{module_gcamusa_LB174.nonghg_bld}
\alias{module_gcamusa_LB174.nonghg_bld}
\title{module_gcamusa_LB174.nonghg_bld}
\usage{
module_gcamusa_LB174.nonghg_bld(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L174.nonghg_tg_state_bld_F_Yb}. The corresponding file in the
original data system was \code{LB174.nonghg_bld_USA.R} (gcam-usa level1).
}
\description{
Calculate non-ghg emission totals and non-ghg emission shares of total emissions.
}
\details{
Buildings sector non-ghg input emissions by U.S. state / sector / fuel / pollutant / year.
}
\author{
YO July 2018
}
