% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB175.nonghg_indenergy_USA.R
\name{module_gcamusa_LB175.nonghg_indenergy}
\alias{module_gcamusa_LB175.nonghg_indenergy}
\title{module_gcamusa_LB175.nonghg_indenergy}
\usage{
module_gcamusa_LB175.nonghg_indenergy(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L175.nonghg_tg_state_indenergy_F_Yb}. The corresponding file in the
original data system was \code{LB175.nonghg_indenergy_USA.R} (gcam-usa level1).
}
\description{
Calculate input emissions for industrial energy use sector by input fuel in the USA.
}
\details{
Industrial sector non-ghg input emissions by U.S. state / sector / fuel / pollutant / year.
}
\author{
YO July 2018
}
