% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcamusa_xml_regional_biomass.R
\name{module_gcamusa_regional_biomass_xml}
\alias{module_gcamusa_regional_biomass_xml}
\title{module_gcamusa_regional_biomass_xml}
\usage{
module_gcamusa_regional_biomass_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{regional_biomass_USA.xml}.
The corresponding file in the original data system was \code{batch_regional_biomass_USA.xml} (gcamusa XML batch).
}
\description{
Construct XML data structure for \code{regional_biomass_USA.xml}.
}
