% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA154.Transport_India.R
\name{module_india_LA154.Transport}
\alias{module_india_LA154.Transport}
\title{module_india_LA154.Transport}
\usage{
module_india_LA154.Transport(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L154.india_in_EJ_trn_India_F}, \code{L154.in_EJ_trn_India_F}. The corresponding file in the
original data system was \code{LA154.Transport.R} (gcam-usa level1).
}
\description{
Change the data structure for transport sector
}
\details{
Transportation energy data was downscaled in proportion to GHG Platform state-level transportation energy data

Transportation nonmotor data was downscaled in proportion to state population
}
\author{
Malyan_Ankur_CEEW @updated Rishabh_Varma_CEEW 2025
}
