% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmodeltime_L200.modeltime.R
\name{module_modeltime_L200.modeltime}
\alias{module_modeltime_L200.modeltime}
\title{module_modeltime_L200.modeltime}
\usage{
module_modeltime_L200.modeltime(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L200.ModelTime}, \code{L200.ModelTimeInterYears}, \code{L200.MAGICC}, \code{L200.hector}. The corresponding file in the
original data system was \code{L200.modeltime.R} (modeltime level2).
}
\description{
Generate the model time information needed to interact with Hector or MAGICC.
}
\details{
Generate the model time information needed to interact with Hector or MAGICC.
}
\author{
BBL
}
