% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzPOLICY_L300.elasticity.R
\name{module_policy_L300.elasticity}
\alias{module_policy_L300.elasticity}
\title{module_policy_L300.elasticity}
\usage{
module_policy_L300.elasticity(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L344.bld_shell}
}
\description{
Produce new final energy elasticities
}
\author{
RLH April 20123
}
