% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzPOLICY_L301.ceilings_floors.R
\name{module_policy_L301.ceilings_floors}
\alias{module_policy_L301.ceilings_floors}
\title{module_policy_L301.ceilings_floors}
\usage{
module_policy_L301.ceilings_floors(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L301.policy_port_stnd}, \code{L301.policy_RES_coefs},
\code{L301.RES_secout}, \code{L301.input_tax}, \code{L301.input_subsidy}
}
\description{
Produce ceilings/floors and markets by region
}
\author{
RLH April 2023
}
