% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzPOLICY_L302.inputtaxsubsidy.R
\name{module_policy_L302.inputtaxsubsidy}
\alias{module_policy_L302.inputtaxsubsidy}
\title{module_policy_L302.inputtaxsubsidy}
\usage{
module_policy_L302.inputtaxsubsidy(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L302.InputTax}, \code{L302.InputSubsidy}
}
\description{
Produce fuel standards coefficients
}
\author{
RLH April 2023
}
