% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzPOLICY_L303.shareweight_overwrite.R
\name{module_policy_L303.shareweight_overwrite}
\alias{module_policy_L303.shareweight_overwrite}
\title{module_policy_L303.shareweight_overwrite}
\usage{
module_policy_L303.shareweight_overwrite(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L303.shareweight_overwrite_subsector}, \code{L303.shareweight_overwrite_stubtech}
}
\description{
Shareweights to overwrite
}
\author{
RLH April 2023
}
