% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzPOLICY_L304.Coef_Eff.R
\name{module_policy_L304.Coef_Eff}
\alias{module_policy_L304.Coef_Eff}
\title{module_policy_L304.Coef_Eff}
\usage{
module_policy_L304.Coef_Eff(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L304.StubTechCoef}, \code{L304.StubTechEff}
}
\description{
Produce tech efficiency/coefficients
}
\author{
RLH April 2023
}
