% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzPOLICY_L305.FixedOutputTech.R
\name{module_policy_L305.FixedOutputTech}
\alias{module_policy_L305.FixedOutputTech}
\title{module_policy_L305.FixedOutputTech}
\usage{
module_policy_L305.FixedOutputTech(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L305.StubTechFixedOutput},\code{L305.StubTechLifetime}, \code{L305.GlbTechFixedOutput}
}
\description{
Produce fuel standards coefficients
}
\author{
RLH April 2023
}
