% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzPOLICY_L306.MAC_rename.R
\name{module_policy_L306.MAC_rename}
\alias{module_policy_L306.MAC_rename}
\title{module_policy_L306.MAC_rename}
\usage{
module_policy_L306.MAC_rename(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L306.ResMAC_fos},\code{L252.MAC_higwp},\code{L306.AgMAC},\code{L306.MAC_an},\code{L306.ResMAC_fos_tc_average},\code{L306.ResMAC_fos_phaseInTime},\code{L306.MAC_higwp_tc_average},\code{L306.MAC_higwp_phaseInTime}
}
\description{
Produce fuel standards coefficients
}
\author{
RLH April 2023
}
