% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzPOLICY_L307.exoShutdownScalar.R
\name{module_policy_L307.exoShutdownScalar}
\alias{module_policy_L307.exoShutdownScalar}
\title{module_policy_L307.exoShutdownScalar}
\usage{
module_policy_L307.exoShutdownScalar(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L307.exoShutdownScalar}
}
\description{
Exogenous shutdown decider scalars
}
\author{
RLH April 2023
}
