% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzPOLICY_L3221.CCap.R
\name{module_policy_L3221.CCap}
\alias{module_policy_L3221.CCap}
\title{module_policy_L3221.CCap}
\usage{
module_policy_L3221.CCap(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L3221.CCap_constraint}, \code{L3221.CCap_link_regions}, \code{L3221.CCap_tech}, \code{L3221.CCap_tranTech}, \code{L3221.CCap_resource}
}
\description{
Produce custom carbon cap targets and markets
}
\author{
RLH April 2023
}
