% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzPOLICY_L3222.CTax.R
\name{module_policy_L3222.CTax}
\alias{module_policy_L3222.CTax}
\title{module_policy_L3222.CTax}
\usage{
module_policy_L3222.CTax(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L3222.CTax}, \code{L3222.CTax_Link}
}
\description{
Produce custom carbon cap targets and markets
}
\author{
RLH April 2023
}
