% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzPOLICY_L3231.land_constraints.R
\name{module_policy_L3231.land_constraints}
\alias{module_policy_L3231.land_constraints}
\title{module_policy_L3231.land_constraints}
\usage{
module_policy_L3231.land_constraints(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L3231.landConstrain_mngd_nodes}, \code{L3231.landConstrain_unmngd_nodes}, \code{L3231.landConstrain_constraint}
}
\description{
Produce land constraints for specified countries and years
}
\author{
RLH April 2023
}
