% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzPOLICY_L326.aeei.R
\name{module_policy_L326.aeei}
\alias{module_policy_L326.aeei}
\title{module_policy_L326.aeei}
\usage{
module_policy_L326.aeei(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L326.aeei}
}
\description{
Produce aeei (autonomous energy efficiency improvement index) parameters by region and sector
}
\author{
RLH April 20123
}
