% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzPOLICY_L354.FuelStandard.R
\name{module_policy_L354.FuelStandards}
\alias{module_policy_L354.FuelStandards}
\title{module_policy_L354.FuelStandards}
\usage{
module_policy_L354.FuelStandards(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L354.FuelStandards}
}
\description{
Produce fuel standards coefficients
}
\author{
RLH April 2023
}
