% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzPOLICY_xml_policy_shareweight_overwrite.R
\name{module_policy_shareweight_overwrite_xml}
\alias{module_policy_shareweight_overwrite_xml}
\title{module_policy_shareweight_overwrite_xml}
\usage{
module_policy_shareweight_overwrite_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{policy_shareweight_overwrite.xml}.
}
\description{
Construct XML data structure for \code{policy_shareweight_overwrite.xml}.
}
