% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_xml_bld_agg.R
\name{module_socio_bld_agg_xml}
\alias{module_socio_bld_agg_xml}
\title{module_socio_bld_agg_xml}
\usage{
module_socio_bld_agg_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{bld_agg.xml}. The corresponding file in the
original data system was \code{batch_bld_agg.xml.R} (socio XML).
}
\description{
Construct XML data structure for \code{bld_agg.xml} and all the SSP and gSSP ones as well.
}
