% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_data_Maddison_population.R
\name{module_socio_data_Maddison_population}
\alias{module_socio_data_Maddison_population}
\title{module_socio_data_Maddison_population}
\usage{
module_socio_data_Maddison_population(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{Maddison_population}.
}
\description{
Dedicated data chunk to read \code{Maddison_population.csv} file.
}
\details{
The \code{Maddison_population.csv} file has a blank column, which
generates a warning message in the general purpose \code{\link{load_csv_files}}
that we'd rather not suppress universally. We also need to skip a goofy initial couple of lines.
}
\author{
BBL
}
