% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_xml_socioeconomics_IncomeElasticity.R
\name{module_socio_IncomeElasticity_xml}
\alias{module_socio_IncomeElasticity_xml}
\title{module_socio_bld_agg_xml}
\usage{
module_socio_IncomeElasticity_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{socioeconomics_incelas_SSP[1-5].xml}, \code{bld_agg_SSP[1-5].xml}.
}
\description{
Construct XML data structure for \code{bld_agg.xml} and all the SSP ones as well.
}
