% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L100.GDP_hist.R
\name{module_socio_L100.GDP_hist}
\alias{module_socio_L100.GDP_hist}
\title{module_socio_L100.GDP_hist}
\usage{
module_socio_L100.GDP_hist(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.gdp_mil90usd_ctry_Yh}.
}
\description{
preprocessing historical GDP data from gcamfaostat
}
\details{
preprocessing historical GDP data from gcamfaostat
}
\author{
XZ 2024
}
