% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L100.GTAP.R
\name{module_socio_L100.GTAP}
\alias{module_socio_L100.GTAP}
\title{module_socio_L100.GTAP}
\usage{
module_socio_L100.GTAP(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.GTAP_capital_stock}, \code{L100.GTAPCostShare_AgLU_reg_comm}.
}
\description{
preprocess GTAP data
}
\details{
Select national accounts data from Penn World Table for all countries.
}
\author{
xz 2025
}
