% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L100.NationalAccounts.R
\name{module_socio_L100.NationalAccounts}
\alias{module_socio_L100.NationalAccounts}
\title{module_socio_L100.NationalAccounts}
\usage{
module_socio_L100.NationalAccounts(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.PWT_nationalAccounts}.
}
\description{
Preprocess PWT data for National accounts.
}
\details{
Select national accounts data from Penn World Table for all countries.
}
\author{
SHK October 2020 xz 2025
}
