% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L100.Population_downscale_ctry.R
\name{module_socio_L100.Population_downscale_ctry}
\alias{module_socio_L100.Population_downscale_ctry}
\title{module_socio_L100.Population_downscale_ctry}
\usage{
module_socio_L100.Population_downscale_ctry(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.Pop_thous_ctry_Yh}, \code{L100.Pop_thous_SSP_ctry_Yfut}. The corresponding file in the
original data system was \code{L100.Population_downscale_ctry.R} (socioeconomics level1).
}
\description{
Clean and interpolate both Maddison historical population data (1700-2010) and SSP population scenarios.
}
\details{
(1) Cleans Maddison historical population data and interpolates to country and year (1700-2010). (2) Cleans SSP population scenarios for smooth join with final base year population.
}
\author{
STW May 2017
}
