% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L100.Population_hist.R
\name{module_socio_L100.Population_hist}
\alias{module_socio_L100.Population_hist}
\title{module_socio_L100.Population_hist}
\usage{
module_socio_L100.Population_hist(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.Pop_thous_ctry_Yh}
}
\description{
Clean and interpolate both Maddison historical population data (1700-max(UN_HISTORICAL_YEARS))
}
\details{
(1) Cleans Maddison historical population data and interpolates to country and year (1700-2010).
}
\author{
STW May 2017 XZ 2024
}
