% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L100.SSP_database.R
\name{module_socio_L100.SSP_database}
\alias{module_socio_L100.SSP_database}
\title{module_socio_L100.SSP_database}
\usage{
module_socio_L100.SSP_database(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.Pop_thous_SSP_ctry_Yfut_raw},
 \code{L100.LaborForce_mil_SSP_ctry_Yfut_raw}, \code{L100.GDP_bilusd_SSP_ctry_Yfut_raw}
}
\description{
preprocess SSP database for population, GDP, and labor force
}
\details{
preprocess SSP database for population, GDP, and labor force
}
\author{
XZ 2025
}
