% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L101.Population.R
\name{module_socio_L101.Population}
\alias{module_socio_L101.Population}
\title{module_socio_L101.Population}
\usage{
module_socio_L101.Population(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.Pop_thous_ctry_Yh}, \code{L100.Pop_thous_SSP_ctry_Yfut},
\code{L101.Pop_thous_R_Yh}, \code{L101.Pop_thous_SSP_R_Yfut}
}
\description{
Stitch historical and future (by SSP) population
}
\details{
Connect future growth to historical population data and aggregate to GCAM regions
}
\author{
STW May 2017 XZ 2024
}
