% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L102.GDP.R
\name{module_socio_L102.GDP}
\alias{module_socio_L102.GDP}
\title{module_socio_L102.GDP}
\usage{
module_socio_L102.GDP(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L102.gdp_mil90usd_Scen_R_Y},
\code{L102.pcgdp_thous90USD_Scen_R_Y}, \code{L102.pcgdp_thous90USD_ctry_Yh},
\code{L102.PPP_MER_R}.
}
\description{
Prepare historical and future GDP time series.  On the historical side, this
amounts to aggregating country-level GDP to GCAM regions.  On the future side
we create a time series for each of a variety of future scenarios.  The
outputs include GDP, pcGDP, and the PPP-MER conversion factor, all tabulated
by GCAM region.
}
\details{
The scenarios generated include the SSPs and the gSSPs (SSPs modified by
near-term IMF projections).  GDP outputs are in millions of 1990 USD, Market
Exchange Rate (measured in 2010) is used for foreign currency.  Per-capita
values are in thousands of 1990 USD.
}
\author{
RPL March 2017
}
