% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L103.NationalAccounts.R
\name{module_socio_L103.NationalAccounts}
\alias{module_socio_L103.NationalAccounts}
\title{module_socio_L103.NationalAccounts}
\usage{
module_socio_L103.NationalAccounts(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L103.National_Accounts_mil90usd_R_Yh}, \code{L103.LaborForceShare_Scen_R_Y}.
There is no corresponding file in the original data system.
}
\description{
National accounts information for GDP macro.
}
\details{
Select national accounts data from Penn World Table for all countries.
}
\author{
SHK October 2020 XZ June 2025
}
