% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L106.income_distributions.R
\name{module_socio_L106.income_distributions}
\alias{module_socio_L106.income_distributions}
\title{module_socio_L106.income_distributions}
\usage{
module_socio_L106.income_distributions(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}.
}
\description{
Read in raw income distribution data and transform into inputs other chunks can use
}
\author{
EL August 2022
}
