% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L201.Pop_GDP_scenarios.R
\name{module_socio_L201.Pop_GDP_scenarios}
\alias{module_socio_L201.Pop_GDP_scenarios}
\title{module_socio_L201.Pop_GDP_scenarios}
\usage{
module_socio_L201.Pop_GDP_scenarios(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L201.GDP_Scen}, \code{L201.Pop_Scen},
\code{L201.TotalFactorProductivity_Scen}, \code{L201.PPPConvert}
}
\description{
Labor productivity and population by scenario and region.
}
\details{
Produces default interest rate by region, historical and future population by region and SSP scenario,
and uses per-capita GDP to calculate labor productivity by region and scenario.
}
\author{
HM & RH June 2017
}
