% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L203.NationalAccounts.R
\name{module_socio_L203.NationalAccounts}
\alias{module_socio_L203.NationalAccounts}
\title{module_socio_L203.NationalAccounts}
\usage{
module_socio_L203.NationalAccounts(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L203.NationalAccounts}, \code{L203.SavingsRateParams},
\code{L203.GDP_macro_function}, \code{L203.FactorProductivity}
}
\description{
National accounts information for GDP macro.
}
\details{
National accounts data and GDP macro function parameters for GCAM regions.
}
\author{
SHK 2020 PP 2023 XZ 2025
}
