% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L2323.iron_steel_Inc_Elas_scenarios.R
\name{module_socio_L2323.iron_steel_Inc_Elas_scenarios}
\alias{module_socio_L2323.iron_steel_Inc_Elas_scenarios}
\title{module_socio_L2323.iron_steel_Inc_Elas_scenarios}
\usage{
module_socio_L2323.iron_steel_Inc_Elas_scenarios(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2323.IncomeElasticity_iron_steel_Scen}.
}
\description{
Calculates iron and steel income elasticity for each GCAM region by linear interpolation of assumption data
}
\details{
Takes per-capita GDP from ssp scenarios in each region.
Then calculates iron_steel income elasticity for each region by linear interpolation of assumption data.
}
\author{
RH April 2017
}
