% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L252.Trn_Inc_Elas_scenarios.R
\name{module_socio_L252.Trn_Inc_Elas_scenarios}
\alias{module_socio_L252.Trn_Inc_Elas_scenarios}
\title{module_socio_L252.Trn_Inc_Elas_scenarios}
\usage{
module_socio_L252.Trn_Inc_Elas_scenarios(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L252.IncomeElasticity_trn_GCAM3}, \code{object}. The corresponding file in the
original data system was \code{L252.Trn_Inc_Elas_scenarios.R} (socioeconomics level2).
}
\description{
Calculates transportation income elasticity for SSP scenarios using linear interpolation of assumption data
}
\details{
Uses per-capita GDP data for SSP scenarios to generate transportation income elasticity using linear interpolation of assumption data.
}
\author{
RH May 2017
}
