% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L280.GDP_macro.R
\name{module_socio_L280.GDP_macro}
\alias{module_socio_L280.GDP_macro}
\title{module_socio_L280.GDP_macro}
\usage{
module_socio_L280.GDP_macro(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L280.nationalAccounts}, \code{L280.SavingsRateParams}, \code{L280.GDP_macro_function}, \code{L280.FactorProductivity}.
There is no corresponding file in the original data system.
}
\description{
National accounts information for GDP macro.
}
\details{
National accounts data and GDP macro function parameters for GCAM regions.
}
\author{
SHK October 2020
}
