% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L281.macro_account_tracking.R
\name{module_socio_L281.macro_account_tracking}
\alias{module_socio_L281.macro_account_tracking}
\title{module_socio_L281.macro_account_tracking}
\usage{
module_socio_L281.macro_account_tracking(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L281.BasePriceSectorMapping}, \code{L281.GlobalTechAccountOutputUseBasePrice_fd}, \code{L281.TrialValueResource}
\code{L281.TechAccountOutput_entrade}, \code{L281.TechAccountInput_entrade}, \code{L281.TechAccountInput_NG_entrade}, \code{L281.GlobalTechAccountInput_entrade}.
\code{L280.factor_input_productivity}.
There is no corresponding file in the original data system.
}
\description{
Generate tibbles to add "hooks" into GCAM to be able to track certain accounts such as
final energy service and net energy trade.  In addition generate the trial value market/resources
for those accounts as well as the capital investment tracking markets.
}
\details{
National accounts data and GDP macro function parameters for GCAM regions.
}
\author{
SHK October 2020
}
