% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L110.ADAGE_socioeconomics.R
\name{module_socioeconomics_L110.ADAGE_socioeconomics}
\alias{module_socioeconomics_L110.ADAGE_socioeconomics}
\title{module_socioeconomics_L110.ADAGE_socioeconomics}
\usage{
module_socioeconomics_L110.ADAGE_socioeconomics(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L110.Pop_thous_R_Y_ADAGE},
  \code{L110.pcgdp_thoususd_R_Y_ADAGE}.
}
\description{
Future population and GDP by GCAM region, consistent with ADAGE model assumptions.
}
\author{
RPL March 2017
}
