% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2324.Off_road_Inc_Elas_scenarios.R
\name{module_socioeconomics_L2324.Off_road_Inc_Elas_scenarios}
\alias{module_socioeconomics_L2324.Off_road_Inc_Elas_scenarios}
\title{module_socioeconomics_L2324.Off_road_Inc_Elas_scenarios}
\usage{
module_socioeconomics_L2324.Off_road_Inc_Elas_scenarios(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2324.Off_road_incelas_gcam3}, \code{object}. The corresponding file in the
original data system was \code{L2324.Off_road_Inc_Elas_scenarios.R} (socioeconomics level2).
}
\description{
Calculates Off_road income elasticity for each GCAM region by linear interpolation of assumption data
}
\details{
Takes per-capita GDP from ssp scenarios in each region.
Then calculates Off_road income elasticity for each region by linear interpolation of assumption data.
}
\author{
RH April 2017
}
