% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_xml_socioeconomics_macro.R
\name{module_socioeconomics_macro_xml}
\alias{module_socioeconomics_macro_xml}
\title{module_socio_macro_xml}
\usage{
module_socioeconomics_macro_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{socioeconomics_macro.xml}. There is no corresponding file in the
original data system.
}
\description{
Construct XML data structure for \code{socioeconomics_macro.xml}.
}
