% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_desalination_xml.R
\name{module_water_batch_desalination_xml}
\alias{module_water_batch_desalination_xml}
\title{module_water_batch_desalination_xml}
\usage{
module_water_batch_desalination_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{desalination.xml}.
}
\description{
Construct XML data structure for \code{desalination.xml}.
}
