% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_EFW_WWtrt_coefs_SSP_xml.R
\name{module_water_batch_EFW_WWtrt_coefs_SSP_xml}
\alias{module_water_batch_EFW_WWtrt_coefs_SSP_xml}
\title{module_water_batch_EFW_WWtrt_coefs_SSP_xml}
\usage{
module_water_batch_EFW_WWtrt_coefs_SSP_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs:\code{EFW_WWtrt_coefs_SSP1.xml},
\code{EFW_WWtrt_coefs_SSP2.xml}, \code{EFW_WWtrt_coefs_SSP3.xml},
\code{EFW_WWtrt_coefs_SSP4.xml}, and \code{EFW_WWtrt_coefs_SSP5.xml}.
}
\description{
Construct XML data structure for all the \code{EFW_WWtrt_coefs_[g]SSP[1-5].xml} files.
}
