% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_batch_electricity_water_nointerp_xml.R
\name{module_water_batch_electricity_water_nointerp.xml}
\alias{module_water_batch_electricity_water_nointerp.xml}
\title{module_water_batch_electricity_water_nointerp.xml}
\usage{
module_water_batch_electricity_water_nointerp.xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{electricity_water_nointerp.xml}. The corresponding file in the
original data system was \code{batch_electricity_water.xml.R} (water XML).
}
\description{
Construct XML data structure for \code{electricity_water_nointerp.xml}.
}
