% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L100.water.supply.runoff.R
\name{module_water_L100.water.supply.runoff}
\alias{module_water_L100.water.supply.runoff}
\title{module_water_L100.water.supply.runoff}
\usage{
module_water_L100.water.supply.runoff(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.runoff_max_bm3}, \code{L100.runoff_accessible}. The corresponding file in the
original data system was \code{L100.water.supply.runoff.R} (Water level1).
}
\description{
Prepare GCAM basin runoff and accessible water fractions using Xanthos output.
}
\details{
Reads Xanthos outputs and converts to maximum and accessible water for all GCAM model years.
}
\author{
ST September 2018
}
