% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zwater_L102.water_supply_unlimited.R
\name{module_water_L102.water_supply_unlimited}
\alias{module_water_L102.water_supply_unlimited}
\title{module_water_L102.water_supply_unlimited}
\usage{
module_water_L102.water_supply_unlimited(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L102.unlimited_mapped_water_price_B_W_Y_75USDm3},
\code{L102.unlimited_nonmapped_water_price_R_W_Y_75USDm3}. The corresponding file in the
original data system was \code{L102.water_supply_unlimited.R} (water level1).
}
\description{
Build out unlimited water prices using defaults in Constants.R.
}
\details{
Generates unlimited water supply for all combinations of basin / region and water_type.
}
\author{
ST Oct 2018
}
