% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L132.water.demand.manufacturing.R
\name{module_water_L132.water.demand.manufacturing}
\alias{module_water_L132.water.demand.manufacturing}
\title{module_water_L132.water.demand.manufacturing}
\usage{
module_water_L132.water.demand.manufacturing(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L132.water_coef_manufacturing_R_W_m3_GJ}. The corresponding file in the
original data system was \code{L132.water.demand.manufacturing.R} (water level1).
}
\description{
Computes manufacturing water withdrawals and consumption by nation/region and historical year
}
\details{
Computes manufacturing energy use coefficients for water withdrawal
and consumption for all regions.
}
\author{
GPK June 2018
}
