% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L171.desalination.R
\name{module_water_L171.desalination}
\alias{module_water_L171.desalination}
\title{module_water_L171.desalination}
\usage{
module_water_L171.desalination(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L171.out_km3_R_desalfromelec_Yh},
  \code{L171.share_R_desal_basin}, \code{L171.out_km3_R_desal_F_tech_Yh}, \code{L171.in_km3_ctry_desal_Yh},
  \code{L171.in_EJ_R_desal_F_Yh}.
}
\description{
Desalinated water production by region, basin, and technology
}
\details{
Generate estimates of desalinated water production for municipal and industrial purposes, by various cuts:
  by country, by GCAM region, by technology, by basin, and also estimate the desalinated water production at combined
  electric power and desalination facilities.
}
\author{
GPK January 2019
}
