% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L172.EFW_irrigation.R
\name{module_water_L172.EFW_irrigation}
\alias{module_water_L172.EFW_irrigation}
\title{module_water_L172.EFW_irrigation}
\usage{
module_water_L172.EFW_irrigation(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L172.Coef_GJm3_IrrEnergy_R},
  \code{L172.in_EJ_R_irr_F_Yh}.
}
\description{
Energy requirements for irrigation water abstraction, by GCAM region
}
\details{
Generate estimates of energy consumption for irrigation water abstraction (i.e., pumping and conveyance),
  including energy input-output coefficients, by GCAM region, fuel, and historical year.
}
\author{
GPK January 2019
}
