% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L173.EFW_manufacturing.R
\name{module_water_L173.EFW_manufacturing}
\alias{module_water_L173.EFW_manufacturing}
\title{module_water_L173.EFW_manufacturing}
\usage{
module_water_L173.EFW_manufacturing(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L173.water_km3_R_indEFW_Yh},
  \code{L173.trtshr_ctry_Yh}, \code{L173.in_desal_km3_ctry_ind_Yh}, \code{L173.in_desal_km3_ctry_muni_Yh},
  \code{L173.in_EJ_R_indEFWtot_F_Yh}, \code{L173.IO_GJkm3_R_indEFW_F_Yh},  \code{L173.WWtrtfrac_R_ind_Yh}.
}
\description{
Energy requirements for manufacturing water use
}
\details{
Generate estimates of energy consumption, energy input-output coefficients, and water flow volumes for
  manufacturing water abstraction, treatment, and wastewater treatment, by GCAM region, fuel, and historical year.
}
\author{
GPK January 2019
}
