% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L174.EFW_municipal.R
\name{module_water_L174.EFW_municipal}
\alias{module_water_L174.EFW_municipal}
\title{module_water_L174.EFW_municipal}
\usage{
module_water_L174.EFW_municipal(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L174.water_km3_R_muniEFW_Yh},
  \code{L174.in_EJ_ctry_muniEFW_F_Yh}, \code{L174.in_EJ_R_muniEFWtot_F_Yh}, \code{L174.IO_GJkm3_R_muniEFW_F_Yh},
  \code{L174.WWtrtfrac_R_muni_Yh}.
}
\description{
Energy requirements for municipal water use
}
\details{
Generate estimates of energy consumption, energy input-output coefficients, and water flow volumes for
  municipal water abstraction, treatment, distribution, and wastewater treatment, by GCAM region, fuel, and
  historical year.
}
\author{
GPK January 2019
}
