% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zwater_L202.water_resources_unlimited.R
\name{module_water_L202.water_resources_unlimited}
\alias{module_water_L202.water_resources_unlimited}
\title{module_water_L202.water_resources_unlimited}
\usage{
module_water_L202.water_resources_unlimited(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L202.UnlimitRsrc_mapped}, \code{L202.UnlimitRsrc_nonmapped},
\code{L202.UnlimitRsrcPrice_mapped}, \code{L202.UnlimitRsrcPrice_nonmapped}. The corresponding file in the
original data system was \code{L202.water_resources_unlimited.R} (water level2).
}
\description{
Create unlimited resource markets for water types, and read in fixed prices for water types.
}
\details{
Create unlimited resource markets (i.e., 32 GCAM regions) for water types (i.e., water consumption, withdrawals, biophysical water consumption and seawater),
and read in fixed prices for water types.
}
\author{
YL July 2017 / ST Oct 2018
}
