% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L232.water.demand.manufacturing.R
\name{module_water_L232.water.demand.manufacturing}
\alias{module_water_L232.water.demand.manufacturing}
\title{module_water_L232.water.demand.manufacturing}
\usage{
module_water_L232.water.demand.manufacturing(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L232.TechCoef}. The corresponding file in the
original data system was \code{L232.water.demand.manufacturing.R} (water level2).
}
\description{
Computes manufacturing water withdrawal/consumption coefficients (m3/GJ output) by region and year
}
\details{
Describe in detail what this chunk does.
}
\author{
GPK June 2018
}
