% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zwater_L233.water_demand_livestock.R
\name{module_water_L233.water_demand_livestock}
\alias{module_water_L233.water_demand_livestock}
\title{module_water_L233.water_demand_livestock}
\usage{
module_water_L233.water_demand_livestock(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L233.TechCoef}. The corresponding file in the
original data system was \code{L233.water_demand_livestock.R} (water level2).
}
\description{
Generates water coefficients for region-specific livestock for model years.
}
\details{
This chunk generates water coefficients (amount of water needed to produce one unit of livestock, unit: m^3/Mt)
for region-specific livestock for#' model years (1975, 1990,2005, 2010....2100), with the information of supplysector,
subsector,technology, and energy input.
}
\author{
YL July 2017
}
