% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L272.EFW_irrigation.R
\name{module_water_L272.EFW_irrigation}
\alias{module_water_L272.EFW_irrigation}
\title{module_water_L272.EFW_irrigation}
\usage{
module_water_L272.EFW_irrigation(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{L272.Supplysector_irr},
  \code{L272.FinalEnergyKeyword_irr}, \code{L272.SubsectorLogit_irr}, \code{L272.SubsectorShrwtFllt_irr},
  \code{L272.SubsectorInterpTo_irr}, \code{L272.StubTech_irr}, \code{L272.GlobalTechCoef_irr},
  \code{L272.GlobalTechShrwt_irr}, \code{L272.StubTechCoef_irr}. This did not
  exist in the original data system.
}
\description{
Generate irrigation energy-for-water input file
}
\details{
This chunk translates input assumptions from CSV to XML, and calibrates the input-output coefficients by
  technology in each region and model base year
}
\author{
GPK January 2019
}
