% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zwater_xml_unlimited_water_supply.R
\name{module_water_unlimited_water_supply_xml}
\alias{module_water_unlimited_water_supply_xml}
\title{module_water_unlimited_water_supply_xml}
\usage{
module_water_unlimited_water_supply_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{unlimited_water_supply.xml}. The corresponding file in the
original data system was \code{batch_unlimited_water_supply.xml.R} (water XML).
}
\description{
Construct XML data structure for \code{unlimited_water_supply.xml}.
}
