% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zwater_xml_water_demand_municipal.R
\name{module_water_water_demand_municipal_xml}
\alias{module_water_water_demand_municipal_xml}
\title{module_water_water_demand_municipal_xml}
\usage{
module_water_water_demand_municipal_xml(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{water_demand_municipal.xml}. The corresponding file in the
original data system was \code{batch_water_demand_municipal.xml.R} (water XML).
}
\description{
Construct XML data structure for \code{water_demand_municipal.xml}.
}
