% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{Moving_average}
\alias{Moving_average}
\title{Moving average}
\usage{
Moving_average(x, periods = 5)
}
\arguments{
\item{x}{A data frame contain the variable for calculation}

\item{periods}{An odd number of the periods in MA. The default is 5, i.e., 2 lags and 2 leads}
}
\value{
A data frame
}
\description{
function to calculate moving average
}
