% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{Moving_average_lagged}
\alias{Moving_average_lagged}
\title{Moving average lagged}
\usage{
Moving_average_lagged(x, periods = 5)
}
\arguments{
\item{x}{A data frame contain the variable for calculation}

\item{periods}{An odd number of the periods in MA. The default is 5, i.e.,
4 preceding years and the current year}
}
\value{
A data frame with moving average
}
\description{
function to calculate moving average using only preceding values
}
\author{
Hassan Niazi, May 2024
}
