% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oldnew_compare}
\alias{oldnew_compare}
\title{oldnew_compare}
\usage{
oldnew_compare(x)
}
\arguments{
\item{x}{Name of an object in the data system, character}
}
\value{
Merged data from old and new data system for this object, if possible, or a
list with the datasets, if not.
}
\description{
Load and merge (if possible) an old-vs-new comparison for a data system object.
}
\note{
This is a development/debugging routine. It's highly dependent on a specific file layout.
}
\examples{
\dontrun{
oldnew_compare("L100.gdp_mil90usd_ctry_Yh.csv")
oldnew_compare("L133.ag_Cost_75USDkg_C")
}
}
