% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{policy_interpolate}
\alias{policy_interpolate}
\title{policy_interpolate}
\usage{
policy_interpolate(df, group_cols, value_col, year_col = year)
}
\arguments{
\item{df}{data}

\item{group_cols}{columns to group, as a list of strings}

\item{value_col}{column to interpolate, unquoted}
}
\value{
assignment of new db
}
\description{
Helper function to interpolate policy constraints/values between years
}
