% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{remove_regions_data_tables}
\alias{remove_regions_data_tables}
\title{remove_regions_data_tables}
\usage{
remove_regions_data_tables(df_list, regions_to_remove)
}
\arguments{
\item{df_list}{list with data and header}

\item{regions_to_remove}{character vector of regions to remove}
}
\value{
data object without specified regions
}
\description{
Helper function to remove regions from a tibble
}
