% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{remove_regions_xml}
\alias{remove_regions_xml}
\title{remove_regions_xml}
\usage{
remove_regions_xml(xml, regions_to_remove, inverse = FALSE)
}
\arguments{
\item{xml}{xml object that is produced at end of xml chunl}

\item{regions_to_remove}{character vector of regions to remove}

\item{inverse}{if TRUE, remove all but the specified regions}
}
\value{
xml object without specified regions in each tibble
}
\description{
Helper function to remove regions from an xml object
}
