% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{rename_gcamdata_chunks}
\alias{rename_gcamdata_chunks}
\title{rename_gcamdata_chunks}
\usage{
rename_gcamdata_chunks()
}
\value{
Nothing
}
\description{
Renames gcamdata chunks to new format, including getting rid of LA/LBs, cleaning up module
names, and replacing "batch" with "xml". Also renames function within scripts
}
\author{
ENL 2023
}
