% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{replace_outlier_EFs}
\alias{replace_outlier_EFs}
\title{replace_outlier_EFs}
\usage{
replace_outlier_EFs(df, to_group, names, ef_col_name)
}
\arguments{
\item{df}{Base tibble to start from that contains EFs, and may include NAs}

\item{to_group}{Character vector indicating the column names to group by.
This relates to how specific the median will be, whether it is by sector, sector and subsector, etc.}

\item{names}{Character vector indicating the column names of the returned tibble}

\item{ef_col_name}{Name of the column containing emission factors}
}
\value{
tibble with corresponding region, year, Non.CO2, GCAM sector, subsector, stub.technology, and modified EFs
}
\description{
Helper function to replace emission factors (EFs) outside a threshold with a median EF
Used for emission factors in several sectors
}
