% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{replace_with_eurostat}
\alias{replace_with_eurostat}
\title{replace_with_eurostat}
\usage{
replace_with_eurostat(global_df, eur_df)
}
\arguments{
\item{global_df}{Tibble of global data}

\item{eur_df}{Tibble of european data}
}
\value{
new tibble
}
\description{
Helper function to take a dataset with global data and replace the european regions with eurostat data
}
