% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{resource_reserve_back_calculate}
\alias{resource_reserve_back_calculate}
\title{resource_reserve_back_calculate}
\usage{
resource_reserve_back_calculate(data)
}
\arguments{
\item{data}{A data frame which is already nested for a single region / "technology".  The data
must include a columns for \code{value} with the annual production by \code{year} and
in addition requires a column \code{lifetime} with the assumed production lifetime of this "technology".}
}
\value{
A data.frame with columns \code{year} and \code{value} which are the reserve additions
by year which exactly replicate the given annual production and lifetime assumption.
}
\description{
Back calculate reserve additions to be exactly enough given our historical production
and assumed production lifetime.  Note production lifetimes may not cover the entire
historical period and production may dip below capacity making the calculation a bit more
tricky.  The \code{resource_reserve_back_calculate} replicates the resource / reserve
behavior in GCAM to help project forward production by each historical vintage so we can
accurately back calculate the reserves despite these complications.
}
\section{Warning}{
 Results of this calculation are dependent on users selection for \code{MODEL_BASE_YEARS}
}

