% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{return_multiple_xmls}
\alias{return_multiple_xmls}
\title{Method to return multiple xmls}
\usage{
return_multiple_xmls(list_of_xmls, all_xml_names)
}
\arguments{
\item{all_xml_names}{List of xml names}
}
\value{
A list of xmls that a function can output
}
\description{
Method to return multiple xmls
}
