% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{run_xml_tests}
\alias{run_xml_tests}
\title{Compare all XML files in a directory to their counterparts in the output}
\usage{
run_xml_tests(olddir, newdir = XML_DIR)
}
\arguments{
\item{olddir}{Directory containing the old (reference) output xml files}

\item{newdir}{Directory containing the new output xml files}
}
\value{
Number of discrepancies found.
}
\description{
The 'old' directory should contain reference versions of the files.  The
'new' directory should be the directory where the output xml files are
stored.  The old directory is searched recursively; the new directory is not
(because the new system currently dumps all of its outputs into a single
output directory).
}
