% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-helpers.R
\name{safe_complete}
\alias{safe_complete}
\title{Helper function: call complete only if group is not empty}
\usage{
safe_complete(.data, ...)
}
\value{
The completed data frame.
}
\description{
Helper function: call complete only if group is not empty
}
