% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{screen_undesired}
\alias{screen_undesired}
\title{screen_undesired}
\usage{
screen_undesired(fn)
}
\arguments{
\item{fn}{The function to be tested. This is the actual function object, not
the name of the function.}
}
\value{
Nx2 Character matrix of flagged lines and the test that tripped them
(empty vector, if none)
}
\description{
Screen a function for use of functions undesired, but not forbidden, by data system style guide.
}
\details{
Certain functions are undesired, but not necessarily forbidden. This function tests a function for calls to forbidden functions
and flags the offending lines.
}
\author{
RLH Sept 2022
}
