% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{screen_years}
\alias{screen_years}
\title{screen_years}
\usage{
screen_years(fn)
}
\arguments{
\item{fn}{The function to be tested. This is the actual function object, not
the name of the function.}
}
\value{
Nx2 Character matrix of flagged lines and the test that tripped them
(empty vector, if none)
}
\description{
Screen a function for hardcoded years.
}
\details{
Years should be declared as constants, not hardcoded.
}
\author{
KVC 10 Mar 2026
}
