% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{set_water_input_name}
\alias{set_water_input_name}
\title{set_water_input_name}
\usage{
set_water_input_name(water_sector, water_type, water_mapping,
  GLU = NA_character_)
}
\arguments{
\item{water_sector}{A character vector containing water sector information}

\item{water_type}{A character vector containing water type information}

\item{water_mapping}{A tibble with a mapping between \code{water.sector} and  \code{supplysector}}

\item{GLU}{An optional character vector containing GLU information (only used for irrigation with mapped water types)}
}
\value{
A vector of names of form supplysector_watertype or supplysector_GLU_watertype.
}
\description{
Get the appropriate minicam.energy.input name to use in the GCAM supplysector.
}
\details{
Get the appropriate minicam.energy.input name to use in the GCAM supplysector
by looking up using a mapping to the water.sector and water_type. The minicam.energy.input
name to use will have to be some water mapping sector for water_types that are "mapped".
}
\author{
BBL April 2017
}
