% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{set_xml_file_helper}
\alias{set_xml_file_helper}
\title{set_xml_file_helper}
\usage{
set_xml_file_helper(xml, fq_name)
}
\arguments{
\item{xml}{The xml pipeline object}

\item{fq_name}{The full path to the XML file}
}
\value{
The updated XML object.
}
\description{
set_xml_file_helper
}
