% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{tech_name_expansion}
\alias{tech_name_expansion}
\title{tech_name_expansion}
\usage{
tech_name_expansion(df, sector = "supplysector", tech = "technology", mapping)
}
\arguments{
\item{sector}{sector name to match}

\item{tech}{column to add prefix}

\item{mapping}{segment name mapping data}
}
\value{
new tibble
}
\description{
Helper function to replace normal electricity techs with segment names
}
