% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{verify_identical_prebuilt}
\alias{verify_identical_prebuilt}
\title{verify_identical_prebuilt}
\usage{
verify_identical_prebuilt(..., pb = NULL)
}
\arguments{
\item{...}{The objects}

\item{pb}{\code{PREBUILT_DATA} object; overridden only for testing}
}
\value{
A logical indicating whether a mismatch occurred.
}
\description{
Check whether objects are identical to their prebuilt versions.
}
\note{
Called primarily for its side effects: a warning is issued for each non-identical object.
}
