% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{warn_datachunk_bypass}
\alias{warn_datachunk_bypass}
\title{warn_datachunk_bypass}
\usage{
warn_datachunk_bypass()
}
\value{
Number of data chunk objects being read directly.
}
\description{
Check whether chunks are bypassing a data chunk
}
\details{
Data chunks read one or more particular files that are quirky: typically, files that generate
warnings or errors when read using the default \code{\link{get_data}} routine, but also ones that need
some special preprocessing. We look for chunks that read a file
}
