% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{warn_mismarked_fileinputs}
\alias{warn_mismarked_fileinputs}
\title{warn_mismarked_fileinputs}
\usage{
warn_mismarked_fileinputs()
}
\value{
Number of data chunk objects marked as \code{from_file} but without a directory path.
}
\description{
Look for mislabeled chunk file inputs
}
