% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_csv_from_cache}
\alias{write_csv_from_cache}
\title{write_csv_from_cache}
\usage{
write_csv_from_cache(output, dir = NULL, target_list = NULL, quiet = F)
}
\arguments{
\item{output}{The name of the output to load and write}

\item{dir}{Cache directory (.drake folder)}

\item{quiet}{Print messages?}

\item{target}{List of all objects in cache}
}
\description{
Loads a target from the drake cache and writes it as a csv to the outputs folder
}
