% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{write_to_all_grid_regions}
\alias{write_to_all_grid_regions}
\title{write_to_all_grid_regions}
\usage{
write_to_all_grid_regions(data, names)
}
\arguments{
\item{data}{Base tibble to start from}

\item{names}{Character vector indicating the column names of the returned tibble}
}
\value{
Tibble with data written out to all USA grid regions
}
\description{
write out data to all GCAM-USA grid regions
}
\note{
Used for GCAM-USA when repeating information for each USA electricity grid region
}
