% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{write_to_all_trn_provinces}
\alias{write_to_all_trn_provinces}
\title{write_to_all_trn_provinces}
\usage{
write_to_all_trn_provinces(data, names)
}
\arguments{
\item{data}{Base tibble to start from}

\item{names}{Character vector indicating the column names of the returned tibble}
}
\value{
Tibble with data written out to all China provinces
}
\description{
write out data to all trn provinces (includes HK and MC)
}
\note{
Used for China national data by GCAM region, which is repeated for each China province
}
