% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakout_helpers.R
\name{write_to_breakout_regions}
\alias{write_to_breakout_regions}
\title{write_to_breakout_regions}
\usage{
write_to_breakout_regions(
  data = NULL,
  composite_region = NULL,
  disag_regions = NULL
)
}
\arguments{
\item{data}{Data frame with a "region" column}

\item{composite_region}{Composite (larger) region that data are being copied from (e.g. "Thailand")}

\item{disag_regions}{Disaggregated regions that data are being copied to (e.g. c("Bangkok","Rest of Thailand"))}
}
\value{
Table with both regions.
}
\description{
Copy Thailand data to disaggregated Bangkok and Rest of Thailand regions
}
\details{
This function is designed to take data from a composite (standard) region that is
generic in nature and apply it similarly to disaggregated regions
}
