% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\docType{data}
\name{xml.XML_SUFFIX}
\alias{xml.XML_SUFFIX}
\title{A suffix to be appended to the XML name in \code{set_xml_file_helper}.  Potentially
useful when used to generate permutations of inputs.  We need to go through a package
data so as to do this in a way that is opaque to drake}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
xml.XML_SUFFIX
}
\description{
A suffix to be appended to the XML name in \code{set_xml_file_helper}.  Potentially
useful when used to generate permutations of inputs.  We need to go through a package
data so as to do this in a way that is opaque to drake
}
\keyword{datasets}
