% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{xml_region_filter}
\alias{xml_region_filter}
\title{Filter xml to specified regions only}
\usage{
xml_region_filter(xml, newnm, regions)
}
\arguments{
\item{xml}{XML object}

\item{newnm}{Name for new xml}

\item{regions}{Vector of region names}
}
\value{
Filtered xml
}
\description{
Filter xml to specified regions only
}
