% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{xml_split_by_region}
\alias{xml_split_by_region}
\title{Filter xml to specified regions only}
\usage{
xml_split_by_region(xml, num_splits, GCAM_region_names)
}
\arguments{
\item{xml}{XML object}

\item{num_splits}{Integer number of xmls to split into}

\item{GCAM_region_names}{Dataframe with names of regions and their ID}
}
\value{
Filtered xml
}
\description{
Filter xml to specified regions only
}
