% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{add_carbon_info}
\alias{add_carbon_info}
\title{add_carbon_info}
\usage{
add_carbon_info(data, carbon_info_table, matchvars = c("region", "GLU",
  Cdensity_LT = "Land_Type"))
}
\arguments{
\item{data}{= input data tibble to receive carbon info}

\item{carbon_info_table}{= table with veg and soil carbon densities, and mature.age}

\item{matchvars}{=  a character vector for by = in left_join(data, carbon_info_table, by = ...)}
}
\value{
the original table with carbon density info added
}
\description{
function to translate from soil, veg, and mature age data (already in a table) to the required read-in model parameters
}
