% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_chunks}
\alias{find_chunks}
\title{find_chunks}
\usage{
find_chunks(pattern = "^module_[a-zA-Z\\\\.]*_.*$", include_disabled = FALSE)
}
\arguments{
\item{pattern}{Regular expression pattern to search for}

\item{include_disabled}{Return names of disabled chunks?}
}
\value{
A data frame with fields 'name', 'module', and 'chunk'.
}
\description{
Get a list of chunks in this package.
These are functions with a name of "module_{modulename}_{chunkname}".
}
\details{
If a chunk name ends with \code{_DISABLED}, by default its name
will not be returned.
}
