% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{join.gdp.ts}
\alias{join.gdp.ts}
\title{join.gdp.ts}
\usage{
join.gdp.ts(past, future, grouping)
}
\arguments{
\item{past}{Tibble with the past time series (year, gdp, and grouping).}

\item{future}{Tibble with the future data (year, gdp, scenario, and
grouping).}

\item{grouping}{Name of the grouping column (generally either 'iso' or
'GCAM_region_ID', but could be anything}
}
\value{
Time series with the past and future joined as described in details.
}
\description{
Join past GDP time series to future.
}
\details{
When we have to join two GDP time series, we usually find that they don't
match up at year of overlap (the "base year").  What we do in these cases is
we compute, for the later time series, ratios of GDPs in the future years to
those in the base year.  We then multiply the future ratios by the past base
year value.  That future time series can then be grafted onto the past
without leaving a seam.

In practice, the past is often a single time series, while the future is
often a collection of scenarios.  Therefore, we assume that the past time
series has no scenario column.  If the future does not have a scenario
column, it is given a dummy one, which is dropped before the new table is
returned.  Note that we look for lower-case 'scenario' for this.

The base year is calculated automatically.  It is the maximum of the years
that overlap between the two data sets.

We also have to know how to group the data for calculating the gdp ratios.
Normally this will be either by country ('iso') or by GCAM region
('GCAM_region_ID').  The choice of which is passed in as the 'grouping'
argument.

Finally, although we have discussed this function in terms of joining two GDP
time series, in the future time series we use only the ratios of GDP to base
year GDP.  Therefore, any time series with the correct ratios will work.  For
example, if we have a time series of growth rates, we can convert those to
ratios using \code{\link[base]{cumprod}} and pass those ratios as the future
time series.  For similar reasons, even if the two time series have different
units (e.g., different dollar-years or PPP vs. MER), they can still be
joined.  The units of the output time series will be the same as the units of
\code{past}.
}
