% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_L100.FAO_SUA_PrimaryEquivalent.R
\name{module_aglu_L100.FAO_SUA_PrimaryEquivalent}
\alias{module_aglu_L100.FAO_SUA_PrimaryEquivalent}
\title{module_aglu_L100.FAO_SUA_PrimaryEquivalent}
\usage{
module_aglu_L100.FAO_SUA_PrimaryEquivalent(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs, a vector of output names, or (if
  \code{command} is "MAKE") all the generated outputs: \code{GCAM_AgLU_SUA_APE},
  \code{FAO_AgProd_Kt_All},\code{FAO_AgArea_Kha_All},\code{FAO_Food_Macronutrient_All},
  \code{FAO_Food_MacronutrientRate_MaxValue}
}
\description{
Generate supply utilization balance in primary equivalent
}
\details{
This chunk compiles balanced supply utilization data in primary equivalent in GCAM region and commodities.
A method to generate primary equivalent is created for the new FAOSTAT supply utilization data (2010 to 2019).
New SUA balance is connected to the old one (before 2010). Production and harvested area data with FAO region and item
for primary production are provided. For FAO food items, macronutrient values are calculated at SUA item level.
Data processing was consistent across scales. Note that GCAM regions and commodities in aggregation mapping can
be changed in corresponding mappings. The output data is not averaged over time.
}
\author{
XZ 2022
# In some cases, prescale sink item SUA using output_specific_extraction_rate can improve the processing.
# e.g., when coproduction shares are not fixed.
}
