% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaglu_L202.an_input.R
\name{module_aglu_L202.an_input}
\alias{module_aglu_L202.an_input}
\title{module_aglu_L202.an_input}
\usage{
module_aglu_L202.an_input(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L202.RenewRsrc}, \code{L202.RenewRsrcPrice}, \code{L202.maxSubResource},
\code{L202.RenewRsrcCurves}, \code{L202.ResTechShrwt}, \code{L202.UnlimitedRenewRsrcCurves}, \code{L202.UnlimitedRenewRsrcPrice},
\code{L202.Supplysector_in}, \code{L202.SubsectorAll_in}, \code{L202.StubTech_in}, \code{L202.StubTechInterp_in},
\code{L202.GlobalTechCoef_in}, \code{L202.GlobalTechShrwt_in}, \code{L202.StubTechProd_in},
\code{L202.Supplysector_an}, \code{L202.SubsectorAll_an}, \code{L202.GlobalTechShrwt_an}, \code{L202.StubTechInterp_an}
\code{L202.StubTechProd_an}, \code{L202.StubTechCoef_an}, \code{L202.StubTechCost_an},
. The corresponding file in the
original data system was \code{L202.an_input.R} (aglu level2).
}
\description{
Produce a wide range of animal-related resource tables: production, import, resource curves.
}
\details{
This chunk produces 22 animal-related resource tables: production, import, resource curves.
}
\author{
BBL August 2017
}
