% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L244.building_CHINA.R
\name{module_gcam.china_L244.building_CHINA}
\alias{module_gcam.china_L244.building_CHINA}
\title{module_gcam.china_L244.building_CHINA}
\usage{
module_gcam.china_L244.building_CHINA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L244.DeleteConsumer_CHINAbld}, \code{L244.DeleteSupplysector_CHINAbld}, \code{L244.SubregionalShares_CHINAbld},
\code{L244.PriceExp_IntGains_CHINAbld}, \code{L244.Floorspace_CHINAbld}, \code{L244.DemandFunction_serv_CHINAbld}, \code{L244.DemandFunction_flsp_CHINAbld},
\code{L244.Satiation_flsp_CHINAbld}, \code{L244.SatiationAdder_CHINAbld}, \code{L244.ThermalBaseService_CHINAbld}, \code{L244.GenericBaseService_CHINAbld},
\code{L244.ThermalServiceSatiation_CHINAbld}, \code{L244.GenericServiceSatiation_CHINAbld}, \code{L244.Intgains_scalar_CHINAbld},
\code{L244.ShellConductance_CHINAbld}, \code{L244.Supplysector_CHINAbld}, \code{L244.FinalEnergyKeyword_CHINAbld}, \code{L244.SubsectorShrwt_CHINAbld},
\code{L244.SubsectorShrwtFllt_CHINAbld}, \code{L244.SubsectorInterp_CHINAbld}, \code{L244.SubsectorInterpTo_CHINAbld},
\code{L244.SubsectorLogit_CHINAbld}, \code{L244.StubTech_CHINAbld}, \code{L244.StubTechCalInput_CHINAbld}, \code{L244.StubTechMarket_CHINAbld},
\code{L244.GlobalTechIntGainOutputRatio_CHINAbld}, \code{L244.GlobalTechInterpTo_CHINAbld}, \code{L244.GlobalTechEff_CHINAbld},
\code{L244.GlobalTechShrwt_CHINAbld}, \code{L244.GlobalTechCost_CHINAbld}, \code{L244.GlobalTechSCurve_CHINAbld},
\code{L244.FuelPrefElast_CHINAbld}.
The corresponding file in the original data system was \code{L244.building_USA.R} (gcam-usa level2).
}
\description{
Creates GCAM-CHINA building output files for writing to xml.
}
\details{
Creates GCAM-CHINA building output files for writing to xml.
}
\author{
BY January 2020
}
