% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA101.Energy_Balance_china.R
\name{module_gcam.china_LA101.Energy_Balance}
\alias{module_gcam.china_LA101.Energy_Balance}
\title{module_gcam.china_LA101.Energy_Balance}
\usage{
module_gcam.china_LA101.Energy_Balance(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L101.NBS_use_all_Mtce}, \code{L101.inNBS_Mtce_province_S_F}. The corresponding file in the
original data system was \code{LA101.Energy_Balance.R} (gcam-china level1).
}
\description{
Briefly describe what this chunk does.
}
\details{
See above
}
\author{
Yang July 2018
}
