% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zgcameurope_L210.resources.R
\name{module_gcameurope_L210.resources}
\alias{module_gcameurope_L210.resources}
\title{module_gcameurope_L210.resources}
\usage{
module_gcameurope_L210.resources(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L210.Rsrc_EUR}, \code{L210.RenewRsrc_EUR}, \code{L210.UnlimitRsrc_EUR}, \code{L210.RsrcPrice_EUR}, \code{L210.RenewRsrcPrice_EUR},
\code{L210.UnlimitRsrcPrice_EUR}, \code{L210.RsrcTechChange_EUR}, \code{L210.SmthRenewRsrcTechChange_EUR}, \code{L210.SmthRenewRsrcTechChange_offshore_wind_EUR},
\code{L210.RsrcCalProd_EUR}, \code{L210.ReserveCalReserve_EUR}, \code{L210.RsrcCurves_fos_EUR}, \code{L210.RsrcCurves_U_EUR}, \code{L210.SmthRenewRsrcCurves_MSW_EUR},
\code{L210.SmthRenewRsrcCurves_wind_EUR}, \code{L210.SmthRenewRsrcCurves_offshore_wind_EUR}, \code{L210.SmthRenewRsrcCurvesGdpElast_roofPV_EUR},
\code{L210.GrdRenewRsrcCurves_geo_EUR}, \code{L210.GrdRenewRsrcMax_geo_EUR}, \code{L210.GrdRenewRsrcCurves_EGS_EUR}, \code{L210.GrdRenewRsrcMax_EGS_EUR},
\code{L210.ResSubresourceProdLifetime_EUR}, \code{L210.ResReserveTechLifetime_EUR}, \code{L210.ResReserveTechDeclinePhase_EUR}, \code{L210.ResReserveTechProfitShutdown_EUR},
\code{L210.ResTechShrwt_EUR}, \code{L210.ResTechShrwt_EGS_EUR}.\code{L210.GrdRenewRsrcCurves_tradbio_EUR}.\code{L210.GrdRenewRsrcMax_tradbio_EUR}.
}
\description{
Resource market information, prices, TechChange parameters, supply curves, and environmental costs.
}
\details{
Resource market information, prices, TechChange parameters, supply curves, and environmental costs.
}
\author{
RLH November 2017
}
