% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L223.electricity_USA.R
\name{module_gcamusa_L223.electricity_USA}
\alias{module_gcamusa_L223.electricity_USA}
\title{module_gcamusa_L223.electricity_USA}
\usage{
module_gcamusa_L223.electricity_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L223.Sector_Investment},
\code{L223.SubsectorLogit_Investment}, \code{L223.SubsectorInterp_Investment}, \code{L223.SubsectorInterpTo_Investment},
\code{L223.StubTech_Investment}, \code{L223.GlobalTechEff_Investment}, \code{L223.StubTechMarket_Investment},
\code{L223.GlobalTechOMfixed_Investment},\code{L223.GlobalTechOMvar_Investment}, \code{L223.GlobalTechCapital_Investment},
\code{L223.GlobalTechShrwt_Investment}, \code{L223.GlobalTechCapFac_Investment}, \code{L223.TechCapFac_Investment},
\code{L223.GlobalTechCapture_Investment}, \code{L223.GlobalTechCost_Investment}, \code{L223.GlobalTechHack_Investment},
\code{L223.Sector_Investment_StateShare}, \code{L223.Subsector_Investment_StateShare}, \code{L223.CapacityTech},
\code{L223.SubsectorShrwtFllt_Investment_StateShare}, \code{L223.TechCoef_Investment_StateShare},
\code{L223.TechShrwt_Investment_StateShare},\code{L223.Sector_Investment_LoadCurve}, \code{L223.TechShrwt_Dispatch},
\code{L223.SectorUseTrialMarket_Investment_LoadCurve}, \code{L223.SubsectorLogit_Investment_LoadCurve},
\code{L223.SubsectorShrwtFllt_Investment_LoadCurve}, \code{L223.TechShrwt_Investment_LoadCurve},
\code{L223.TechCoef_Investment_LoadCurve}, \code{L223.TechPMult_Investment_LoadCurve}, \code{L223.TechEff_Dispatch},
\code{L223.TechCost_Investment_LoadCurve}, \code{L223.DispatchSector}, \code{L223.Sector_Dispatch},
\code{L223.SubsectorLogit_Dispatch}, \code{L223.SubsectorShrwtFllt_Dispatch}, \code{L223.CapacityTech_FutureTechs},
\code{L223.TechOMvar_Dispatch}, \code{L223.TechLifetime_Dispatch}, \code{L223.TechSCurve_Dispatch},
\code{L223.TechCapFac_Dispatch}, \code{L223.TechCarbonCapture_Dispatch}, \code{L223.TechCapFac_Cal},
\code{L223.TechEff_Cal}, \code{L223.Sector_Dispatch_Grid}, \code{L223.DispatchSectorCalProd},
\code{L223.DispatchSectorDemandSegments}, \code{L223.InterestRate_FERC}, \code{L223.Pop_FERC}, \code{L223.BaseGDP_FERC},
\code{L223.LaborForceFillout_FERC}.
The corresponding file in the
original data system was \code{L223.electricity_USA.R} (gcam-usa level2 - dispatch branch).
}
\description{
Generates GCAM-USA model inputs for electrcity sector by grid regions and states.
}
\details{
This chunk generates input files to create an annualized electricity generation sector for each state
and creates the demand for the state-level electricity sectors in the grid regions.
}
\author{
YO Feb 2020
}
