% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2237.wind_reeds_USA.R
\name{module_gcamusa_L2237.wind_reeds_USA}
\alias{module_gcamusa_L2237.wind_reeds_USA}
\title{module_gcamusa_L2237.wind_reeds_USA}
\usage{
module_gcamusa_L2237.wind_reeds_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2237.SmthRenewRsrcCurves_wind_reeds_USA}, \code{L2237.StubTechCapFactor_wind_reeds_USA},
\code{L2237.SmthRenewRsrcTechChange_wind_reeds_USA}, \code{L2237.StubTechCost_wind_reeds_USA},
\code{L2237.ResTechShrwt_wind_reeds_USA}.
The corresponding file in the original data system was \code{L2237.wind_reeds_USA.R} (gcam-usa level2).
}
\description{
Create updated wind resource supply curves consistent with ReEDS.
}
\details{
Create state-level wind resource supply curves
}
\author{
MTB September 2018
}
