% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2238.PV_reeds_USA.R
\name{module_gcamusa_L2238.PV_reeds_USA}
\alias{module_gcamusa_L2238.PV_reeds_USA}
\title{module_gcamusa_L2238.PV_reeds_USA}
\usage{
module_gcamusa_L2238.PV_reeds_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2238.DeleteStubTechMinicamEnergyInput_investment_PV_reeds_USA},
\code{L2238.DeleteInput_dispatch_PV_reeds_USA}, \code{L2238.RenewRsrc_PV_reeds_USA},
\code{L2238.GrdRenewRsrcCurves_PV_reeds_USA}, \code{L2238.GrdRenewRsrcMax_PV_reeds_USA},
\code{L2238.StubTechEffFlag_investment_PV_reeds_USA}, \code{L2238.TechEff_Dispatch_PV_reeds_USA},
\code{L2238.RenewRsrcTechChange_PV_reeds_USA}, and \code{L2238.StubTechCost_PV_reeds_USA},
\code{L2238.ResTechShrwt_PV_reeds_USA}, \code{L2238.CapacityTechInputPMult_dispatch_PV_reeds_USA}.
The corresponding file in the original data system was \code{L2238.PV_reeds_USA.R} (gcam-usa level2).
}
\description{
Create updated solar PV resource supply curves consistent with ReEDS.
Also add non-ReEDS states (AK,DC,HI) based on NREL technical potential data.
}
\details{
Create state-level solar PV resource supply curves
}
\author{
MTB September 2018 / AJS June 2019 / YO April 2020
L223.StubTechMarket_Investment',
L223.TechEff_Dispatch',
L223.GlobalTechCapital_Investment',
L223.GlobalIntTechCapital_elec',
L223.GlobalIntTechOMfixed_elec'
L223.GlobalTechCapital_Investment',
L223.GlobalIntTechCapital_elec',
L223.GlobalIntTechOMfixed_elec') ->
L223.StubTechMarket_Investment',
L223.GlobalTechCapital_Investment',
L223.GlobalIntTechCapital_elec',
L223.GlobalIntTechOMfixed_elec') ->
}
