% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2239.CSP_reeds_USA.R
\name{module_gcamusa_L2239.CSP_reeds_USA}
\alias{module_gcamusa_L2239.CSP_reeds_USA}
\title{module_gcam.usa_L2239.CSP_reeds_USA}
\usage{
module_gcamusa_L2239.CSP_reeds_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2239.DeleteUnlimitRsrc_reeds_USA}, \code{L2239.DeleteStubTechMinicamEnergyInput_CSP_reeds_USA},
\code{L2239.RenewRsrc_CSP_reeds_USA}, \code{L2239.GrdRenewRsrcCurves_CSP_reeds_USA},
\code{L2239.GrdRenewRsrcMax_CSP_reeds_USA}, \code{L2239.StubTechEffFlag_CSP_reeds_USA},
\code{L2239.StubTechCapFactor_CSP_reeds_USA}, \code{L2239.RenewRsrcTechChange_CSP_reeds_USA},
\code{L2239.StubTechCost_CSP_reeds_USA}, \code{L2239.ResTechShrwt_CSP_reeds_USA}.
The corresponding file in the original data system was \code{L2239.CSP_reeds_USA.R} (gcam-usa level2).
}
\description{
Create updated solar CSP resource supply curves consistent with ReEDS.
}
\details{
Create state-level solar CSP resource supply curves
}
\author{
MTB September 2018
}
