% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_L2242.elec_hydro_USA.R
\name{module_gcamusa_L2242.elec_hydro_USA}
\alias{module_gcamusa_L2242.elec_hydro_USA}
\title{module_gcamusa_L2242.elec_hydro_USA}
\usage{
module_gcamusa_L2242.elec_hydro_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L2242.CapacityTech_hydro_future}, \code{L2242.TechLifetime_hydro},
\code{L2242.CapacityTechSegmentCapFac_hydro}, \code{L2242.CapacityTechSegmentCapFac_hydro}.
The corresponding file in the original data system was \code{L2242.elec_hydro_USA.R} (gcam-usa level2).
}
\description{
Update hydropower capacity technology info to match historical (2019) generation for 2020 and all future years.
}
\details{
Update state-level hydro-electricity fixed outputs
}
\author{
MTB September 2018
}
