% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LA131.enduse_noEFW_USA.R
\name{module_gcamusa_LA131.enduse_noEFW_USA}
\alias{module_gcamusa_LA131.enduse_noEFW_USA}
\title{module_gcamusa_LA131.enduse_noEFW_USA}
\usage{
module_gcamusa_LA131.enduse_noEFW_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L131.in_EJ_USA_Senduse_F_Yh_noEFW}.
}
\description{
Generate the following two tables:
\itemize{
 \item{Final scaled energy input by USA / end-use sector (incl CHP) / fuel / historical year (no EFW)}
}
}
\details{
This chunk performs the same steps as module_energy_LA131.enduse, but works from the energy balances prior to the
energy-for-water deductions, in order to generate information for GCAM-USA which does not have energy-for-water
}
\author{
GPK November 2020
}
