% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB102.FERC_load_curves_USA.R
\name{module_gcamusa_LB102.FERC_load_curves_USA}
\alias{module_gcamusa_LB102.FERC_load_curves_USA}
\title{module_gcamusa_LB102.FERC_load_curves_USA}
\usage{
module_gcamusa_LB102.FERC_load_curves_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L102.date_load_curve_mapping_S_gcamusa}, \code{L102.load_segments_gcamusa},
\code{L102.invest_segments_gcamusa}. The corresponding file in the
original data system was \code{LA102.FERC_load_curves.R} (gcamusa dispatch level1).
}
\description{
Process FERC hours data into load segments by grid regions for dispatch model.
}
\details{
Compute load curve related parameters for investment and disparch segments
}
\author{
PLP YO May 2020
}
