% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB105.EIA_elec_vintage_USA.R
\name{module_gcamusa_LB105.EIA_elec_vintage_USA}
\alias{module_gcamusa_LB105.EIA_elec_vintage_USA}
\title{module_gcamusa_LB105.EIA_elec_vintage_USA}
\usage{
module_gcamusa_LB105.EIA_elec_vintage_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L105.elec_capacity_state_vintage_gcamusa}, \code{L105.elec_generation_state_vintage_gcamusa},
\code{L105.elec_fuelconsumption_state_vintage_gcamusa}.
There is no corresponding file in the original data system. This was originally a preprocessing code
}
\description{
Process 2015 EIA Form 923 and Form 860
}
\details{
Process 2015 EIA Form 923 and Form 860
}
\author{
YO Jun 2020
}
