% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zchunk_LB1235.elec_load_segments_USA.R
\name{module_gcamusa_LB1235.elec_load_segments_USA}
\alias{module_gcamusa_LB1235.elec_load_segments_USA}
\title{module_gcamusa_LB1235.elec_load_segments_USA}
\usage{
module_gcamusa_LB1235.elec_load_segments_USA(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L1235.grid_elec_supply_USA}, \code{L1235.elecS_horizontal_vertical_USA},
\code{L1235.elecS_horizontal_vertical_GCAM_coeff_USA}, \code{L1235.elecS_demand_fraction_USA}.

The corresponding file in the original data system was \code{LB1235.elec_load_segments.R} (gcam-usa level1).
}
\description{
Read in demand fraction and time fraction and compute load curve related parameters
Calculate an initial estimate of generation by fuel (EJ) in the horizontal segments
}
\details{
Compute load curve related parameters and nitial estimate of generation by fuel in the horizontal segments.
}
\author{
MTB August 2018
}
