% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsocio_L100.GDP_hist.R
\name{module_socio_L100.GDP_hist}
\alias{module_socio_L100.GDP_hist}
\title{module_socio_L100.GDP_hist}
\usage{
module_socio_L100.GDP_hist(command, ...)
}
\arguments{
\item{command}{API command to execute}

\item{...}{other optional parameters, depending on command}
}
\value{
Depends on \code{command}: either a vector of required inputs,
a vector of output names, or (if \code{command} is "MAKE") all
the generated outputs: \code{L100.gdp_mil90usd_ctry_Yh}. The corresponding file in the
original data system was \code{L100.GDP_hist.R} (socioeconomics level1).
}
\description{
Prepare GDP database for later use: filter missing values and convert units to 1990 USD.
}
\details{
At present the GDP database used requires no downscaling and all
major countries are included, so really no processing steps are needed.
All that happens in this file right now is filtering out \code{NA} values
and converting the units to GCAM's GDP unit (million 1990 USD).
}
\author{
BBL February 2017
}
