% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PREBUILT_DATA}
\alias{PREBUILT_DATA}
\title{A list of prebuilt data objects. These are used when the proprietary IEA
energy data files are not available, and thus
\code{\link{module_energy_LA100.IEA_downscale_ctry}} is not able to run.
Its immediate downstream dependencies then used the prebuilt versions of
their outputs stored in this object.}
\format{
A list object where [[object_name]] <- tibble:
}
\usage{
PREBUILT_DATA
}
\description{
A list of prebuilt data objects. These are used when the proprietary IEA
energy data files are not available, and thus
\code{\link{module_energy_LA100.IEA_downscale_ctry}} is not able to run.
Its immediate downstream dependencies then used the prebuilt versions of
their outputs stored in this object.
}
\keyword{datasets}
